% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.covid.global_deaths}
\alias{alt.covid.global_deaths}
\title{Get historical deaths for given country.}
\usage{
alt.covid.global_deaths(country)
}
\arguments{
\item{country}{(character length 1) Country to search for}
}
\description{
Wrapper for Python function alt.covid.global_deaths from OpenBB Terminal SDK
}
\examples{
alt.covid.global_deaths(country='Canada')
}
