% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.covid.ov}
\alias{alt.covid.ov}
\title{Get historical cases and deaths by country.}
\usage{
alt.covid.ov(country, limit = 100)
}
\arguments{
\item{country}{(character length 1) Country to get data for}

\item{limit}{(integer length 1) Number of raw data to show}
}
\description{
Wrapper for Python function alt.covid.ov from OpenBB Terminal SDK
}
\examples{
alt.covid.ov(country='Canada', limit=100)
}
