% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.covid.slopes}
\alias{alt.covid.slopes}
\title{Load cases and find slope over period.}
\usage{
alt.covid.slopes(days_back = 30, limit = 50, threshold = 10000, ascend = FALSE)
}
\arguments{
\item{days_back}{(integer length 1) Number of historical days to consider}

\item{limit}{(integer length 1) Number of rows to show}

\item{threshold}{(integer length 1) Threshold for total number of cases}

\item{ascend}{(logical length 1) Flag to sort in ascending order}
}
\description{
Wrapper for Python function alt.covid.slopes from OpenBB Terminal SDK
}
\examples{
alt.covid.slopes(days_back=30, limit=50, threshold=10000, ascend=FALSE)
}
