% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.covid.stat}
\alias{alt.covid.stat}
\title{Show historical cases and deaths by country.}
\usage{
alt.covid.stat(country, stat = "cases", limit = 10)
}
\arguments{
\item{country}{(character length 1) Country to get data for}

\item{stat}{(character length 1) Statistic to get.  Either "cases", "deaths" or "rates"}

\item{limit}{(integer length 1) Number of raw data to show}
}
\description{
Wrapper for Python function alt.covid.stat from OpenBB Terminal SDK
}
\examples{
alt.covid.stat(country='Canada', stat='cases', limit=10)
}
