% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.covid.stat_chart}
\alias{alt.covid.stat_chart}
\title{Prints table showing historical cases and deaths by country.}
\usage{
alt.covid.stat_chart(
  country,
  stat = "cases",
  raw = FALSE,
  limit = 10,
  export = "",
  sheet_name,
  plot = TRUE
)
}
\arguments{
\item{country}{(character length 1) Country to get data for}

\item{stat}{(character length 1) Statistic to get.  Either "cases", "deaths" or "rates"}

\item{raw}{(logical length 1) Flag to display raw data}

\item{limit}{(integer length 1) Number of raw data to show}

\item{export}{(character length 1) Format to export data}

\item{plot}{(logical length 1) Flag to plot data}
}
\description{
Wrapper for Python function alt.covid.stat_chart from OpenBB Terminal SDK
}
\examples{
alt.covid.stat_chart(country='Canada', stat='cases', raw=FALSE, limit=10, plot=TRUE)
}
