% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.hn}
\alias{alt.hn}
\title{Get top stories from HackerNews.}
\usage{
alt.hn(limit = 10)
}
\arguments{
\item{limit}{(integer length 1) Number of stories to return}
}
\description{
Wrapper for Python function alt.hn from OpenBB Terminal SDK
}
