% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.hn_chart}
\alias{alt.hn_chart}
\title{View top stories from HackerNews.}
\usage{
alt.hn_chart(limit = 10, export = "", sheet_name)
}
\arguments{
\item{limit}{(integer length 1) Number of stories to return}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function alt.hn_chart from OpenBB Terminal SDK
}
