% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.github_data}
\alias{alt.oss.github_data}
\title{Get repository stats.}
\usage{
alt.oss.github_data(url, kwargs)
}
\arguments{
\item{url}{(character length 1) github api endpoint}
}
\description{
Wrapper for Python function alt.oss.github_data from OpenBB Terminal SDK
}
