% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.history}
\alias{alt.oss.history}
\title{Get repository star history.}
\usage{
alt.oss.history(repo)
}
\arguments{
\item{repo}{(character length 1) Repo to search for Format: org/repo, e.g., openbb-finance/openbbterminal}
}
\description{
Wrapper for Python function alt.oss.history from OpenBB Terminal SDK
}
