% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.history_chart}
\alias{alt.oss.history_chart}
\title{Plots repo summary \link{Source: https://api.github.com}.}
\usage{
alt.oss.history_chart(repo, export = "", sheet_name, external_axes)
}
\arguments{
\item{repo}{(character length 1) Repository to display star history. Format: org/repo, e.g., openbb-finance/openbbterminal}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function alt.oss.history_chart from OpenBB Terminal SDK
}
