% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.ross}
\alias{alt.oss.ross}
\title{Get startups from ROSS index \link{Source: https://runacap.com/}.}
\usage{
alt.oss.ross()
}
\description{
Wrapper for Python function alt.oss.ross from OpenBB Terminal SDK
}
