% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.ross_chart}
\alias{alt.oss.ross_chart}
\title{Plots list of startups from ross index \link{Source: https://runacap.com/}}
\usage{
alt.oss.ross_chart(
  limit = 10,
  sortby = "Stars AGR [\%]",
  ascend = FALSE,
  show_chart = FALSE,
  show_growth = TRUE,
  chart_type = "stars",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{limit}{(integer length 1) Number of startups to search}

\item{sortby}{(character length 1) Key by which to sort data. Default: Stars AGR \link{\%}}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{show_chart}{(logical length 1) Flag to show chart with startups}

\item{show_growth}{(logical length 1) Flag to show growth line chart}

\item{chart_type}{(character length 1) Chart type {stars,forks}}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function alt.oss.ross_chart from OpenBB Terminal SDK
}
\examples{
alt.oss.ross_chart(limit=10, sortby='Stars AGR [\%]', ascend=FALSE, show_chart=FALSE, show_growth=TRUE, chart_type='stars')
}
