% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.search}
\alias{alt.oss.search}
\title{Get repos sorted by stars or forks. Can be filtered by categories.}
\usage{
alt.oss.search(sortby = "stars", page = 1, categories = "")
}
\arguments{
\item{sortby}{(character length 1) Sort repos by {stars, forks}}

\item{page}{(integer length 1) Page number to get repos}

\item{categories}{(character length 1) Check for repo categories. If more than one separate with a comma: e.g., finance,investment. Default: None}
}
\description{
Wrapper for Python function alt.oss.search from OpenBB Terminal SDK
}
\examples{
alt.oss.search(sortby='stars', page=1)
}
