% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.summary}
\alias{alt.oss.summary}
\title{Get repository summary.}
\usage{
alt.oss.summary(repo)
}
\arguments{
\item{repo}{(character length 1) Repo to search for Format: org/repo, e.g., openbb-finance/openbbterminal}
}
\description{
Wrapper for Python function alt.oss.summary from OpenBB Terminal SDK
}
