% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.top}
\alias{alt.oss.top}
\title{Get repos sorted by stars or forks. Can be filtered by categories.}
\usage{
alt.oss.top(sortby, limit = 50, categories = "")
}
\arguments{
\item{sortby}{(character length 1) Sort repos by {stars, forks}}

\item{limit}{(integer length 1) Number of repos to search for}

\item{categories}{(character length 1) Check for repo categories. If more than one separate with a comma: e.g., finance,investment. Default: None}
}
\description{
Wrapper for Python function alt.oss.top from OpenBB Terminal SDK
}
\examples{
alt.oss.top(limit=50)
}
