% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.oss.top_chart}
\alias{alt.oss.top_chart}
\title{Plots repo summary \link{Source: https://api.github.com}.}
\usage{
alt.oss.top_chart(
  sortby,
  categories = "",
  limit = 10,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{sortby}{(character length 1) Sort repos by {stars, forks}}

\item{categories}{(character length 1) Check for repo categories. If more than one separate with a comma: e.g., finance,investment. Default: None}

\item{limit}{(integer length 1) Number of repos to look at}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function alt.oss.top_chart from OpenBB Terminal SDK
}
\examples{
alt.oss.top_chart(limit=10)
}
