% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.realestate.get_estate_sales}
\alias{alt.realestate.get_estate_sales}
\title{All sales for specified postcode.}
\usage{
alt.realestate.get_estate_sales(postcode, limit = 25)
}
\arguments{
\item{postcode}{(character length 1) Postcode}
}
\description{
Wrapper for Python function alt.realestate.get_estate_sales from OpenBB Terminal SDK
}
