% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.realestate.get_region_stats}
\alias{alt.realestate.get_region_stats}
\title{Get regional house price statistics.}
\usage{
alt.realestate.get_region_stats(
  region,
  start_date = "2010-01-01",
  end_date = ""
)
}
\arguments{
\item{region}{(character length 1) region}
}
\description{
Wrapper for Python function alt.realestate.get_region_stats from OpenBB Terminal SDK
}
