% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alt.realestate.get_towns_sold_prices}
\alias{alt.realestate.get_towns_sold_prices}
\title{Get towns sold house price data.}
\usage{
alt.realestate.get_towns_sold_prices(town, start_date, end_date, limit = 25)
}
\arguments{
\item{town}{(character length 1) town}
}
\description{
Wrapper for Python function alt.realestate.get_towns_sold_prices from OpenBB Terminal SDK
}
