% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.candle}
\alias{crypto.candle}
\title{Plot candle chart from dataframe. \link{Source: Binance}}
\usage{
crypto.candle(
  symbol,
  data,
  start_date,
  end_date,
  interval,
  exchange = "binance",
  to_symbol = "usdt",
  source = "CCXT",
  volume = TRUE,
  ylabel = "",
  title = "",
  external_axes,
  yscale = "linear",
  raw = FALSE
)
}
\arguments{
\item{symbol}{(character length 1) Ticker name}

\item{exchange}{(character length 1) Exchange to use}

\item{to_symbol}{(character length 1) Currency to use}

\item{source}{(character length 1) Source to use}

\item{volume}{(logical length 1) If volume data shall be plotted, by default True}

\item{ylabel}{(character length 1) Y-label of the graph, by default ""}

\item{title}{(character length 1) Title of graph, by default ""}

\item{yscale}{(character length 1) Scaling for y axis.  Either linear or log}
}
\description{
Wrapper for Python function crypto.candle from OpenBB Terminal SDK
}
