% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.chart}
\alias{crypto.chart}
\title{Load data for Technical Analysis}
\usage{
crypto.chart(
  prices_df,
  to_symbol = "",
  from_symbol = "",
  source = "",
  exchange = "",
  interval = "",
  external_axes,
  yscale = "linear"
)
}
\arguments{
\item{prices_df}{(data.frame) Cryptocurrency}

\item{to_symbol}{(character length 1) Coin (only used for chart title), by default ""}

\item{from_symbol}{(character length 1) Currency (only used for chart title), by default ""}

\item{yscale}{(character length 1) Scale for y axis of plot Either linear or log}
}
\description{
Wrapper for Python function crypto.chart from OpenBB Terminal SDK
}
\examples{
crypto.chart(yscale='linear')
}
