% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.active}
\alias{crypto.dd.active}
\title{Returns active addresses of a certain symbol}
\usage{
crypto.dd.active(symbol, interval = "24h", start_date = "2010-01-01", end_date)
}
\arguments{
\item{symbol}{(character length 1) Asset to search active addresses (e.g., BTC)}

\item{interval}{(character length 1) Interval frequency (e.g., 24h)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.dd.active from OpenBB Terminal SDK
}
\examples{
crypto.dd.active(symbol='AMZN', interval='24h', start_date='2010-01-01')
}
