% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.active_chart}
\alias{crypto.dd.active_chart}
\title{Plots active addresses of a certain symbol over time}
\usage{
crypto.dd.active_chart(
  symbol,
  start_date = "2010-01-01",
  end_date,
  interval = "24h",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Asset to search active addresses (e.g., BTC)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{interval}{(character length 1) Interval frequency (possible values are: 24h, 1w, 1month)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.active_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.active_chart(symbol='AMZN', start_date='2010-01-01', interval='24h')
}
