% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.all_binance_trading_pairs}
\alias{crypto.dd.all_binance_trading_pairs}
\title{Returns all available pairs on Binance in DataFrame format. DataFrame has 3 columns symbol, baseAsset, quoteAsset}
\usage{
crypto.dd.all_binance_trading_pairs()
}
\description{
Wrapper for Python function crypto.dd.all_binance_trading_pairs from OpenBB Terminal SDK
}
