% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ath}
\alias{crypto.dd.ath}
\title{Get all time high for a coin in a given currency}
\usage{
crypto.dd.ath(symbol, currency = "USD")
}
\arguments{
\item{symbol}{(character length 1) Coin to get all time high for}

\item{currency}{(character length 1) Currency to get all time high in}
}
\description{
Wrapper for Python function crypto.dd.ath from OpenBB Terminal SDK
}
