% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.balance}
\alias{crypto.dd.balance}
\title{Get account holdings for asset. \link{Source: Binance}}
\usage{
crypto.dd.balance(from_symbol, to_symbol = "USDT")
}
\arguments{
\item{from_symbol}{(character length 1) Cryptocurrency}

\item{to_symbol}{(character length 1) Cryptocurrency}
}
\description{
Wrapper for Python function crypto.dd.balance from OpenBB Terminal SDK
}
\examples{
crypto.dd.balance(to_symbol='USDT')
}
