% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.balance_chart}
\alias{crypto.dd.balance_chart}
\title{Prints table showing account holdings for asset. \link{Source: Binance}}
\usage{
crypto.dd.balance_chart(
  from_symbol,
  to_symbol = "USDT",
  export = "",
  sheet_name
)
}
\arguments{
\item{from_symbol}{(character length 1) Cryptocurrency}

\item{to_symbol}{(character length 1) Cryptocurrency}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx}
}
\description{
Wrapper for Python function crypto.dd.balance_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.balance_chart(to_symbol='USDT')
}
