% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.basic}
\alias{crypto.dd.basic}
\title{Basic coin information \link{Source: CoinPaprika}}
\usage{
crypto.dd.basic(symbol = "BTC")
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}
}
\description{
Wrapper for Python function crypto.dd.basic from OpenBB Terminal SDK
}
\examples{
crypto.dd.basic(symbol='BTC')
}
