% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.basic_chart}
\alias{crypto.dd.basic_chart}
\title{Prints table showing basic information for coin. Like:}
\usage{
crypto.dd.basic_chart(symbol = "BTC", export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx}
}
\description{
Wrapper for Python function crypto.dd.basic_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.basic_chart(symbol='BTC')
}
