% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.binance_available_quotes_for_each_coin}
\alias{crypto.dd.binance_available_quotes_for_each_coin}
\title{Helper methods that for every coin available on Binance add all quote assets. \link{Source: Binance}}
\usage{
crypto.dd.binance_available_quotes_for_each_coin()
}
\description{
Wrapper for Python function crypto.dd.binance_available_quotes_for_each_coin from OpenBB Terminal SDK
}
