% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.candle}
\alias{crypto.dd.candle}
\title{Get candles for chosen trading pair and time interval. \link{Source: Coinbase}}
\usage{
crypto.dd.candle(symbol, interval = "24hour")
}
\arguments{
\item{symbol}{(character length 1) Trading pair of coins on Coinbase e.g ETH-USDT or UNI-ETH}

\item{interval}{(character length 1) Time interval. One from 1min, 5min ,15min, 1hour, 6hour, 24hour, 1day}
}
\description{
Wrapper for Python function crypto.dd.candle from OpenBB Terminal SDK
}
