% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.change}
\alias{crypto.dd.change}
\title{Returns 30d change of the supply held in exchange wallets of a certain symbol.}
\usage{
crypto.dd.change(
  symbol,
  exchange = "binance",
  start_date = "2010-01-01",
  end_date
)
}
\arguments{
\item{symbol}{(character length 1) Asset symbol to search supply (e.g., BTC)}

\item{exchange}{(character length 1) Exchange to check net position change (e.g., binance)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.dd.change from OpenBB Terminal SDK
}
\examples{
crypto.dd.change(symbol='AMZN', exchange='binance', start_date='2010-01-01')
}
