% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.change_chart}
\alias{crypto.dd.change_chart}
\title{Plots 30d change of the supply held in exchange wallets.}
\usage{
crypto.dd.change_chart(
  symbol,
  exchange = "binance",
  start_date = "2010-01-01",
  end_date,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Asset to search active addresses (e.g., BTC)}

\item{exchange}{(character length 1) Exchange to check net position change (possible values are: aggregated, binance,
bittrex, coinex, gate.io, gemini, huobi, kucoin, poloniex, bibox, bigone, bitfinex,
hitbtc, kraken, okex, bithumb, zb.com, cobinhood, bitmex, bitstamp, coinbase, coincheck, luno)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.change_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.change_chart(symbol='AMZN', exchange='binance', start_date='2010-01-01')
}
