% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.check_valid_binance_str}
\alias{crypto.dd.check_valid_binance_str}
\title{Check if symbol is in defined binance. \link{Source: Binance}}
\usage{
crypto.dd.check_valid_binance_str(symbol)
}
\description{
Wrapper for Python function crypto.dd.check_valid_binance_str from OpenBB Terminal SDK
}
