% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.close}
\alias{crypto.dd.close}
\title{Returns the price of a cryptocurrency}
\usage{
crypto.dd.close(
  symbol,
  start_date = "2010-01-01",
  end_date,
  print_errors = TRUE
)
}
\arguments{
\item{symbol}{(character length 1) Crypto to check close price (BTC or ETH)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{print_errors}{(logical length 1) Flag to print errors. Default: True}
}
\description{
Wrapper for Python function crypto.dd.close from OpenBB Terminal SDK
}
\examples{
crypto.dd.close(symbol='ETH', start_date='2010-01-01', print_errors=TRUE)
}
