% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.coin_market_chart}
\alias{crypto.dd.coin_market_chart}
\title{Get prices for given coin. \link{Source: CoinGecko}}
\usage{
crypto.dd.coin_market_chart(
  symbol = "",
  vs_currency = "usd",
  days = 30,
  kwargs
)
}
\arguments{
\item{vs_currency}{(character length 1) currency vs which display data}

\item{days}{(integer length 1) number of days to display the data}
}
\description{
Wrapper for Python function crypto.dd.coin_market_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.coin_market_chart(vs_currency='usd', days=30)
}
