% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.dev}
\alias{crypto.dd.dev}
\title{Get developer stats for a coin}
\usage{
crypto.dd.dev(symbol)
}
\arguments{
\item{symbol}{(character length 1) Coin to get stats for}
}
\description{
Wrapper for Python function crypto.dd.dev from OpenBB Terminal SDK
}
