% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.eb}
\alias{crypto.dd.eb}
\title{Returns the total amount of coins held on exchange addresses in units and percentage.}
\usage{
crypto.dd.eb(symbol, exchange = "aggregated", start_date, end_date)
}
\arguments{
\item{symbol}{(character length 1) Asset to search active addresses (e.g., BTC)}

\item{exchange}{(character length 1) Exchange to check net position change (possible values are: aggregated, binance, bittrex,
coinex, gate.io, gemini, huobi, kucoin, poloniex, bibox, bigone, bitfinex, hitbtc, kraken,
okex, bithumb, zb.com, cobinhood, bitmex, bitstamp, coinbase, coincheck, luno), by default "aggregated"}
}
\description{
Wrapper for Python function crypto.dd.eb from OpenBB Terminal SDK
}
\examples{
crypto.dd.eb(symbol='AMZN', exchange='aggregated')
}
