% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.eb_chart}
\alias{crypto.dd.eb_chart}
\title{Plots total amount of coins held on exchange addresses in units and percentage.}
\usage{
crypto.dd.eb_chart(
  symbol,
  exchange = "aggregated",
  start_date,
  end_date,
  percentage = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Asset to search active addresses (e.g., BTC)}

\item{exchange}{(character length 1) Exchange to check net position change (possible values are: aggregated, binance, bittrex,
coinex, gate.io, gemini, huobi, kucoin, poloniex, bibox, bigone, bitfinex, hitbtc, kraken,
okex, bithumb, zb.com, cobinhood, bitmex, bitstamp, coinbase, coincheck, luno), by default "aggregated"}

\item{percentage}{(logical length 1) Show percentage instead of stacked value.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.eb_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.eb_chart(symbol='AMZN', exchange='aggregated', percentage=FALSE)
}
