% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.events_chart}
\alias{crypto.dd.events_chart}
\title{Prints table showing all events for given coin id. \link{Source: CoinPaprika}}
\usage{
crypto.dd.events_chart(
  symbol = "BTC",
  limit = 10,
  sortby = "date",
  ascend = FALSE,
  links = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid
(see for possible values:
https://api.coinpaprika.com/docs#tag/Coins/paths/~1coins~1\%7Bcoin_id\%7D~1events/get).}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{links}{(logical length 1) Flag to display urls}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.events_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.events_chart(symbol='BTC', limit=10, sortby='date', ascend=FALSE, links=FALSE)
}
