% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ex}
\alias{crypto.dd.ex}
\title{Get all exchanges for given coin id. \link{Source: CoinPaprika}}
\usage{
crypto.dd.ex(
  symbol = "BTC",
  sortby = "adjusted_volume_24h_share",
  ascend = TRUE
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid (see for possible values:
https://api.coinpaprika.com/v1).}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.dd.ex from OpenBB Terminal SDK
}
\examples{
crypto.dd.ex(symbol='BTC', sortby='adjusted_volume_24h_share', ascend=TRUE)
}
