% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ex_chart}
\alias{crypto.dd.ex_chart}
\title{Prints table showing all exchanges for given coin id. \link{Source: CoinPaprika}}
\usage{
crypto.dd.ex_chart(
  symbol = "btc",
  limit = 10,
  sortby = "adjusted_volume_24h_share",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid (see for possible values:
https://api.coinpaprika.com/v1).}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.ex_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.ex_chart(symbol='btc', limit=10, sortby='adjusted_volume_24h_share', ascend=TRUE)
}
