% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.fr}
\alias{crypto.dd.fr}
\title{Returns coin fundraising}
\usage{
crypto.dd.fr(symbol)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check fundraising}
}
\description{
Wrapper for Python function crypto.dd.fr from OpenBB Terminal SDK
}
\examples{
crypto.dd.fr(symbol='ETH')
}
