% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.get_mt}
\alias{crypto.dd.get_mt}
\title{Returns available messari timeseries}
\usage{
crypto.dd.get_mt(only_free = TRUE)
}
\arguments{
\item{only_free}{(logical length 1) Display only timeseries available for free}
}
\description{
Wrapper for Python function crypto.dd.get_mt from OpenBB Terminal SDK
}
\examples{
crypto.dd.get_mt(only_free=TRUE)
}
