% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.get_mt_chart}
\alias{crypto.dd.get_mt_chart}
\title{Prints table showing messari timeseries list}
\usage{
crypto.dd.get_mt_chart(
  limit = 10,
  query = "",
  only_free = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) number to show}

\item{query}{(character length 1) Query to search across all messari timeseries}

\item{only_free}{(logical length 1) Display only timeseries available for free}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.get_mt_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.get_mt_chart(limit=10, only_free=TRUE)
}
