% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.gh}
\alias{crypto.dd.gh}
\title{Returns  a list of developer activity for a given coin and time interval.}
\usage{
crypto.dd.gh(
  symbol,
  dev_activity = FALSE,
  interval = "1d",
  start_date,
  end_date
)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check github activity}

\item{dev_activity}{(logical length 1) Whether to filter only for development activity}

\item{interval}{(character length 1) Interval frequency (e.g., 1d)}
}
\description{
Wrapper for Python function crypto.dd.gh from OpenBB Terminal SDK
}
\examples{
crypto.dd.gh(symbol='ETH', dev_activity=FALSE, interval='1d')
}
