% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.gov}
\alias{crypto.dd.gov}
\title{Returns coin governance}
\usage{
crypto.dd.gov(symbol)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check governance}
}
\description{
Wrapper for Python function crypto.dd.gov from OpenBB Terminal SDK
}
\examples{
crypto.dd.gov(symbol='ETH')
}
