% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.headlines}
\alias{crypto.dd.headlines}
\title{Gets Sentiment analysis provided by FinBrain's API \link{Source: finbrain}.}
\usage{
crypto.dd.headlines(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the sentiment analysis from}
}
\description{
Wrapper for Python function crypto.dd.headlines from OpenBB Terminal SDK
}
\examples{
crypto.dd.headlines(symbol='AMZN')
}
