% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.headlines_chart}
\alias{crypto.dd.headlines_chart}
\title{Sentiment analysis from FinBrain for Cryptocurrencies}
\usage{
crypto.dd.headlines_chart(
  symbol,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency}

\item{raw}{(logical length 1) Display raw table data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.headlines_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.headlines_chart(symbol='ETH')
}
