% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.inv}
\alias{crypto.dd.inv}
\title{Returns coin investors}
\usage{
crypto.dd.inv(symbol)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check investors}
}
\description{
Wrapper for Python function crypto.dd.inv from OpenBB Terminal SDK
}
\examples{
crypto.dd.inv(symbol='ETH')
}
