% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.mcapdom}
\alias{crypto.dd.mcapdom}
\title{Returns market dominance of a coin over time}
\usage{
crypto.dd.mcapdom(symbol, interval = "1d", start_date, end_date)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check market cap dominance}

\item{interval}{(character length 1) Interval frequency (possible values are: 5m, 15m, 30m, 1h, 1d, 1w)}
}
\description{
Wrapper for Python function crypto.dd.mcapdom from OpenBB Terminal SDK
}
\examples{
crypto.dd.mcapdom(symbol='ETH', interval='1d')
}
