% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.mcapdom_chart}
\alias{crypto.dd.mcapdom_chart}
\title{Plots market dominance of a coin over time}
\usage{
crypto.dd.mcapdom_chart(
  symbol,
  start_date,
  end_date,
  interval = "1d",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check market cap dominance}

\item{interval}{(character length 1) Interval frequency (possible values are: 5m, 15m, 30m, 1h, 1d, 1w)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.mcapdom_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.mcapdom_chart(symbol='ETH', interval='1d')
}
