% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.mkt}
\alias{crypto.dd.mkt}
\title{All markets for given coin and currency \link{Source: CoinPaprika}}
\usage{
crypto.dd.mkt(
  symbol = "BTC",
  quotes = "USD",
  sortby = "pct_volume_share",
  ascend = TRUE
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{quotes}{(character length 1) Comma separated list of quotes to return.
Example: quotes=USD,BTC
Allowed values:
BTC, ETH, USD, EUR, PLN, KRW, GBP, CAD, JPY, RUB, TRY, NZD, AUD, CHF, UAH, HKD, SGD, NGN,
PHP, MXN, BRL, THB, CLP, CNY, CZK, DKK, HUF, IDR, ILS, INR, MYR, NOK, PKR, SEK, TWD, ZAR,
VND, BOB, COP, PEN, ARS, ISK}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid (see for possible values:
https://api.coinpaprika.com/v1).}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.dd.mkt from OpenBB Terminal SDK
}
\examples{
crypto.dd.mkt(symbol='BTC', quotes='USD', sortby='pct_volume_share', ascend=TRUE)
}
