% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.mkt_chart}
\alias{crypto.dd.mkt_chart}
\title{Prints table showing all markets for given coin id. \link{Source: CoinPaprika}}
\usage{
crypto.dd.mkt_chart(
  from_symbol = "BTC",
  to_symbol = "USD",
  limit = 20,
  sortby = "pct_volume_share",
  ascend = TRUE,
  links = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{from_symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{to_symbol}{(character length 1) Quoted currency}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid (see for possible values:
https://api.coinpaprika.com/v1).}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{links}{(logical length 1) Flag to display urls}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.mkt_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.mkt_chart(from_symbol='BTC', to_symbol='USD', limit=20, sortby='pct_volume_share', ascend=TRUE, links=FALSE)
}
