% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.mt_chart}
\alias{crypto.dd.mt_chart}
\title{Plots messari timeseries}
\usage{
crypto.dd.mt_chart(
  symbol,
  timeseries_id,
  start_date,
  end_date,
  interval = "1d",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check market cap dominance}

\item{timeseries_id}{(character length 1) Obtained by api.crypto.dd.get_mt command}

\item{interval}{(character length 1) Interval frequency (possible values are: 5m, 15m, 30m, 1h, 1d, 1w)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.mt_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.mt_chart(symbol='ETH', interval='1d')
}
