% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.news}
\alias{crypto.dd.news}
\title{Get recent posts from CryptoPanic news aggregator platform. \link{Source: https://cryptopanic.com/}}
\usage{
crypto.dd.news(
  limit = 60,
  post_kind = "news",
  filter_,
  region = "en",
  source,
  symbol,
  sortby = "published_at",
  ascend = TRUE
)
}
\arguments{
\item{limit}{(integer length 1) number of news to fetch}

\item{post_kind}{(character length 1) Filter by category of news. Available values: news or media.}

\item{region}{(character length 1) Filter news by regions. Available regions are: en (English), de (Deutsch), nl (Dutch),
es (Español), fr (Français), it (Italiano), pt (Português), ru (Русский)}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in ascend order.}
}
\description{
Wrapper for Python function crypto.dd.news from OpenBB Terminal SDK
}
\examples{
crypto.dd.news(limit=60, post_kind='news', region='en', sortby='published_at', ascend=TRUE)
}
