% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.news_chart}
\alias{crypto.dd.news_chart}
\title{Prints table showing recent posts from CryptoPanic news aggregator platform.}
\usage{
crypto.dd.news_chart(
  post_kind = "news",
  region = "en",
  filter_,
  source,
  symbol,
  limit = 25,
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{post_kind}{(character length 1) Filter by category of news. Available values: news or media.}

\item{region}{(character length 1) Filter news by regions. Available regions are: en (English), de (Deutsch), nl (Dutch),
es (Español), fr (Français), it (Italiano), pt (Português), ru (Русский)}

\item{limit}{(integer length 1) number of news to display}

\item{ascend}{(logical length 1) Sort in ascending order.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.news_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.news_chart(limit=25, post_kind='news', region='en', ascend=TRUE)
}
