% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.nonzero}
\alias{crypto.dd.nonzero}
\title{Returns addresses with non-zero balance of a certain symbol}
\usage{
crypto.dd.nonzero(symbol, start_date = "2010-01-01", end_date)
}
\arguments{
\item{symbol}{(character length 1) Asset to search (e.g., BTC)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.dd.nonzero from OpenBB Terminal SDK
}
\examples{
crypto.dd.nonzero(symbol='AMZN', start_date='2010-01-01')
}
