% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ob}
\alias{crypto.dd.ob}
\title{Returns orderbook for a coin in a given exchange}
\usage{
crypto.dd.ob(exchange, symbol, to_symbol)
}
\arguments{
\item{exchange}{(character length 1) exchange id}

\item{symbol}{(character length 1) coin symbol}

\item{to_symbol}{(character length 1) currency to compare coin against}
}
\description{
Wrapper for Python function crypto.dd.ob from OpenBB Terminal SDK
}
\examples{
crypto.dd.ob(symbol='AMZN')
}
