% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ob_chart}
\alias{crypto.dd.ob_chart}
\title{Plots order book for a coin in a given exchange}
\usage{
crypto.dd.ob_chart(
  exchange,
  symbol,
  to_symbol,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{exchange}{(character length 1) exchange id}

\item{symbol}{(character length 1) coin symbol}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.ob_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.ob_chart(symbol='AMZN')
}
