% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.oi}
\alias{crypto.dd.oi}
\title{Returns open interest by exchange for a certain symbol}
\usage{
crypto.dd.oi(symbol, interval = 0)
}
\arguments{
\item{symbol}{(character length 1) Crypto Symbol to search open interest futures (e.g., BTC)}

\item{interval}{(integer length 1) Frequency (possible values are: 0 for ALL, 2 for 1H, 1 for 4H, 4 for 12H), by default 0}
}
\description{
Wrapper for Python function crypto.dd.oi from OpenBB Terminal SDK
}
\examples{
crypto.dd.oi(symbol='ETH', interval=0)
}
