% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.oi_chart}
\alias{crypto.dd.oi_chart}
\title{Plots open interest by exchange for a certain cryptocurrency}
\usage{
crypto.dd.oi_chart(symbol, interval = 0, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to search open interest (e.g., BTC)}

\item{interval}{(integer length 1) Frequency (possible values are: 0 for ALL, 2 for 1H, 1 for 4H, 4 for 12H), by default 0}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.oi_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.oi_chart(symbol='ETH', interval=0)
}
