% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.pi}
\alias{crypto.dd.pi}
\title{Returns coin product info}
\usage{
crypto.dd.pi(symbol)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check product info}
}
\description{
Wrapper for Python function crypto.dd.pi from OpenBB Terminal SDK
}
\examples{
crypto.dd.pi(symbol='ETH')
}
