% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.pr}
\alias{crypto.dd.pr}
\title{Fetch data to calculate potential returns of a certain coin. \link{Source: CoinGecko}}
\usage{
crypto.dd.pr(main_coin, to_symbol, limit, price)
}
\arguments{
\item{main_coin}{(character length 1) Coin loaded to check potential returns for (e.g., algorand)}
}
\description{
Wrapper for Python function crypto.dd.pr from OpenBB Terminal SDK
}
\examples{
crypto.dd.pr(to_symbol='Coin to compare main_coin with (e.g., bitcoin)')
}
