% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.pr_chart}
\alias{crypto.dd.pr_chart}
\title{Prints table showing potential returns of a certain coin. \link{Source: CoinGecko}}
\usage{
crypto.dd.pr_chart(
  to_symbol,
  from_symbol,
  limit,
  price,
  export = "",
  sheet_name
)
}
\arguments{
\item{to_symbol}{(character length 1) Coin loaded to check potential returns for (e.g., algorand)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.pr_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.pr_chart(from_symbol='Coin to compare main_coin with (e.g., bitcoin)')
}
