% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ps}
\alias{crypto.dd.ps}
\title{Get all most important ticker related information for given coin id \link{Source: CoinPaprika}}
\usage{
crypto.dd.ps(symbol = "BTC", quotes = "USD")
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{quotes}{(character length 1) Comma separated quotes to return e.g quotes = USD, BTC}
}
\description{
Wrapper for Python function crypto.dd.ps from OpenBB Terminal SDK
}
\examples{
crypto.dd.ps(symbol='BTC', quotes='USD')
}
