% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.ps_chart}
\alias{crypto.dd.ps_chart}
\title{Prints table showing ticker information for single coin \link{Source: CoinPaprika}}
\usage{
crypto.dd.ps_chart(
  from_symbol = "BTC",
  to_symbol = "USD",
  export = "",
  sheet_name
)
}
\arguments{
\item{from_symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{to_symbol}{(character length 1) Quoted currency}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx}
}
\description{
Wrapper for Python function crypto.dd.ps_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.ps_chart(from_symbol='BTC', to_symbol='USD')
}
