% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.rm}
\alias{crypto.dd.rm}
\title{Returns coin roadmap}
\usage{
crypto.dd.rm(symbol, ascend = TRUE)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check roadmap}

\item{ascend}{(logical length 1) reverse order}
}
\description{
Wrapper for Python function crypto.dd.rm from OpenBB Terminal SDK
}
\examples{
crypto.dd.rm(symbol='ETH', ascend=TRUE)
}
