% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.rm_chart}
\alias{crypto.dd.rm_chart}
\title{Plots coin roadmap}
\usage{
crypto.dd.rm_chart(
  symbol,
  ascend = TRUE,
  limit = 5,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check roadmap}

\item{ascend}{(logical length 1) reverse order}

\item{limit}{(integer length 1) number to show}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.rm_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.rm_chart(symbol='ETH', ascend=TRUE, limit=5)
}
