% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.score}
\alias{crypto.dd.score}
\title{Get scores for a coin from CoinGecko}
\usage{
crypto.dd.score(symbol)
}
\arguments{
\item{symbol}{(character length 1) Coin to get scores for}
}
\description{
Wrapper for Python function crypto.dd.score from OpenBB Terminal SDK
}
