% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.show_available_pairs_for_given_symbol}
\alias{crypto.dd.show_available_pairs_for_given_symbol}
\title{Return all available quoted assets for given symbol. \link{Source: Coinbase}}
\usage{
crypto.dd.show_available_pairs_for_given_symbol(symbol = "ETH")
}
\arguments{
\item{symbol}{(character length 1) Uppercase symbol of coin e.g BTC, ETH, UNI, LUNA, DOT ...}
}
\description{
Wrapper for Python function crypto.dd.show_available_pairs_for_given_symbol from OpenBB Terminal SDK
}
\examples{
crypto.dd.show_available_pairs_for_given_symbol(symbol='ETH')
}
