% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.social}
\alias{crypto.dd.social}
\title{Get social media stats for a coin}
\usage{
crypto.dd.social(symbol)
}
\arguments{
\item{symbol}{(character length 1) Coin to get social stats for}
}
\description{
Wrapper for Python function crypto.dd.social from OpenBB Terminal SDK
}
