% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.stats}
\alias{crypto.dd.stats}
\title{Get 24 hr stats for the product. Volume is in base currency units.}
\usage{
crypto.dd.stats(symbol)
}
\arguments{
\item{symbol}{(character length 1) Trading pair of coins on Coinbase e.g ETH-USDT or UNI-ETH}
}
\description{
Wrapper for Python function crypto.dd.stats from OpenBB Terminal SDK
}
\examples{
crypto.dd.stats(symbol='UNI-ETH')
}
