% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.stats_chart}
\alias{crypto.dd.stats_chart}
\title{Prints table showing 24 hr stats for the product. Volume is in base currency units.}
\usage{
crypto.dd.stats_chart(symbol, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Trading pair of coins on Coinbase e.g ETH-USDT or UNI-ETH}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.stats_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.stats_chart(symbol='UNI-ETH')
}
