% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.team}
\alias{crypto.dd.team}
\title{Returns coin team}
\usage{
crypto.dd.team(symbol)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check team}
}
\description{
Wrapper for Python function crypto.dd.team from OpenBB Terminal SDK
}
\examples{
crypto.dd.team(symbol='ETH')
}
