% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.team_chart}
\alias{crypto.dd.team_chart}
\title{Prints table showing coin team}
\usage{
crypto.dd.team_chart(symbol, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check coin team}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.team_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.team_chart(symbol='ETH')
}
