% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.tk}
\alias{crypto.dd.tk}
\title{Returns coin tokenomics}
\usage{
crypto.dd.tk(symbol, coingecko_id)
}
\arguments{
\item{symbol}{(character length 1) Crypto symbol to check tokenomics}

\item{coingecko_id}{(character length 1) ID from coingecko}
}
\description{
Wrapper for Python function crypto.dd.tk from OpenBB Terminal SDK
}
\examples{
crypto.dd.tk(symbol='ETH')
}
