% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.tokenomics}
\alias{crypto.dd.tokenomics}
\title{Get tokenomics for given coin. \link{Source: CoinGecko}}
\usage{
crypto.dd.tokenomics(symbol = "")
}
\arguments{
\item{symbol}{(character length 1) coin symbol to check tokenomics}
}
\description{
Wrapper for Python function crypto.dd.tokenomics from OpenBB Terminal SDK
}
