% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.trades}
\alias{crypto.dd.trades}
\title{Returns trades for a coin in a given exchange}
\usage{
crypto.dd.trades(exchange_id, symbol, to_symbol)
}
\arguments{
\item{exchange_id}{(character length 1) exchange id}

\item{symbol}{(character length 1) coin symbol}

\item{to_symbol}{(character length 1) currency to compare coin against}
}
\description{
Wrapper for Python function crypto.dd.trades from OpenBB Terminal SDK
}
\examples{
crypto.dd.trades(symbol='AMZN')
}
