% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.trades_chart}
\alias{crypto.dd.trades_chart}
\title{Prints table showing trades for a coin in a given exchange}
\usage{
crypto.dd.trades_chart(
  exchange,
  symbol,
  to_symbol,
  limit = 10,
  export = "",
  sheet_name
)
}
\arguments{
\item{exchange}{(character length 1) exchange id}

\item{symbol}{(character length 1) coin symbol}

\item{to_symbol}{(character length 1) currency to compare coin against}

\item{limit}{(integer length 1) number of trades to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.dd.trades_chart from OpenBB Terminal SDK
}
\examples{
crypto.dd.trades_chart(symbol='AMZN', limit=10)
}
