% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.trading_pair_info}
\alias{crypto.dd.trading_pair_info}
\title{Get information about chosen trading pair. \link{Source: Coinbase}}
\usage{
crypto.dd.trading_pair_info(symbol)
}
\arguments{
\item{symbol}{(character length 1) Trading pair of coins on Coinbase e.g ETH-USDT or UNI-ETH}
}
\description{
Wrapper for Python function crypto.dd.trading_pair_info from OpenBB Terminal SDK
}
\examples{
crypto.dd.trading_pair_info(symbol='UNI-ETH')
}
