% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.trading_pairs}
\alias{crypto.dd.trading_pairs}
\title{Helper method that return all trading pairs on binance. Methods ause this data for input for e.g}
\usage{
crypto.dd.trading_pairs()
}
\description{
Wrapper for Python function crypto.dd.trading_pairs from OpenBB Terminal SDK
}
