% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.dd.twitter}
\alias{crypto.dd.twitter}
\title{Get twitter timeline for given coin id. Not more than last 50 tweets \link{Source: CoinPaprika}}
\usage{
crypto.dd.twitter(symbol = "BTC", sortby = "date", ascend = TRUE)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency symbol (e.g. BTC)}

\item{sortby}{(character length 1) Key by which to sort data. Every column name is valid
(see for possible values:
https://api.coinpaprika.com/docs#tag/Coins/paths/~1coins~1\%7Bcoin_id\%7D~1twitter/get).}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.dd.twitter from OpenBB Terminal SDK
}
\examples{
crypto.dd.twitter(symbol='BTC', sortby='date', ascend=TRUE)
}
