% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.anchor_data}
\alias{crypto.defi.anchor_data}
\title{Returns anchor protocol earnings data of a certain terra address}
\usage{
crypto.defi.anchor_data(address = "")
}
\arguments{
\item{address}{(character length 1) Terra address. Valid terra addresses start with 'terra'}
}
\description{
Wrapper for Python function crypto.defi.anchor_data from OpenBB Terminal SDK
}
