% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.anchor_data_chart}
\alias{crypto.defi.anchor_data_chart}
\title{Plots anchor protocol earnings data of a certain terra address}
\usage{
crypto.defi.anchor_data_chart(
  address = "",
  export = "",
  sheet_name,
  show_transactions = FALSE,
  external_axes
)
}
\arguments{
\item{address}{(character length 1) Terra address. Valid terra addresses start with 'terra'}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}

\item{show_transactions}{(logical length 1) Flag to show history of transactions in Anchor protocol for address. Default False}
}
\description{
Wrapper for Python function crypto.defi.anchor_data_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.anchor_data_chart(show_transactions=FALSE)
}
