% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.aterra}
\alias{crypto.defi.aterra}
\title{Returns historical data of an asset in a certain terra address}
\usage{
crypto.defi.aterra(
  asset = "ust",
  address = "terra1tmnqgvg567ypvsvk6rwsga3srp7e3lg6u0elp8"
)
}
\arguments{
\item{asset}{(character length 1) Terra asset {ust,luna,sdt}}

\item{address}{(character length 1) Terra address. Valid terra addresses start with 'terra'}
}
\description{
Wrapper for Python function crypto.defi.aterra from OpenBB Terminal SDK
}
\examples{
crypto.defi.aterra(asset='ust', address='terra1tmnqgvg567ypvsvk6rwsga3srp7e3lg6u0elp8')
}
