% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.aterra_chart}
\alias{crypto.defi.aterra_chart}
\title{Plots the 30-day history of specified asset in terra address}
\usage{
crypto.defi.aterra_chart(
  asset = "",
  address = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{asset}{(character length 1) Terra asset {ust,luna,sdt}}

\item{address}{(character length 1) Terra address. Valid terra addresses start with 'terra'}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.aterra_chart from OpenBB Terminal SDK
}
