% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.ayr}
\alias{crypto.defi.ayr}
\title{Displays the 30-day history of the Anchor Yield Reserve.}
\usage{
crypto.defi.ayr()
}
\description{
Wrapper for Python function crypto.defi.ayr from OpenBB Terminal SDK
}
