% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.ayr_chart}
\alias{crypto.defi.ayr_chart}
\title{Plots the 30-day history of the Anchor Yield Reserve.}
\usage{
crypto.defi.ayr_chart(export = "", sheet_name, external_axes)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file, by default False}
}
\description{
Wrapper for Python function crypto.defi.ayr_chart from OpenBB Terminal SDK
}
