% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.dtvl}
\alias{crypto.defi.dtvl}
\title{Returns information about historical tvl of a defi protocol.}
\usage{
crypto.defi.dtvl(protocol)
}
\arguments{
\item{protocol}{(character length 1) Name of the protocol}
}
\description{
Wrapper for Python function crypto.defi.dtvl from OpenBB Terminal SDK
}
