% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.dtvl_chart}
\alias{crypto.defi.dtvl_chart}
\title{Plots historical TVL of different dApps}
\usage{
crypto.defi.dtvl_chart(dapps = "", export = "", sheet_name, external_axes)
}
\arguments{
\item{dapps}{(character length 1) dApps to search historical TVL. Should be split by , e.g.: anchor,sushiswap,pancakeswap}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.dtvl_chart from OpenBB Terminal SDK
}
