% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gacc}
\alias{crypto.defi.gacc}
\title{Get terra blockchain account growth history \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.gacc(cumulative = TRUE)
}
\arguments{
\item{cumulative}{(logical length 1) distinguish between periodical and cumulative account growth data}
}
\description{
Wrapper for Python function crypto.defi.gacc from OpenBB Terminal SDK
}
\examples{
crypto.defi.gacc(cumulative=TRUE)
}
