% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gacc_chart}
\alias{crypto.defi.gacc_chart}
\title{Plots terra blockchain account growth history \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.gacc_chart(
  kind = "total",
  cumulative = FALSE,
  limit = 90,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{kind}{(character length 1) display total account count or active account count. One from list \link{active, total}}

\item{cumulative}{(logical length 1) Flag to show cumulative or discrete values. For active accounts only discrete value are available.}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.gacc_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.gacc_chart(limit=90, kind='total', cumulative=FALSE)
}
