% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gdapps}
\alias{crypto.defi.gdapps}
\title{Display top dApps (in terms of TVL) grouped by chain.}
\usage{
crypto.defi.gdapps(limit = 50)
}
\arguments{
\item{limit}{(integer length 1) Number of top dApps to display}
}
\description{
Wrapper for Python function crypto.defi.gdapps from OpenBB Terminal SDK
}
\examples{
crypto.defi.gdapps(limit=50)
}
