% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gdapps_chart}
\alias{crypto.defi.gdapps_chart}
\title{Plots top dApps (in terms of TVL) grouped by chain.}
\usage{
crypto.defi.gdapps_chart(limit = 50, export = "", sheet_name, external_axes)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.gdapps_chart from OpenBB Terminal SDK
}
