% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gov_proposals}
\alias{crypto.defi.gov_proposals}
\title{Get terra blockchain governance proposals list \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.gov_proposals(
  status = "",
  sortby = "id",
  ascend = TRUE,
  limit = 10
)
}
\arguments{
\item{status}{(character length 1) status of proposal, one from list: \link{'Voting','Deposit','Passed','Rejected'}}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{limit}{(integer length 1) Number of records to display}
}
\description{
Wrapper for Python function crypto.defi.gov_proposals from OpenBB Terminal SDK
}
\examples{
crypto.defi.gov_proposals(sortby='id', ascend=TRUE, limit=10)
}
