% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.gov_proposals_chart}
\alias{crypto.defi.gov_proposals_chart}
\title{Prints table showing terra blockchain governance proposals list \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.gov_proposals_chart(
  limit = 10,
  status = "all",
  sortby = "id",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{status}{(character length 1) status of proposal, one from list: \link{'Voting','Deposit','Passed','Rejected'}}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascend}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.gov_proposals_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.gov_proposals_chart(limit=10, status='all', sortby='id', ascend=TRUE)
}
