% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.luna_supply_chart}
\alias{crypto.defi.luna_supply_chart}
\title{Plots and prints table showing Luna circulating supply stats}
\usage{
crypto.defi.luna_supply_chart(
  days = 30,
  export = "",
  sheet_name,
  supply_type = "lunaSupplyChallengeStats",
  limit = 5,
  external_axes
)
}
\arguments{
\item{days}{(integer length 1) Number of days}

\item{export}{(character length 1) Export type}

\item{supply_type}{(character length 1) Supply type to unpack json}

\item{limit}{(integer length 1) Number of results display on the terminal
Default: 5}
}
\description{
Wrapper for Python function crypto.defi.luna_supply_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.luna_supply_chart(days=30, supply_type='lunaSupplyChallengeStats', limit=5)
}
