% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.newsletters}
\alias{crypto.defi.newsletters}
\title{Scrape all substack newsletters from url list.}
\usage{
crypto.defi.newsletters()
}
\description{
Wrapper for Python function crypto.defi.newsletters from OpenBB Terminal SDK
}
