% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.newsletters_chart}
\alias{crypto.defi.newsletters_chart}
\title{Prints table showing DeFi related substack newsletters.}
\usage{
crypto.defi.newsletters_chart(limit = 10, export = "", sheet_name)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.newsletters_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.newsletters_chart(limit=10)
}
