% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.pairs}
\alias{crypto.defi.pairs}
\title{Get lastly added trade-able pairs on Uniswap with parameters like:}
\usage{
crypto.defi.pairs(
  last_days = 14,
  min_volume = 100,
  min_liquidity = 0,
  min_tx = 100
)
}
\arguments{
\item{last_days}{(integer length 1) How many days back to look for added pairs.}

\item{min_volume}{(integer length 1) Minimum volume}

\item{min_liquidity}{(integer length 1) Minimum liquidity}

\item{min_tx}{(integer length 1) Minimum number of transactions done in given pool.}
}
\description{
Wrapper for Python function crypto.defi.pairs from OpenBB Terminal SDK
}
\examples{
crypto.defi.pairs(last_days=14, min_volume=100, min_liquidity=0, min_tx=100)
}
