% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.pairs_chart}
\alias{crypto.defi.pairs_chart}
\title{Prints table showing Lastly added pairs on Uniswap DEX.}
\usage{
crypto.defi.pairs_chart(
  limit = 20,
  days = 7,
  min_volume = 20,
  min_liquidity = 0,
  min_tx = 100,
  sortby = "created",
  ascend = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{days}{(integer length 1) Number of days the pair has been active,}

\item{min_volume}{(integer length 1) Minimum trading volume,}

\item{min_liquidity}{(integer length 1) Minimum liquidity}

\item{min_tx}{(integer length 1) Minimum number of transactions}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.pairs_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.pairs_chart(limit=20, days=7, min_volume=20, min_liquidity=0, min_tx=100, sortby='created', ascend=FALSE)
}
