% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.pools}
\alias{crypto.defi.pools}
\title{Get uniswap pools by volume. \link{Source: https://thegraph.com/en/}}
\usage{
crypto.defi.pools()
}
\description{
Wrapper for Python function crypto.defi.pools from OpenBB Terminal SDK
}
