% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.sinfo}
\alias{crypto.defi.sinfo}
\title{Get staking info for provided terra account \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.sinfo(address = "")
}
\arguments{
\item{address}{(character length 1) terra blockchain address e.g. terra1jvwelvs7rdk6j3mqdztq5tya99w8lxk6l9hcqg}
}
\description{
Wrapper for Python function crypto.defi.sinfo from OpenBB Terminal SDK
}
