% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.sinfo_chart}
\alias{crypto.defi.sinfo_chart}
\title{Prints table showing staking info for provided terra account address \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.sinfo_chart(address = "", limit = 10, export = "", sheet_name)
}
\arguments{
\item{address}{(character length 1) terra blockchain address e.g. terra1jvwelvs7rdk6j3mqdztq5tya99w8lxk6l9hcqg}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.sinfo_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.sinfo_chart(limit=10)
}
