% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.sratio}
\alias{crypto.defi.sratio}
\title{Get terra blockchain staking ratio history \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.sratio(limit = 200)
}
\arguments{
\item{limit}{(integer length 1) The number of ratios to show}
}
\description{
Wrapper for Python function crypto.defi.sratio from OpenBB Terminal SDK
}
\examples{
crypto.defi.sratio(limit=200)
}
