% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.sratio_chart}
\alias{crypto.defi.sratio_chart}
\title{Plots terra blockchain staking ratio history \link{Source: https://fcd.terra.dev/v1}}
\usage{
crypto.defi.sratio_chart(limit = 90, export = "", sheet_name, external_axes)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.sratio_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.sratio_chart(limit=90)
}
