% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.sreturn}
\alias{crypto.defi.sreturn}
\title{Get terra blockchain staking returns history \link{Source: https://fcd.terra.dev/v1}}
\usage{
crypto.defi.sreturn(limit = 200)
}
\arguments{
\item{limit}{(integer length 1) The number of returns to show}
}
\description{
Wrapper for Python function crypto.defi.sreturn from OpenBB Terminal SDK
}
\examples{
crypto.defi.sreturn(limit=200)
}
