% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.stats}
\alias{crypto.defi.stats}
\title{Get base statistics about Uniswap DEX. \link{Source: https://thegraph.com/en/}}
\usage{
crypto.defi.stats()
}
\description{
Wrapper for Python function crypto.defi.stats from OpenBB Terminal SDK
}
