% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.stats_chart}
\alias{crypto.defi.stats_chart}
\title{Prints table showing base statistics about Uniswap DEX. \link{Source: https://thegraph.com/en/}}
\usage{
crypto.defi.stats_chart(export = "", sheet_name)
}
\description{
Wrapper for Python function crypto.defi.stats_chart from OpenBB Terminal SDK
}
