% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.stvl}
\alias{crypto.defi.stvl}
\title{Returns historical values of the total sum of TVLs from all listed protocols.}
\usage{
crypto.defi.stvl()
}
\description{
Wrapper for Python function crypto.defi.stvl from OpenBB Terminal SDK
}
