% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.stvl_chart}
\alias{crypto.defi.stvl_chart}
\title{Plots historical values of the total sum of TVLs from all listed protocols.}
\usage{
crypto.defi.stvl_chart(limit = 5, export = "", sheet_name, external_axes)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display, by default 5}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.stvl_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.stvl_chart(limit=5)
}
