% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.swaps}
\alias{crypto.defi.swaps}
\title{Get the last 100 swaps done on Uniswap \link{Source: https://thegraph.com/en/}}
\usage{
crypto.defi.swaps(limit = 100, sortby = "timestamp", ascend = FALSE)
}
\arguments{
\item{limit}{(integer length 1) Number of swaps to return. Maximum possible number: 1000.}

\item{sortby}{(character length 1) Key by which to sort data. The table can be sorted by every of its columns
(see https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v2).}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.defi.swaps from OpenBB Terminal SDK
}
\examples{
crypto.defi.swaps(limit=100, sortby='timestamp', ascend=FALSE)
}
