% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.swaps_chart}
\alias{crypto.defi.swaps_chart}
\title{Prints table showing last swaps done on Uniswap}
\usage{
crypto.defi.swaps_chart(
  limit = 10,
  sortby = "timestamp",
  ascend = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data. The table can be sorted by every of its columns
(see https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v2).}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.swaps_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.swaps_chart(limit=10, sortby='timestamp', ascend=FALSE)
}
