% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.tokens}
\alias{crypto.defi.tokens}
\title{Get list of tokens trade-able on Uniswap DEX. \link{Source: https://thegraph.com/en/}}
\usage{
crypto.defi.tokens(skip = 0, limit = 100, sortby = "index", ascend = FALSE)
}
\arguments{
\item{skip}{(integer length 1) Skip n number of records.}

\item{limit}{(integer length 1) Show n number of records.}

\item{sortby}{(character length 1) The column to sort by}

\item{ascend}{(logical length 1) Whether to sort in ascending order}
}
\description{
Wrapper for Python function crypto.defi.tokens from OpenBB Terminal SDK
}
\examples{
crypto.defi.tokens(skip=0, limit=100, sortby='index', ascend=FALSE)
}
