% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.tokens_chart}
\alias{crypto.defi.tokens_chart}
\title{Prints table showing tokens trade-able on Uniswap DEX.}
\usage{
crypto.defi.tokens_chart(
  skip = 0,
  limit = 20,
  sortby = "index",
  ascend = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{skip}{(integer length 1) Number of records to skip}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.tokens_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.tokens_chart(skip=0, limit=20, sortby='index', ascend=FALSE)
}
