% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.validators}
\alias{crypto.defi.validators}
\title{Get information about terra validators \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.validators(sortby = "votingPower", ascend = TRUE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data. Choose from:
validatorName, tokensAmount, votingPower, commissionRate, status, uptime}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.defi.validators from OpenBB Terminal SDK
}
