% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.validators_chart}
\alias{crypto.defi.validators_chart}
\title{Prints table showing information about terra validators \link{Source: https://fcd.terra.dev/swagger}}
\usage{
crypto.defi.validators_chart(
  limit = 10,
  sortby = "votingPower",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data. Choose from:
validatorName, tokensAmount, votingPower, commissionRate, status, uptime}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.validators_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.validators_chart(limit=10, sortby='votingPower', ascend=TRUE)
}
