% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.vaults}
\alias{crypto.defi.vaults}
\title{Get DeFi Vaults Information. DeFi Vaults are pools of funds with an assigned strategy which main goal is to}
\usage{
crypto.defi.vaults(chain, protocol, kind, ascend = TRUE, sortby = "apy")
}
\description{
Wrapper for Python function crypto.defi.vaults from OpenBB Terminal SDK
}
