% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.defi.vaults_chart}
\alias{crypto.defi.vaults_chart}
\title{Prints table showing Top DeFi Vaults - pools of funds with an assigned strategy which main goal is to}
\usage{
crypto.defi.vaults_chart(
  chain,
  protocol,
  kind,
  limit = 10,
  sortby = "apy",
  ascend = TRUE,
  link = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{link}{(logical length 1) Flag to show links}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.defi.vaults_chart from OpenBB Terminal SDK
}
\examples{
crypto.defi.vaults_chart(limit=10, sortby='apy', ascend=TRUE, link=FALSE)
}
