% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.categories_keys}
\alias{crypto.disc.categories_keys}
\title{Get list of categories keys}
\usage{
crypto.disc.categories_keys()
}
\description{
Wrapper for Python function crypto.disc.categories_keys from OpenBB Terminal SDK
}
