% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.coin_list}
\alias{crypto.disc.coin_list}
\title{Get list of coins available on CoinGecko \link{Source: CoinGecko}}
\usage{
crypto.disc.coin_list()
}
\description{
Wrapper for Python function crypto.disc.coin_list from OpenBB Terminal SDK
}
