% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.coins}
\alias{crypto.disc.coins}
\title{Get N coins from CoinGecko \link{Source: CoinGecko}}
\usage{
crypto.disc.coins(
  limit = 250,
  category = "",
  sortby = "Symbol",
  ascend = FALSE
)
}
\arguments{
\item{limit}{(integer length 1) Number of top coins to grab from CoinGecko}

\item{category}{(character length 1) Category of the coins we want to retrieve}

\item{sortby}{(character length 1) Key to sort data}

\item{ascend}{(logical length 1) Sort data in ascending order}
}
\description{
Wrapper for Python function crypto.disc.coins from OpenBB Terminal SDK
}
\examples{
crypto.disc.coins(limit=250, sortby='Symbol', ascend=FALSE)
}
