% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.coins_chart}
\alias{crypto.disc.coins_chart}
\title{Prints table showing top coins \link{Source: CoinGecko}}
\usage{
crypto.disc.coins_chart(
  category,
  limit = 250,
  sortby = "Symbol",
  export = "",
  sheet_name,
  ascend = FALSE
)
}
\arguments{
\item{category}{(character length 1) If no category is passed it will search for all coins. (E.g., smart-contract-platform)}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key to sort data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}

\item{ascend}{(logical length 1) Sort data in ascending order}
}
\description{
Wrapper for Python function crypto.disc.coins_chart from OpenBB Terminal SDK
}
\examples{
crypto.disc.coins_chart(limit=250, sortby='Symbol', ascend=FALSE)
}
