% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.coins_for_given_exchange}
\alias{crypto.disc.coins_for_given_exchange}
\title{Helper method to get all coins available on binance exchange \link{Source: CoinGecko}}
\usage{
crypto.disc.coins_for_given_exchange(exchange_id = "binance", page = 1)
}
\arguments{
\item{exchange_id}{(character length 1) id of exchange}

\item{page}{(integer length 1) number of page. One page contains 100 records}
}
\description{
Wrapper for Python function crypto.disc.coins_for_given_exchange from OpenBB Terminal SDK
}
\examples{
crypto.disc.coins_for_given_exchange(exchange_id='binance', page=1)
}
