% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.cpsearch}
\alias{crypto.disc.cpsearch}
\title{Search CoinPaprika. \link{Source: CoinPaprika}}
\usage{
crypto.disc.cpsearch(query, category, modifier, sortby = "id", ascend = TRUE)
}
\arguments{
\item{query}{(character length 1) phrase for search}

\item{sortby}{(character length 1) Key to sort data. The table can be sorted by every of its columns. Refer to
API documentation (see https://api.coinpaprika.com/docs#tag/Tools/paths/~1search/get)}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.disc.cpsearch from OpenBB Terminal SDK
}
\examples{
crypto.disc.cpsearch(sortby='id', ascend=TRUE)
}
