% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.cpsearch_chart}
\alias{crypto.disc.cpsearch_chart}
\title{Prints table showing Search over CoinPaprika. \link{Source: CoinPaprika}}
\usage{
crypto.disc.cpsearch_chart(
  query,
  category = "all",
  limit = 10,
  sortby = "id",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{query}{(character length 1) Search query}

\item{category}{(character length 1) Categories to search: currencies|exchanges|icos|people|tags|all. Default: all}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key to sort data. The table can be sorted by every of its columns. Refer to
API documentation (see https://api.coinpaprika.com/docs#tag/Tools/paths/~1search/get)}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.disc.cpsearch_chart from OpenBB Terminal SDK
}
\examples{
crypto.disc.cpsearch_chart(category='all', limit=10, sortby='id', ascend=TRUE)
}
