% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.gainers_chart}
\alias{crypto.disc.gainers_chart}
\title{Prints table showing Largest Gainers - coins which gain the most in given period. \link{Source: CoinGecko}}
\usage{
crypto.disc.gainers_chart(
  interval = "1h",
  limit = 20,
  sortby = "market_cap_rank",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{interval}{(character length 1) Time period by which data is displayed. One from \link{1h, 24h, 7d, 14d, 30d, 60d, 1y}}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key to sort data. The table can be sorted by every of its columns. Refer to
API documentation (see /coins/markets in https://www.coingecko.com/en/api/documentation)}

\item{ascend}{(logical length 1) Sort data in ascending order}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.disc.gainers_chart from OpenBB Terminal SDK
}
\examples{
crypto.disc.gainers_chart(interval='1h', limit=20, sortby='market_cap_rank')
}
