% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.losers}
\alias{crypto.disc.losers}
\title{Shows Largest Losers - coins which lose the most in given period. \link{Source: CoinGecko}}
\usage{
crypto.disc.losers(
  interval = "1h",
  limit = 50,
  sortby = "market_cap_rank",
  ascend = TRUE
)
}
\arguments{
\item{interval}{(character length 1) Time interval by which data is displayed. One from \link{1h, 24h, 7d, 14d, 30d, 60d, 1y}}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key to sort data. The table can be sorted by every of its columns. Refer to
API documentation (see /coins/markets in https://www.coingecko.com/en/api/documentation)}

\item{ascend}{(logical length 1) Sort data in ascending order}
}
\description{
Wrapper for Python function crypto.disc.losers from OpenBB Terminal SDK
}
\examples{
crypto.disc.losers(interval='1h', limit=50, sortby='market_cap_rank')
}
