% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_coins}
\alias{crypto.disc.top_coins}
\title{Get top cryptp coins.}
\usage{
crypto.disc.top_coins(source = "CoinGecko", limit = 10)
}
\arguments{
\item{source}{(character length 1) Source of data, by default "CoinGecko"}

\item{limit}{(integer length 1) Number of coins to return, by default 10}
}
\description{
Wrapper for Python function crypto.disc.top_coins from OpenBB Terminal SDK
}
\examples{
crypto.disc.top_coins(source='CoinGecko', limit=10)
}
