% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_dapps}
\alias{crypto.disc.top_dapps}
\title{Get top decentralized applications by daily volume and users \link{Source: https://dappradar.com/}}
\usage{
crypto.disc.top_dapps(sortby = "", limit = 10)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}
}
\description{
Wrapper for Python function crypto.disc.top_dapps from OpenBB Terminal SDK
}
