% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_dexes_chart}
\alias{crypto.disc.top_dexes_chart}
\title{Prints table showing top decentralized exchanges \link{Source: https://dappradar.com/}}
\usage{
crypto.disc.top_dexes_chart(limit = 10, export = "", sheet_name, sortby = "")
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}

\item{sortby}{(character length 1) Key by which to sort data}
}
\description{
Wrapper for Python function crypto.disc.top_dexes_chart from OpenBB Terminal SDK
}
\examples{
crypto.disc.top_dexes_chart(limit=10)
}
