% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_games}
\alias{crypto.disc.top_games}
\title{Get top blockchain games by daily volume and users \link{Source: https://dappradar.com/}}
\usage{
crypto.disc.top_games(sortby = "", limit = 10)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{limit}{(integer length 1) Number of records to display}
}
\description{
Wrapper for Python function crypto.disc.top_games from OpenBB Terminal SDK
}
\examples{
crypto.disc.top_games(limit=10)
}
