% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_nfts}
\alias{crypto.disc.top_nfts}
\title{Get top nft collections \link{Source: https://dappradar.com/}}
\usage{
crypto.disc.top_nfts(sortby = "", limit = 10)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}
}
\description{
Wrapper for Python function crypto.disc.top_nfts from OpenBB Terminal SDK
}
