% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.top_nfts_chart}
\alias{crypto.disc.top_nfts_chart}
\title{Prints table showing top nft collections \link{Source: https://dappradar.com/}}
\usage{
crypto.disc.top_nfts_chart(limit = 10, sortby = "", export = "", sheet_name)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.disc.top_nfts_chart from OpenBB Terminal SDK
}
\examples{
crypto.disc.top_nfts_chart(limit=10)
}
