% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.trending}
\alias{crypto.disc.trending}
\title{Returns trending coins \link{Source: CoinGecko}}
\usage{
crypto.disc.trending()
}
\description{
Wrapper for Python function crypto.disc.trending from OpenBB Terminal SDK
}
