% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.disc.trending_chart}
\alias{crypto.disc.trending_chart}
\title{Prints table showing trending coins \link{Source: CoinGecko}}
\usage{
crypto.disc.trending_chart(export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.disc.trending_chart from OpenBB Terminal SDK
}
