% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.find}
\alias{crypto.find}
\title{Find similar coin by coin name,symbol or id.}
\usage{
crypto.find(query, source = "CoinGecko", key = "symbol", limit = 10)
}
\arguments{
\item{query}{(character length 1) Cryptocurrency}

\item{source}{(character length 1) Data source of coins.  CoinGecko or CoinPaprika or Binance or Coinbase}

\item{key}{(character length 1) Searching key (symbol, id, name)}

\item{limit}{(integer length 1) Number of records to display}
}
\description{
Wrapper for Python function crypto.find from OpenBB Terminal SDK
}
\examples{
crypto.find(source='CoinGecko', key='symbol', limit=10)
}
