% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.load}
\alias{crypto.load}
\title{Load crypto currency to get data for}
\usage{
crypto.load(
  symbol,
  start_date,
  interval,
  exchange = "binance",
  to_symbol = "usd",
  end_date,
  source = "CoinGecko"
)
}
\arguments{
\item{symbol}{(character length 1) Coin to get}

\item{exchange}{(character length 1) The exchange to get data from.}

\item{to_symbol}{(character length 1) Quote Currency (Defaults to usdt)}

\item{source}{(character length 1) The source of the data
Choose from: CCXT, CoinGecko, YahooFinance}
}
\description{
Wrapper for Python function crypto.load from OpenBB Terminal SDK
}
\examples{
crypto.load(symbol='AMZN', interval=1440, exchange='binance', vs_currency='usdt', source='CCXT')
}
