% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.collections}
\alias{crypto.nft.collections}
\title{Get nft collections \link{Source: https://nftpricefloor.com/}}
\usage{
crypto.nft.collections()
}
\description{
Wrapper for Python function crypto.nft.collections from OpenBB Terminal SDK
}
