% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.collections_chart}
\alias{crypto.nft.collections_chart}
\title{Display NFT collections. \link{Source: https://nftpricefloor.com/}}
\usage{
crypto.nft.collections_chart(
  show_fp = FALSE,
  show_sales = FALSE,
  limit = 5,
  export = "",
  sheet_name
)
}
\arguments{
\item{show_fp}{(logical length 1) Show NFT Price Floor for top collections}

\item{limit}{(integer length 1) Number of NFT collections to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.nft.collections_chart from OpenBB Terminal SDK
}
\examples{
crypto.nft.collections_chart(show_fp=FALSE, limit=5)
}
