% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.fp}
\alias{crypto.nft.fp}
\title{Get nft collections \link{Source: https://nftpricefloor.com/}}
\usage{
crypto.nft.fp(slug)
}
\arguments{
\item{slug}{(character length 1) nft collection slug}
}
\description{
Wrapper for Python function crypto.nft.fp from OpenBB Terminal SDK
}
