% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.fp_chart}
\alias{crypto.nft.fp_chart}
\title{Display NFT collection floor price over time. \link{Source: https://nftpricefloor.com/}}
\usage{
crypto.nft.fp_chart(
  slug,
  limit = 10,
  export = "",
  sheet_name,
  external_axes,
  raw = FALSE
)
}
\arguments{
\item{slug}{(character length 1) NFT collection slug}

\item{limit}{(integer length 1) Number of raw data to show}

\item{export}{(character length 1) Format to export data}

\item{raw}{(logical length 1) Flag to display raw data}
}
\description{
Wrapper for Python function crypto.nft.fp_chart from OpenBB Terminal SDK
}
\examples{
crypto.nft.fp_chart(raw=FALSE, limit=10)
}
