% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.stats}
\alias{crypto.nft.stats}
\title{Get stats of a nft collection \link{Source: opensea.io}}
\usage{
crypto.nft.stats(slug)
}
\arguments{
\item{slug}{(character length 1) Opensea collection slug. If the name of the collection is Mutant Ape Yacht Club the slug is mutant-ape-yacht-club}
}
\description{
Wrapper for Python function crypto.nft.stats from OpenBB Terminal SDK
}
