% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.nft.stats_chart}
\alias{crypto.nft.stats_chart}
\title{Prints table showing collection stats. \link{Source: opensea.io}}
\usage{
crypto.nft.stats_chart(slug, export, sheet_name)
}
\arguments{
\item{slug}{(character length 1) Opensea collection slug.
If the name of the collection is Mutant Ape Yacht Club the slug is mutant-ape-yacht-club}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.nft.stats_chart from OpenBB Terminal SDK
}
