% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.baas}
\alias{crypto.onchain.baas}
\title{Get an average bid and ask prices, average spread for given crypto pair for chosen time period.}
\usage{
crypto.onchain.baas(
  symbol = "WETH",
  to_symbol = "USDT",
  limit = 10,
  sortby = "date",
  ascend = TRUE
)
}
\arguments{
\item{symbol}{(character length 1) ERC20 token symbol}

\item{to_symbol}{(character length 1) Quoted currency.}

\item{limit}{(integer length 1) Last n days to query data}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.baas from OpenBB Terminal SDK
}
\examples{
crypto.onchain.baas(symbol='WETH', to_symbol='USDT', limit=10, sortby='date', ascend=TRUE)
}
