% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.btc_supply}
\alias{crypto.onchain.btc_supply}
\title{Returns BTC circulating supply \link{Source: https://api.blockchain.info/}}
\usage{
crypto.onchain.btc_supply()
}
\description{
Wrapper for Python function crypto.onchain.btc_supply from OpenBB Terminal SDK
}
