% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.btc_transac}
\alias{crypto.onchain.btc_transac}
\title{Returns BTC confirmed transactions \link{Source: https://api.blockchain.info/}}
\usage{
crypto.onchain.btc_transac()
}
\description{
Wrapper for Python function crypto.onchain.btc_transac from OpenBB Terminal SDK
}
