% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.btcsingleblock}
\alias{crypto.onchain.btcsingleblock}
\title{Returns BTC block data in json format. \link{Source: https://blockchain.info/}}
\usage{
crypto.onchain.btcsingleblock(blockhash)
}
\description{
Wrapper for Python function crypto.onchain.btcsingleblock from OpenBB Terminal SDK
}
