% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.btcsingleblock_chart}
\alias{crypto.onchain.btcsingleblock_chart}
\title{Returns BTC block data. \link{Source: https://api.blockchain.info/}}
\usage{
crypto.onchain.btcsingleblock_chart(blockhash, export = "", sheet_name)
}
\arguments{
\item{blockhash}{(character length 1) Hash of the block you are looking for.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.btcsingleblock_chart from OpenBB Terminal SDK
}
