% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.dex_trades_monthly}
\alias{crypto.onchain.dex_trades_monthly}
\title{Get list of trades on Decentralized Exchanges monthly aggregated.}
\usage{
crypto.onchain.dex_trades_monthly(
  trade_amount_currency = "USD",
  limit = 90,
  ascend = TRUE
)
}
\arguments{
\item{trade_amount_currency}{(character length 1) Currency of displayed trade amount. Default: USD}

\item{limit}{(integer length 1) Last n days to query data. Maximum 365 (bigger numbers can cause timeouts
on server side)}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.dex_trades_monthly from OpenBB Terminal SDK
}
\examples{
crypto.onchain.dex_trades_monthly(trade_amount_currency='USD', limit=90, ascend=TRUE)
}
