% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.dvcp}
\alias{crypto.onchain.dvcp}
\title{Get daily volume for given pair \link{Source: https://graphql.bitquery.io/}}
\usage{
crypto.onchain.dvcp(
  limit = 100,
  symbol = "UNI",
  to_symbol = "USDT",
  sortby = "date",
  ascend = TRUE
)
}
\arguments{
\item{limit}{(integer length 1) Last n days to query data}

\item{symbol}{(character length 1) ERC20 token symbol}

\item{to_symbol}{(character length 1) Quote currency.}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.dvcp from OpenBB Terminal SDK
}
\examples{
crypto.onchain.dvcp(limit=100, symbol='UNI', to_symbol='USDT', sortby='date', ascend=TRUE)
}
