% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.erc20_tokens}
\alias{crypto.onchain.erc20_tokens}
\title{Helper method that loads ~1500 most traded erc20 token.}
\usage{
crypto.onchain.erc20_tokens()
}
\description{
Wrapper for Python function crypto.onchain.erc20_tokens from OpenBB Terminal SDK
}
