% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.gwei}
\alias{crypto.onchain.gwei}
\title{Returns the most recent Ethereum gas fees in gwei}
\usage{
crypto.onchain.gwei()
}
\description{
Wrapper for Python function crypto.onchain.gwei from OpenBB Terminal SDK
}
