% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.gwei_chart}
\alias{crypto.onchain.gwei_chart}
\title{Current gwei fees}
\usage{
crypto.onchain.gwei_chart(export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.gwei_chart from OpenBB Terminal SDK
}
