% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.hist_chart}
\alias{crypto.onchain.hist_chart}
\title{Display information about balance historical transactions. \link{Source: Ethplorer}}
\usage{
crypto.onchain.hist_chart(
  address,
  limit = 10,
  sortby = "timestamp",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{address}{(character length 1) Ethereum blockchain balance e.g. 0x3cD751E6b0078Be393132286c442345e5DC49699}

\item{limit}{(integer length 1) Limit of transactions. Maximum 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in ascending order.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.hist_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.hist_chart(limit=10, sortby='timestamp', ascend='True')
}
