% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.holders}
\alias{crypto.onchain.holders}
\title{Get info about top token holders. \link{Source: Ethplorer}}
\usage{
crypto.onchain.holders(address, sortby = "balance", ascend = TRUE)
}
\arguments{
\item{address}{(character length 1) Token balance e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}
}
\description{
Wrapper for Python function crypto.onchain.holders from OpenBB Terminal SDK
}
\examples{
crypto.onchain.holders(sortby='balance', ascend='True')
}
