% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.holders_chart}
\alias{crypto.onchain.holders_chart}
\title{Display info about top ERC20 token holders. \link{Source: Ethplorer}}
\usage{
crypto.onchain.holders_chart(
  address,
  limit = 10,
  sortby = "balance",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{address}{(character length 1) Token balance e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984}

\item{limit}{(integer length 1) Limit of transactions. Maximum 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.holders_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.holders_chart(limit=10, sortby='balance', ascend='True')
}
