% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.hr}
\alias{crypto.onchain.hr}
\title{Returns dataframe with mean hashrate of btc or eth blockchain and symbol price}
\usage{
crypto.onchain.hr(
  symbol,
  interval = "24h",
  start_date = "2010-01-01",
  end_date
)
}
\arguments{
\item{symbol}{(character length 1) Blockchain to check hashrate (BTC or ETH)}

\item{interval}{(character length 1) Interval frequency (e.g., 24h)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.onchain.hr from OpenBB Terminal SDK
}
\examples{
crypto.onchain.hr(symbol='AMZN', start_date='2010-01-01', interval='24h')
}
