% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.hr_chart}
\alias{crypto.onchain.hr_chart}
\title{Plots dataframe with mean hashrate of btc or eth blockchain and symbol price.}
\usage{
crypto.onchain.hr_chart(
  symbol,
  start_date = "2010-01-01",
  end_date,
  interval = "24h",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Blockchain to check mean hashrate (BTC or ETH)}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{interval}{(character length 1) Interval frequency (possible values are: 24, 1w, 1month)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.hr_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.hr_chart(symbol='AMZN', start_date='2010-01-01', interval='24h')
}
