% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.info}
\alias{crypto.onchain.info}
\title{Get info about ERC20 token. \link{Source: Ethplorer}}
\usage{
crypto.onchain.info(address)
}
\arguments{
\item{address}{(Python type: inspect._empty) Token balance e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984}
}
\description{
Wrapper for Python function crypto.onchain.info from OpenBB Terminal SDK
}
