% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.info_chart}
\alias{crypto.onchain.info_chart}
\title{Display info about ERC20 token. \link{Source: Ethplorer}}
\usage{
crypto.onchain.info_chart(address, social = FALSE, export = "", sheet_name)
}
\arguments{
\item{address}{(character length 1) Token balance e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984}

\item{social}{(logical length 1) Flag to display social media links}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.info_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.info_chart(social=FALSE)
}
