% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.lt}
\alias{crypto.onchain.lt}
\title{Get trades on Decentralized Exchanges aggregated by DEX \link{Source: https://graphql.bitquery.io/}}
\usage{
crypto.onchain.lt(
  trade_amount_currency = "USD",
  limit = 90,
  sortby = "tradeAmount",
  ascend = TRUE
)
}
\arguments{
\item{trade_amount_currency}{(character length 1) Currency of displayed trade amount. Default: USD}

\item{limit}{(integer length 1) Last n days to query data. Maximum 365 (bigger numbers can cause timeouts
on server side)}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.lt from OpenBB Terminal SDK
}
\examples{
crypto.onchain.lt(trade_amount_currency='USD', limit=90, sortby='tradeAmount', ascend=TRUE)
}
