% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.lt_chart}
\alias{crypto.onchain.lt_chart}
\title{Prints table showing Trades on Decentralized Exchanges aggregated by DEX or Month}
\usage{
crypto.onchain.lt_chart(
  trade_amount_currency = "USD",
  kind = "dex",
  limit = 20,
  days = 90,
  sortby = "tradeAmount",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{trade_amount_currency}{(character length 1) Currency of displayed trade amount. Default: USD}

\item{kind}{(character length 1) Aggregate trades by dex or time}

\item{limit}{(integer length 1) Number of records to display}

\item{days}{(integer length 1) Last n days to query data. Maximum 365 (bigger numbers can cause timeouts
on server side)}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.lt_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.lt_chart(kind='dex', trade_amount_currency='USD', limit=20, sortby='tradeAmount', ascend=TRUE, days=90)
}
