% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.prices}
\alias{crypto.onchain.prices}
\title{Get token historical prices with volume and market cap, and average price. \link{Source: Ethplorer}}
\usage{
crypto.onchain.prices(address, sortby = "date", ascend = FALSE)
}
\arguments{
\item{address}{(character length 1) Token e.g. 0xf3db5fa2c66b7af3eb0c0b782510816cbe4813b8}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}
}
\description{
Wrapper for Python function crypto.onchain.prices from OpenBB Terminal SDK
}
\examples{
crypto.onchain.prices(sortby='date', ascend='False')
}
