% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.query_graph}
\alias{crypto.onchain.query_graph}
\title{Helper methods for querying graphql api. \link{Source: https://bitquery.io/}}
\usage{
crypto.onchain.query_graph(url, query)
}
\arguments{
\item{url}{(character length 1) Endpoint url}

\item{query}{(character length 1) Graphql query}
}
\description{
Wrapper for Python function crypto.onchain.query_graph from OpenBB Terminal SDK
}
