% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.th}
\alias{crypto.onchain.th}
\title{Get info about token historical transactions. \link{Source: Ethplorer}}
\usage{
crypto.onchain.th(address, sortby = "timestamp", ascend = FALSE)
}
\arguments{
\item{address}{(character length 1) Token e.g. 0xf3db5fa2c66b7af3eb0c0b782510816cbe4813b8}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}
}
\description{
Wrapper for Python function crypto.onchain.th from OpenBB Terminal SDK
}
\examples{
crypto.onchain.th(sortby='timestamp', ascend='False')
}
