% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.th_chart}
\alias{crypto.onchain.th_chart}
\title{Display info about token history. \link{Source: Ethplorer}}
\usage{
crypto.onchain.th_chart(
  address,
  limit = 10,
  sortby = "timestamp",
  ascend = FALSE,
  hash_ = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{address}{(character length 1) Token balance e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984}

\item{limit}{(integer length 1) Limit of transactions. Maximum 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}

\item{hash_}{(logical length 1) Flag to show transaction hash.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.th_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.th_chart(limit=10, sortby='timestamp', ascend='False', hash_=FALSE)
}
