% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.token_decimals}
\alias{crypto.onchain.token_decimals}
\title{Helper methods that gets token decimals number. \link{Source: Ethplorer}}
\usage{
crypto.onchain.token_decimals(address)
}
\arguments{
\item{address}{(character length 1) Blockchain balance e.g. 0x1f9840a85d5af5bf1d1762f925bdaddc4201f984}
}
\description{
Wrapper for Python function crypto.onchain.token_decimals from OpenBB Terminal SDK
}
