% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.top}
\alias{crypto.onchain.top}
\title{Get top 50 tokens. \link{Source: Ethplorer}}
\usage{
crypto.onchain.top(sortby = "rank", ascend = FALSE)
}
\arguments{
\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}
}
\description{
Wrapper for Python function crypto.onchain.top from OpenBB Terminal SDK
}
\examples{
crypto.onchain.top(sortby='rank', ascend='False')
}
