% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.top_chart}
\alias{crypto.onchain.top_chart}
\title{Display top ERC20 tokens \link{Source: Ethplorer}}
\usage{
crypto.onchain.top_chart(
  limit = 15,
  sortby = "rank",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Limit of transactions. Maximum 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in descending order.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.top_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.top_chart(limit=15, sortby='rank', ascend='True')
}
