% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.ttcp_chart}
\alias{crypto.onchain.ttcp_chart}
\title{Prints table showing most traded crypto pairs on given decentralized exchange in chosen time period.}
\usage{
crypto.onchain.ttcp_chart(
  exchange = "Uniswap",
  days = 10,
  limit = 10,
  sortby = "tradeAmount",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{exchange}{(character length 1) Decentralized exchange name}

\item{days}{(integer length 1) Number of days taken into calculation account.}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.ttcp_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.ttcp_chart(exchange='Uniswap', days=10, sortby='tradeAmount', ascend=TRUE)
}
