% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.tv}
\alias{crypto.onchain.tv}
\title{Get token volume on different Decentralized Exchanges. \link{Source: https://graphql.bitquery.io/}}
\usage{
crypto.onchain.tv(
  symbol = "UNI",
  trade_amount_currency = "USD",
  sortby = "tradeAmount",
  ascend = TRUE
)
}
\arguments{
\item{symbol}{(character length 1) ERC20 token symbol.}

\item{trade_amount_currency}{(character length 1) Currency to display trade amount in.}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.tv from OpenBB Terminal SDK
}
\examples{
crypto.onchain.tv(symbol='UNI', trade_amount_currency='USD', sortby='tradeAmount', ascend=TRUE)
}
