% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.tv_chart}
\alias{crypto.onchain.tv_chart}
\title{Prints table showing token volume on different Decentralized Exchanges.}
\usage{
crypto.onchain.tv_chart(
  symbol = "WBTC",
  trade_amount_currency = "USD",
  limit = 10,
  sortby = "tradeAmount",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) ERC20 token symbol or address}

\item{trade_amount_currency}{(character length 1) Currency of displayed trade amount. Default: USD}

\item{limit}{(integer length 1) Number of records to display}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.tv_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.tv_chart(symbol='WBTC', trade_amount_currency='USD', limit=10, sortby='tradeAmount', ascend=TRUE)
}
