% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.tx}
\alias{crypto.onchain.tx}
\title{Get info about transaction. \link{Source: Ethplorer}}
\usage{
crypto.onchain.tx(tx_hash)
}
\arguments{
\item{tx_hash}{(character length 1) Transaction hash e.g. 0x9dc7b43ad4288c624fdd236b2ecb9f2b81c93e706b2ffd1d19b112c1df7849e6}
}
\description{
Wrapper for Python function crypto.onchain.tx from OpenBB Terminal SDK
}
