% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.tx_chart}
\alias{crypto.onchain.tx_chart}
\title{Display info about transaction. \link{Source: Ethplorer}}
\usage{
crypto.onchain.tx_chart(tx_hash, export = "", sheet_name)
}
\arguments{
\item{tx_hash}{(character length 1) Transaction hash e.g. 0x9dc7b43ad4288c624fdd236b2ecb9f2b81c93e706b2ffd1d19b112c1df7849e6}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.tx_chart from OpenBB Terminal SDK
}
