% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.ueat}
\alias{crypto.onchain.ueat}
\title{Get number of unique ethereum addresses which made a transaction in given time interval.}
\usage{
crypto.onchain.ueat(
  interval = "day",
  limit = 90,
  sortby = "tradeAmount",
  ascend = TRUE
)
}
\arguments{
\item{interval}{(character length 1) Time interval in which count unique ethereum addresses which made transaction. day,
month or week.}

\item{limit}{(integer length 1) Number of records for data query.}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.onchain.ueat from OpenBB Terminal SDK
}
\examples{
crypto.onchain.ueat(interval='day', limit=90, sortby='tradeAmount', ascend=TRUE)
}
