% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.ueat_chart}
\alias{crypto.onchain.ueat_chart}
\title{Prints table showing number of unique ethereum addresses which made a transaction in given time interval}
\usage{
crypto.onchain.ueat_chart(
  interval = "days",
  limit = 10,
  sortby = "date",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{interval}{(character length 1) Time interval in which ethereum address made transaction. month, week or day}

\item{limit}{(integer length 1) Number of records to display. It's calculated base on provided interval.
If interval is month then calculation is made in the way: limit * 30 = time period,
in case if interval is set to week, then time period is calculated as limit * 7.
For better user experience maximum time period in days is equal to 90.}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.ueat_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.ueat_chart(interval='days', limit=10, sortby='date', ascend=TRUE)
}
