% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.whales}
\alias{crypto.onchain.whales}
\title{Whale Alert's API allows you to retrieve live and historical transaction data from major blockchains.}
\usage{
crypto.onchain.whales(
  min_value = 8e+05,
  limit = 100,
  sortby = "date",
  ascend = FALSE
)
}
\arguments{
\item{min_value}{(integer length 1) Minimum value of trade to track.}

\item{limit}{(integer length 1) Limit of transactions. Max 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in ascending order.}
}
\description{
Wrapper for Python function crypto.onchain.whales from OpenBB Terminal SDK
}
\examples{
crypto.onchain.whales(min_value=800000, limit=100, sortby='date', ascend='False')
}
