% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.onchain.whales_chart}
\alias{crypto.onchain.whales_chart}
\title{Display huge value transactions from major blockchains. \link{Source: https://docs.whale-alert.io/}}
\usage{
crypto.onchain.whales_chart(
  min_value = 8e+05,
  limit = 100,
  sortby = "date",
  ascend = FALSE,
  show_address = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{min_value}{(integer length 1) Minimum value of trade to track.}

\item{limit}{(integer length 1) Limit of transactions. Maximum 100}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in ascending order.}

\item{show_address}{(logical length 1) Flag to show addresses of transactions.}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.onchain.whales_chart from OpenBB Terminal SDK
}
\examples{
crypto.onchain.whales_chart(min_value=800000, limit=100, sortby='date', ascend='False', show_address=FALSE)
}
