% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.altindex}
\alias{crypto.ov.altindex}
\title{Get altcoin index overtime}
\usage{
crypto.ov.altindex(period = 30, start_date = "2010-01-01", end_date)
}
\arguments{
\item{period}{(integer length 1) Number of days {30,90,365} to check performance of coins and calculate the altcoin index.
E.g., 365 checks yearly performance, 90 will check seasonal performance (90 days),
30 will check monthly performance (30 days).}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.ov.altindex from OpenBB Terminal SDK
}
