% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.altindex_chart}
\alias{crypto.ov.altindex_chart}
\title{Displays altcoin index overtime}
\usage{
crypto.ov.altindex_chart(
  period = 365,
  start_date = "2010-01-01",
  end_date,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{period}{(integer length 1) Number of days to check the performance of coins and calculate the altcoin index.
E.g., 365 will check yearly performance , 90 will check seasonal performance (90 days),
30 will check monthly performance (30 days).}

\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.altindex_chart from OpenBB Terminal SDK
}
