% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.btcrb}
\alias{crypto.ov.btcrb}
\title{Get bitcoin price data}
\usage{
crypto.ov.btcrb(start_date = "2010-01-01", end_date)
}
\arguments{
\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}
}
\description{
Wrapper for Python function crypto.ov.btcrb from OpenBB Terminal SDK
}
\examples{
crypto.ov.btcrb(start_date='2010-01-01')
}
