% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.btcrb_chart}
\alias{crypto.ov.btcrb_chart}
\title{Displays bitcoin rainbow chart}
\usage{
crypto.ov.btcrb_chart(
  start_date = "2010-01-01",
  end_date,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{start_date}{(character length 1) Initial date, format YYYY-MM-DD}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.btcrb_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.btcrb_chart(start_date='2010-01-01')
}
