% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.categories}
\alias{crypto.ov.categories}
\title{Returns top crypto categories \link{Source: CoinGecko}}
\usage{
crypto.ov.categories(sort_filter = "market_cap_desc")
}
\arguments{
\item{sort_filter}{(character length 1) Can be one of - "market_cap_desc", "market_cap_asc", "name_desc", "name_asc",
"market_cap_change_24h_desc", "market_cap_change_24h_asc"}
}
\description{
Wrapper for Python function crypto.ov.categories from OpenBB Terminal SDK
}
\examples{
crypto.ov.categories(sort_filter='market_cap_desc')
}
