% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.categories_chart}
\alias{crypto.ov.categories_chart}
\title{Shows top cryptocurrency categories by market capitalization}
\usage{
crypto.ov.categories_chart(
  sortby = "market_cap_desc",
  limit = 15,
  export = "",
  sheet_name,
  pie = FALSE
)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}

\item{pie}{(logical length 1) Whether to show the pie chart}
}
\description{
Wrapper for Python function crypto.ov.categories_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.categories_chart(sortby='market_cap_desc', limit=15, pie=FALSE)
}
