% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.cbpairs}
\alias{crypto.ov.cbpairs}
\title{Get a list of available currency pairs for trading. \link{Source: Coinbase}}
\usage{
crypto.ov.cbpairs(limit = 50, sortby = "quote_increment", ascend = TRUE)
}
\arguments{
\item{limit}{(integer length 1) Top n of pairs}

\item{sortby}{(character length 1) Key to sortby data}

\item{ascend}{(logical length 1) Sort descending flag}
}
\description{
Wrapper for Python function crypto.ov.cbpairs from OpenBB Terminal SDK
}
\examples{
crypto.ov.cbpairs(limit=50, sortby='quote_increment', ascend=TRUE)
}
