% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.cbpairs_chart}
\alias{crypto.ov.cbpairs_chart}
\title{Displays a list of available currency pairs for trading. \link{Source: Coinbase}}
\usage{
crypto.ov.cbpairs_chart(
  limit = 20,
  sortby = "quote_increment",
  ascend = TRUE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Top n of pairs}

\item{sortby}{(character length 1) Key to sortby data}

\item{ascend}{(logical length 1) Sort ascending flag}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.cbpairs_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.cbpairs_chart(limit=20, sortby='quote_increment', ascend=TRUE)
}
