% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.coin_list}
\alias{crypto.ov.coin_list}
\title{Get list of all available coins on CoinPaprika  \link{Source: CoinPaprika}}
\usage{
crypto.ov.coin_list()
}
\description{
Wrapper for Python function crypto.ov.coin_list from OpenBB Terminal SDK
}
