% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.contracts}
\alias{crypto.ov.contracts}
\title{Gets all contract addresses for given platform \link{Source: CoinPaprika}}
\usage{
crypto.ov.contracts(
  platform_id = "eth-ethereum",
  sortby = "active",
  ascend = TRUE
)
}
\arguments{
\item{platform_id}{(character length 1) Blockchain platform like eth-ethereum}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascend}
}
\description{
Wrapper for Python function crypto.ov.contracts from OpenBB Terminal SDK
}
\examples{
crypto.ov.contracts(platform_id='eth-ethereum', sortby='active', ascend=TRUE)
}
