% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.contracts_chart}
\alias{crypto.ov.contracts_chart}
\title{Gets all contract addresses for given platform. \link{Source: CoinPaprika}}
\usage{
crypto.ov.contracts_chart(
  symbol,
  sortby = "active",
  ascend = TRUE,
  limit = 15,
  export = "",
  sheet_name
)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.contracts_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.contracts_chart(limit=15, sortby='active', ascend=TRUE)
}
