% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.cr}
\alias{crypto.ov.cr}
\title{Returns crypto {borrow,supply} interest rates for cryptocurrencies across several platforms}
\usage{
crypto.ov.cr(rate_type = "borrow")
}
\arguments{
\item{rate_type}{(character length 1) Interest rate type: {borrow, supply}. Default: supply}
}
\description{
Wrapper for Python function crypto.ov.cr from OpenBB Terminal SDK
}
\examples{
crypto.ov.cr(rate_type='borrow')
}
