% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.cr_chart}
\alias{crypto.ov.cr_chart}
\title{Displays crypto {borrow,supply} interest rates for cryptocurrencies across several platforms}
\usage{
crypto.ov.cr_chart(
  symbols,
  platforms,
  rate_type = "borrow",
  limit = 10,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbols}{(character length 1) Crypto separated by commas. Default: BTC,ETH,USDT,USDC}

\item{platforms}{(character length 1) Platforms separated by commas. Default: BlockFi,Ledn,SwissBorg,Youhodler}

\item{rate_type}{(character length 1) Interest rate type: {borrow, supply}. Default: supply}

\item{limit}{(integer length 1) Number of records to show}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.cr_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.cr_chart(rate_type='borrow', limit=10)
}
