% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.crypto_hack}
\alias{crypto.ov.crypto_hack}
\title{Get crypto hack}
\usage{
crypto.ov.crypto_hack(slug)
}
\arguments{
\item{slug}{(character length 1) slug of crypto hack}
}
\description{
Wrapper for Python function crypto.ov.crypto_hack from OpenBB Terminal SDK
}
