% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.crypto_hack_slugs}
\alias{crypto.ov.crypto_hack_slugs}
\title{Get all crypto hack slugs}
\usage{
crypto.ov.crypto_hack_slugs()
}
\description{
Wrapper for Python function crypto.ov.crypto_hack_slugs from OpenBB Terminal SDK
}
