% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.crypto_hacks}
\alias{crypto.ov.crypto_hacks}
\title{Get major crypto-related hacks}
\usage{
crypto.ov.crypto_hacks(sortby = "Platform", ascend = FALSE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data {Platform,Date,Amount \link{$},Audit,Slug,URL}}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.ov.crypto_hacks from OpenBB Terminal SDK
}
\examples{
crypto.ov.crypto_hacks(sortby='Platform', ascend=FALSE)
}
