% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.crypto_hacks_chart}
\alias{crypto.ov.crypto_hacks_chart}
\title{Display list of major crypto-related hacks. If slug is passed}
\usage{
crypto.ov.crypto_hacks_chart(
  limit = 15,
  sortby = "Platform",
  ascend = FALSE,
  slug = "polyntwork-rekt",
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of hacks to search}

\item{sortby}{(character length 1) Key by which to sort data {Platform,Date,Amount \link{$},Audit,Slug,URL}}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{slug}{(character length 1) Crypto hack slug to check (e.g., polynetwork-rekt)}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.crypto_hacks_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.crypto_hacks_chart(slug='polyntwork-rekt', limit=15, sortby='Platform', ascend=FALSE)
}
