% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.defi}
\alias{crypto.ov.defi}
\title{Get global statistics about Decentralized Finances \link{Source: CoinGecko}}
\usage{
crypto.ov.defi()
}
\description{
Wrapper for Python function crypto.ov.defi from OpenBB Terminal SDK
}
