% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.defi_chart}
\alias{crypto.ov.defi_chart}
\title{Shows global statistics about Decentralized Finances. \link{Source: CoinGecko}}
\usage{
crypto.ov.defi_chart(export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.defi_chart from OpenBB Terminal SDK
}
