% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.derivatives}
\alias{crypto.ov.derivatives}
\title{Get list of crypto derivatives from CoinGecko API \link{Source: CoinGecko}}
\usage{
crypto.ov.derivatives(sortby = "Rank", ascend = FALSE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.ov.derivatives from OpenBB Terminal SDK
}
\examples{
crypto.ov.derivatives(sortby='Rank', ascend=FALSE)
}
