% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.ewf_chart}
\alias{crypto.ov.ewf_chart}
\title{Exchange withdrawal fees}
\usage{
crypto.ov.ewf_chart(export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.ewf_chart from OpenBB Terminal SDK
}
