% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.exchanges}
\alias{crypto.ov.exchanges}
\title{Show top crypto exchanges.}
\usage{
crypto.ov.exchanges(source = "CoinGecko")
}
\arguments{
\item{source}{(character length 1) Source to get exchanges, by default "CoinGecko"}
}
\description{
Wrapper for Python function crypto.ov.exchanges from OpenBB Terminal SDK
}
\examples{
crypto.ov.exchanges(source='CoinGecko')
}
