% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.exmarkets_chart}
\alias{crypto.ov.exmarkets_chart}
\title{Get all markets for given exchange \link{Source: CoinPaprika}}
\usage{
crypto.ov.exmarkets_chart(
  exchange = "binance",
  sortby = "pair",
  ascend = TRUE,
  limit = 15,
  links = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{exchange}{(character length 1) Exchange identifier e.g Binance}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{limit}{(integer length 1) Number of records to display}

\item{links}{(logical length 1) Flag to display urls}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.exmarkets_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.exmarkets_chart(exchange='binance', limit=15, sortby='pair', ascend=TRUE, links=FALSE)
}
