% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.exrates}
\alias{crypto.ov.exrates}
\title{Get list of crypto, fiats, commodity exchange rates from CoinGecko API \link{Source: CoinGecko}}
\usage{
crypto.ov.exrates(sortby = "Name", ascend = FALSE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.ov.exrates from OpenBB Terminal SDK
}
\examples{
crypto.ov.exrates(sortby='Name', ascend=FALSE)
}
