% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.hm_chart}
\alias{crypto.ov.hm_chart}
\title{Shows cryptocurrencies heatmap \link{Source: CoinGecko}}
\usage{
crypto.ov.hm_chart(
  category = "",
  limit = 15,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{limit}{(integer length 1) Number of top cryptocurrencies to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx}
}
\description{
Wrapper for Python function crypto.ov.hm_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.hm_chart(limit=15)
}
