% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.hold}
\alias{crypto.ov.hold}
\title{Returns public companies that holds ethereum or bitcoin \link{Source: CoinGecko}}
\usage{
crypto.ov.hold(endpoint = "bitcoin")
}
\arguments{
\item{endpoint}{(character length 1) "bitcoin" or "ethereum"}
}
\description{
Wrapper for Python function crypto.ov.hold from OpenBB Terminal SDK
}
\examples{
crypto.ov.hold(endpoint='bitcoin')
}
