% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.hold_chart}
\alias{crypto.ov.hold_chart}
\title{Shows overview of public companies that holds ethereum or bitcoin. \link{Source: CoinGecko}}
\usage{
crypto.ov.hold_chart(
  symbol,
  show_bar = FALSE,
  export = "",
  sheet_name,
  limit = 15
)
}
\arguments{
\item{symbol}{(character length 1) Cryptocurrency: ethereum or bitcoin}

\item{show_bar}{(logical length 1) Whether to show a bar graph for the data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx}

\item{limit}{(integer length 1) The number of rows to show}
}
\description{
Wrapper for Python function crypto.ov.hold_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.hold_chart(symbol='ETH', show_bar=FALSE, limit=15)
}
