% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.indexes}
\alias{crypto.ov.indexes}
\title{Get list of crypto indexes from CoinGecko API \link{Source: CoinGecko}}
\usage{
crypto.ov.indexes(sortby = "Name", ascend = TRUE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.ov.indexes from OpenBB Terminal SDK
}
\examples{
crypto.ov.indexes(sortby='Name', ascend=TRUE)
}
