% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.indexes_chart}
\alias{crypto.ov.indexes_chart}
\title{Shows list of crypto indexes. \link{Source: CoinGecko}}
\usage{
crypto.ov.indexes_chart(
  sortby = "Name",
  ascend = TRUE,
  limit = 15,
  export = "",
  sheet_name
)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.indexes_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.indexes_chart(limit=15, sortby='Name', ascend=TRUE)
}
