% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.info}
\alias{crypto.ov.info}
\title{Returns basic coin information for all coins from CoinPaprika API \link{Source: CoinPaprika}}
\usage{
crypto.ov.info(symbols = "USD", sortby = "rank", ascend = TRUE)
}
\arguments{
\item{symbols}{(character length 1) Comma separated quotes to return e.g quotes=USD,BTC}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data descending}
}
\description{
Wrapper for Python function crypto.ov.info from OpenBB Terminal SDK
}
\examples{
crypto.ov.info(symbols='USD', sortby='rank', ascend=TRUE)
}
