% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.markets_chart}
\alias{crypto.ov.markets_chart}
\title{Displays basic market information for all coins from CoinPaprika API. \link{Source: CoinPaprika}}
\usage{
crypto.ov.markets_chart(
  symbol,
  sortby = "rank",
  ascend = TRUE,
  limit = 15,
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) Quoted currency}

\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.markets_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.markets_chart(symbol='AMZN', limit=15, sortby='rank', ascend=TRUE)
}
