% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.news_chart}
\alias{crypto.ov.news_chart}
\title{Display recent posts from CryptoPanic news aggregator platform.}
\usage{
crypto.ov.news_chart(
  post_kind = "news",
  region = "en",
  filter_,
  limit = 25,
  sortby = "published_at",
  ascend = FALSE,
  links = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{post_kind}{(character length 1) Filter by category of news. Available values: news or media.}

\item{region}{(character length 1) Filter news by regions. Available regions are: en (English), de (Deutsch), nl (Dutch),
es (Español), fr (Français), it (Italiano), pt (Português), ru (Русский)}

\item{limit}{(integer length 1) number of news to display}

\item{sortby}{(character length 1) Key to sort by.}

\item{ascend}{(logical length 1) Sort in ascending order.}

\item{links}{(logical length 1) Show urls for news}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.news_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.news_chart(limit=25, post_kind='news', region='en', sortby='published_at', ascend=FALSE, links=FALSE)
}
