% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.platforms}
\alias{crypto.ov.platforms}
\title{List all smart contract platforms like ethereum, solana, cosmos, polkadot, kusama ... \link{Source: CoinPaprika}}
\usage{
crypto.ov.platforms()
}
\description{
Wrapper for Python function crypto.ov.platforms from OpenBB Terminal SDK
}
