% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.platforms_chart}
\alias{crypto.ov.platforms_chart}
\title{List all smart contract platforms like ethereum, solana, cosmos, polkadot, kusama.}
\usage{
crypto.ov.platforms_chart(export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.platforms_chart from OpenBB Terminal SDK
}
