% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.products}
\alias{crypto.ov.products}
\title{Get list of financial products from CoinGecko API}
\usage{
crypto.ov.products(sortby = "Name", ascend = TRUE)
}
\arguments{
\item{sortby}{(character length 1) Key by which to sort data}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.ov.products from OpenBB Terminal SDK
}
\examples{
crypto.ov.products(sortby='Name', ascend=TRUE)
}
