% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.stables}
\alias{crypto.ov.stables}
\title{Returns top stable coins \link{Source: CoinGecko}}
\usage{
crypto.ov.stables(limit = 15, sortby = "Market_Cap_[$]", ascend = FALSE)
}
\arguments{
\item{limit}{(integer length 1) How many rows to show}

\item{sortby}{(character length 1) Key by which to sort data, default is Market_Cap_\link{$}}

\item{ascend}{(logical length 1) Flag to sort data ascending}
}
\description{
Wrapper for Python function crypto.ov.stables from OpenBB Terminal SDK
}
\examples{
crypto.ov.stables(limit=15, sortby='Market_Cap_[$]', ascend=FALSE)
}
