% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.stables_chart}
\alias{crypto.ov.stables_chart}
\title{Shows stablecoins data \link{Source: CoinGecko}}
\usage{
crypto.ov.stables_chart(
  limit = 15,
  export = "",
  sheet_name,
  sortby = "Market_Cap_[$]",
  ascend = FALSE,
  pie = TRUE
)
}
\arguments{
\item{limit}{(integer length 1) Number of records to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}

\item{sortby}{(character length 1) Key by which to sort data, default is Market_Cap_\link{$}}

\item{ascend}{(logical length 1) Flag to sort data ascending}

\item{pie}{(logical length 1) Whether to show a pie chart}
}
\description{
Wrapper for Python function crypto.ov.stables_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.stables_chart(limit=15, sortby='Market_Cap_[$]', ascend=FALSE, pie=TRUE, pie=TRUE)
}
