% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.wf_chart}
\alias{crypto.ov.wf_chart}
\title{Top coins withdrawal fees}
\usage{
crypto.ov.wf_chart(limit = 15, export = "", sheet_name)
}
\arguments{
\item{limit}{(integer length 1) Number of coins to search}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.wf_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.wf_chart(limit=15)
}
