% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.wfpe}
\alias{crypto.ov.wfpe}
\title{Scrapes coin withdrawal fees per exchange}
\usage{
crypto.ov.wfpe(symbol)
}
\arguments{
\item{symbol}{(character length 1) Coin to check withdrawal fees. By default bitcoin}
}
\description{
Wrapper for Python function crypto.ov.wfpe from OpenBB Terminal SDK
}
\examples{
crypto.ov.wfpe(symbol='AMZN')
}
