% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.ov.wfpe_chart}
\alias{crypto.ov.wfpe_chart}
\title{Coin withdrawal fees per exchange}
\usage{
crypto.ov.wfpe_chart(symbol, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Coin to check withdrawal fees}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.ov.wfpe_chart from OpenBB Terminal SDK
}
\examples{
crypto.ov.wfpe_chart(symbol='AMZN')
}
