% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.price}
\alias{crypto.price}
\title{Returns price and confidence interval from pyth live feed. \link{Source: Pyth}}
\usage{
crypto.price(symbol)
}
\arguments{
\item{symbol}{(character length 1) Symbol of the asset to get price and confidence interval from}
}
\description{
Wrapper for Python function crypto.price from OpenBB Terminal SDK
}
\examples{
crypto.price(symbol='AMZN')
}
