% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.tools.apy}
\alias{crypto.tools.apy}
\title{Converts apr into apy}
\usage{
crypto.tools.apy(apr, compounding_times)
}
\arguments{
\item{apr}{(numeric length 1) value in percentage}

\item{compounding_times}{(integer length 1) number of compounded periods in a year}
}
\description{
Wrapper for Python function crypto.tools.apy from OpenBB Terminal SDK
}
