% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.tools.apy_chart}
\alias{crypto.tools.apy_chart}
\title{Displays APY value converted from APR}
\usage{
crypto.tools.apy_chart(
  apr,
  compounding_times,
  narrative = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{apr}{(numeric length 1) value in percentage}

\item{compounding_times}{(integer length 1) number of compounded periods in a year}

\item{narrative}{(logical length 1) display narrative version instead of dataframe}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function crypto.tools.apy_chart from OpenBB Terminal SDK
}
\examples{
crypto.tools.apy_chart(narrative='False')
}
