% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crypto.tools.il}
\alias{crypto.tools.il}
\title{Calculates Impermanent Loss in a custom liquidity pool}
\usage{
crypto.tools.il(price_changeA, price_changeB, proportion, initial_pool_value)
}
\arguments{
\item{price_changeA}{(numeric length 1) price change of crypto A in percentage}

\item{price_changeB}{(numeric length 1) price change of crypto B in percentage}

\item{proportion}{(numeric length 1) percentage of first token in pool}

\item{initial_pool_value}{(numeric length 1) initial value that pool contains}
}
\description{
Wrapper for Python function crypto.tools.il from OpenBB Terminal SDK
}
