% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.bgod}
\alias{econometrics.bgod}
\title{Calculate test statistics for autocorrelation}
\usage{
econometrics.bgod(model, lags = 3)
}
\arguments{
\item{model}{(data.frame) OLS model that has been fit.}

\item{lags}{(integer length 1) The amount of lags.}
}
\description{
Wrapper for Python function econometrics.bgod from OpenBB Terminal SDK
}
\examples{
econometrics.bgod(lags=3)
}
