% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.bgod_chart}
\alias{econometrics.bgod_chart}
\title{Show Breusch-Godfrey autocorrelation test}
\usage{
econometrics.bgod_chart(model, lags = 3, export = "", sheet_name)
}
\arguments{
\item{model}{(Python type: statsmodels.regression.linear_model.RegressionResultsWrapper) OLS model that has been fit.}

\item{lags}{(integer length 1) The amount of lags included.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.bgod_chart from OpenBB Terminal SDK
}
\examples{
econometrics.bgod_chart(lags=3)
}
