% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.bols}
\alias{econometrics.bols}
\title{The between estimator is an alternative, usually less efficient estimator, can can be used to}
\usage{
econometrics.bols(Y, X)
}
\description{
Wrapper for Python function econometrics.bols from OpenBB Terminal SDK
}
