% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.bpag}
\alias{econometrics.bpag}
\title{Calculate test statistics for heteroscedasticity}
\usage{
econometrics.bpag(model)
}
\arguments{
\item{model}{(Python type: statsmodels.regression.linear_model.RegressionResultsWrapper) Model containing residual values.}
}
\description{
Wrapper for Python function econometrics.bpag from OpenBB Terminal SDK
}
