% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.bpag_chart}
\alias{econometrics.bpag_chart}
\title{Show Breusch-Pagan heteroscedasticity test}
\usage{
econometrics.bpag_chart(model, export = "", sheet_name)
}
\arguments{
\item{model}{(Python type: statsmodels.regression.linear_model.RegressionResultsWrapper) OLS model that has been fit.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.bpag_chart from OpenBB Terminal SDK
}
