% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.clean}
\alias{econometrics.clean}
\title{Clean up NaNs from the dataset}
\usage{
econometrics.clean(dataset, fill = "", drop = "", limit)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to clean}

\item{fill}{(character length 1) The method of filling NaNs. Choose from:
rfill, cfill, rbfill, cbfill, rffill, cffill}

\item{drop}{(character length 1) The method of dropping NaNs. Choose from:
rdrop, cdrop}
}
\description{
Wrapper for Python function econometrics.clean from OpenBB Terminal SDK
}
