% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.coint}
\alias{econometrics.coint}
\title{Calculate cointegration tests between variable number of input series}
\usage{
econometrics.coint(datasets, return_z = FALSE)
}
\arguments{
\item{datasets}{(Python type: pandas.core.series.Series) Input series to test cointegration for}

\item{return_z}{(logical length 1) Flag to return the z data to plot}
}
\description{
Wrapper for Python function econometrics.coint from OpenBB Terminal SDK
}
\examples{
econometrics.coint(return_z=FALSE)
}
