% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.coint_chart}
\alias{econometrics.coint_chart}
\title{Estimates long-run and short-run cointegration relationship for series y and x and apply}
\usage{
econometrics.coint_chart(
  datasets,
  significant = FALSE,
  plot = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{datasets}{(Python type: pandas.core.series.Series) Variable number of series to test for cointegration}

\item{significant}{(logical length 1) Show only companies that have p-values lower than this percentage}

\item{plot}{(logical length 1) Whether you wish to plot the z-values of all pairs.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.coint_chart from OpenBB Terminal SDK
}
\examples{
econometrics.coint_chart(plot=TRUE)
}
