% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.comparison}
\alias{econometrics.comparison}
\title{Compare regression results between Panel Data regressions.}
\usage{
econometrics.comparison(regressions, export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.comparison from OpenBB Terminal SDK
}
