% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.dwat}
\alias{econometrics.dwat}
\title{Calculate test statistics for Durbin Watson autocorrelation}
\usage{
econometrics.dwat(model)
}
\arguments{
\item{model}{(Python type: statsmodels.regression.linear_model.RegressionResultsWrapper) Previously fit statsmodels OLS.}
}
\description{
Wrapper for Python function econometrics.dwat from OpenBB Terminal SDK
}
