% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.dwat_chart}
\alias{econometrics.dwat_chart}
\title{Show Durbin-Watson autocorrelation tests}
\usage{
econometrics.dwat_chart(
  model,
  dependent_variable,
  plot = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{model}{(Python type: statsmodels.regression.linear_model.RegressionResultsWrapper) A fit statsmodels OLS model.}

\item{dependent_variable}{(Python type: pandas.core.series.Series) The dependent variable for plotting}

\item{plot}{(logical length 1) Whether to plot the residuals}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.dwat_chart from OpenBB Terminal SDK
}
\examples{
econometrics.dwat_chart(plot=TRUE)
}
