% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.fdols}
\alias{econometrics.fdols}
\title{First differencing is an alternative to using fixed effects when there might be correlation.}
\usage{
econometrics.fdols(Y, X)
}
\description{
Wrapper for Python function econometrics.fdols from OpenBB Terminal SDK
}
