% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.fe}
\alias{econometrics.fe}
\title{When effects are correlated with the regressors the RE and BE estimators are not consistent.}
\usage{
econometrics.fe(Y, X, entity_effects = FALSE, time_effects = FALSE)
}
\arguments{
\item{entity_effects}{(logical length 1) Whether to include entity effects}

\item{time_effects}{(logical length 1) Whether to include time effects}
}
\description{
Wrapper for Python function econometrics.fe from OpenBB Terminal SDK
}
\examples{
econometrics.fe(entity_effects=FALSE, time_effects=FALSE)
}
