% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.get_regression_data}
\alias{econometrics.get_regression_data}
\title{This function creates a DataFrame with the required regression data as}
\usage{
econometrics.get_regression_data(
  regression_variables,
  data,
  regression_type = ""
)
}
\arguments{
\item{regression_type}{(character length 1) The type of regression that is executed.}
}
\description{
Wrapper for Python function econometrics.get_regression_data from OpenBB Terminal SDK
}
