% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.granger}
\alias{econometrics.granger}
\title{Calculate granger tests}
\usage{
econometrics.granger(dependent_series, independent_series, lags = 3)
}
\arguments{
\item{dependent_series}{(Python type: pandas.core.series.Series) The series you want to test Granger Causality for.}

\item{independent_series}{(Python type: pandas.core.series.Series) The series that you want to test whether it Granger-causes time_series_y}

\item{lags}{(integer length 1) The amount of lags for the Granger test. By default, this is set to 3.}
}
\description{
Wrapper for Python function econometrics.granger from OpenBB Terminal SDK
}
\examples{
econometrics.granger(lags=3)
}
