% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.granger_chart}
\alias{econometrics.granger_chart}
\title{Show granger tests}
\usage{
econometrics.granger_chart(
  dependent_series,
  independent_series,
  lags = 3,
  confidence_level = 0.05,
  export = "",
  sheet_name
)
}
\arguments{
\item{dependent_series}{(Python type: pandas.core.series.Series) The series you want to test Granger Causality for.}

\item{independent_series}{(Python type: pandas.core.series.Series) The series that you want to test whether it Granger-causes dependent_series}

\item{lags}{(integer length 1) The amount of lags for the Granger test. By default, this is set to 3.}

\item{confidence_level}{(numeric length 1) The confidence level you wish to use. By default, this is set to 0.05.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.granger_chart from OpenBB Terminal SDK
}
\examples{
econometrics.granger_chart(lags=3, confidence_level=0.05)
}
