% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.load}
\alias{econometrics.load}
\title{Load custom file into dataframe.}
\usage{
econometrics.load(file, data_files, data_examples)
}
\arguments{
\item{file}{(character length 1) Path to file}
}
\description{
Wrapper for Python function econometrics.load from OpenBB Terminal SDK
}
