% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.norm}
\alias{econometrics.norm}
\title{The distribution of returns and generate statistics on the relation to the normal curve.}
\usage{
econometrics.norm(data)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) A series or column of a DataFrame to test normality for}
}
\description{
Wrapper for Python function econometrics.norm from OpenBB Terminal SDK
}
