% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.norm_chart}
\alias{econometrics.norm_chart}
\title{Determine the normality of a timeseries.}
\usage{
econometrics.norm_chart(
  data,
  dataset = "",
  column = "",
  plot = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series of custom data}

\item{dataset}{(character length 1) Dataset name}

\item{column}{(character length 1) Column for y data}

\item{plot}{(logical length 1) Whether you wish to plot a histogram}

\item{export}{(character length 1) Format to export data.}
}
\description{
Wrapper for Python function econometrics.norm_chart from OpenBB Terminal SDK
}
\examples{
econometrics.norm_chart(plot=TRUE)
}
