% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.ols}
\alias{econometrics.ols}
\title{Performs an OLS regression on timeseries data. \link{Source: Statsmodels}}
\usage{
econometrics.ols(Y, X)
}
\arguments{
\item{Y}{(data.frame) Dependent variable series.}

\item{X}{(data.frame) Dataframe of independent variables series.}
}
\description{
Wrapper for Python function econometrics.ols from OpenBB Terminal SDK
}
