% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.options}
\alias{econometrics.options}
\title{Obtain columns-dataset combinations from loaded in datasets that can be used in other commands}
\usage{
econometrics.options(datasets, dataset_name = "")
}
\arguments{
\item{dataset_name}{(character length 1) The dataset you wish to show the options for.}
}
\description{
Wrapper for Python function econometrics.options from OpenBB Terminal SDK
}
