% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.options_chart}
\alias{econometrics.options_chart}
\title{Plot custom data}
\usage{
econometrics.options_chart(datasets, dataset_name, export = "", sheet_name)
}
\arguments{
\item{export}{(character length 1) Format to export image}
}
\description{
Wrapper for Python function econometrics.options_chart from OpenBB Terminal SDK
}
