% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.panel}
\alias{econometrics.panel}
\title{Based on the regression type, this function decides what regression to run.}
\usage{
econometrics.panel(
  Y,
  X,
  regression_type = "OLS",
  entity_effects = FALSE,
  time_effects = FALSE
)
}
\arguments{
\item{Y}{(data.frame) Dataframe containing the dependent variable.}

\item{X}{(data.frame) Dataframe containing the independent variables.}

\item{regression_type}{(character length 1) The type of regression you wish to execute.}

\item{entity_effects}{(logical length 1) Whether to apply Fixed Effects on entities.}

\item{time_effects}{(logical length 1) Whether to apply Fixed Effects on time.}
}
\description{
Wrapper for Python function econometrics.panel from OpenBB Terminal SDK
}
\examples{
econometrics.panel(regression_type='OLS', entity_effects=FALSE, time_effects=FALSE)
}
