% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.panel_chart}
\alias{econometrics.panel_chart}
\title{Based on the regression type, this function decides what regression to run.}
\usage{
econometrics.panel_chart(
  Y,
  X,
  regression_type = "OLS",
  entity_effects = FALSE,
  time_effects = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{regression_type}{(character length 1) The type of regression you wish to execute. Choose from:
OLS, POLS, RE, BOLS, FE}

\item{entity_effects}{(logical length 1) Whether to apply Fixed Effects on entities.}

\item{time_effects}{(logical length 1) Whether to apply Fixed Effects on time.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function econometrics.panel_chart from OpenBB Terminal SDK
}
\examples{
econometrics.panel_chart(regression_type='OLS', entity_effects=FALSE, time_effects=FALSE)
}
