% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.pols}
\alias{econometrics.pols}
\title{PooledOLS is just plain OLS that understands that various panel data structures.}
\usage{
econometrics.pols(Y, X)
}
\description{
Wrapper for Python function econometrics.pols from OpenBB Terminal SDK
}
