% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.re}
\alias{econometrics.re}
\title{The random effects model is virtually identical to the pooled OLS model except that is accounts for the}
\usage{
econometrics.re(Y, X)
}
\description{
Wrapper for Python function econometrics.re from OpenBB Terminal SDK
}
