% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.root}
\alias{econometrics.root}
\title{Calculate test statistics for unit roots}
\usage{
econometrics.root(data, fuller_reg = "c", kpss_reg = "c")
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series or column of DataFrame of target variable}

\item{fuller_reg}{(character length 1) Type of regression of ADF test}

\item{kpss_reg}{(character length 1) Type of regression for KPSS test}
}
\description{
Wrapper for Python function econometrics.root from OpenBB Terminal SDK
}
\examples{
econometrics.root(fuller_reg='c', kpss_reg='c')
}
