% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{econometrics.root_chart}
\alias{econometrics.root_chart}
\title{Determine the normality of a timeseries.}
\usage{
econometrics.root_chart(
  data,
  dataset = "",
  column = "",
  fuller_reg = "c",
  kpss_reg = "c",
  export = "",
  sheet_name
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series of target variable}

\item{dataset}{(character length 1) Name of the dataset}

\item{column}{(character length 1) Name of the column}

\item{fuller_reg}{(character length 1) Type of regression of ADF test. Choose c, ct, ctt, or nc}

\item{kpss_reg}{(character length 1) Type of regression for KPSS test. Choose c or ct}

\item{export}{(character length 1) Format to export data.}
}
\description{
Wrapper for Python function econometrics.root_chart from OpenBB Terminal SDK
}
\examples{
econometrics.root_chart(fuller_reg='c', kpss_reg='c')
}
