% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.available_indices}
\alias{economy.available_indices}
\title{Get available indices}
\usage{
economy.available_indices()
}
\description{
Wrapper for Python function economy.available_indices from OpenBB Terminal SDK
}
