% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.bigmac}
\alias{economy.bigmac}
\title{Display Big Mac Index for given countries}
\usage{
economy.bigmac(country_codes)
}
\description{
Wrapper for Python function economy.bigmac from OpenBB Terminal SDK
}
