% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.bigmac_chart}
\alias{economy.bigmac_chart}
\title{Display Big Mac Index for given countries}
\usage{
economy.bigmac_chart(
  country_codes,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{raw}{(logical length 1) Flag to display raw data, by default False}

\item{export}{(character length 1) Format data, by default ""}
}
\description{
Wrapper for Python function economy.bigmac_chart from OpenBB Terminal SDK
}
\examples{
economy.bigmac_chart(raw=FALSE)
}
