% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.country_codes}
\alias{economy.country_codes}
\title{Get available country codes for Bigmac index}
\usage{
economy.country_codes()
}
\description{
Wrapper for Python function economy.country_codes from OpenBB Terminal SDK
}
