% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.cpi}
\alias{economy.cpi}
\title{Obtain CPI data from FRED. \link{Source: FRED}}
\usage{
economy.cpi(
  countries,
  units = "",
  frequency = "",
  harmonized = FALSE,
  smart_select = TRUE,
  options = FALSE,
  start_date,
  end_date
)
}
\arguments{
\item{countries}{(Python type: list) The country or countries you want to see.}

\item{units}{(character length 1) The units you want to see, can be "growth_previous", "growth_same" or "index_2015".}

\item{frequency}{(character length 1) The frequency you want to see, either "annual", monthly" or "quarterly".}

\item{harmonized}{(logical length 1) Whether you wish to obtain harmonized data.}

\item{smart_select}{(logical length 1) Whether to assist with the selection.}

\item{options}{(logical length 1) Whether to return the options.}
}
\description{
Wrapper for Python function economy.cpi from OpenBB Terminal SDK
}
\examples{
economy.cpi(interval='m', start_year=2010)
}
