% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.cpi_chart}
\alias{economy.cpi_chart}
\title{Plot CPI data. \link{Source: FRED}}
\usage{
economy.cpi_chart(
  countries,
  units = "growth_same",
  frequency = "monthly",
  harmonized = FALSE,
  smart_select = TRUE,
  options = FALSE,
  start_date,
  end_date,
  raw = FALSE,
  export = "",
  sheet_name = "",
  external_axes
)
}
\arguments{
\item{countries}{(Python type: list) List of countries to plot}

\item{units}{(character length 1) Units of the data, either "growth_same", "growth_previous", "index_2015"}

\item{frequency}{(character length 1) Frequency of the data, either "monthly", "quarterly" or "annual"}

\item{harmonized}{(logical length 1) Whether to use harmonized data}

\item{smart_select}{(logical length 1) Whether to automatically select the best series}

\item{options}{(logical length 1) Whether to show options}

\item{raw}{(logical length 1) Show raw data}

\item{export}{(character length 1) Export data to csv or excel file}

\item{sheet_name}{(character length 1) Name of the sheet to export to}
}
\description{
Wrapper for Python function economy.cpi_chart from OpenBB Terminal SDK
}
\examples{
economy.cpi_chart(interval='m', start_year=2010, raw=FALSE)
}
