% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.currencies}
\alias{economy.currencies}
\title{Scrape data for global currencies}
\usage{
economy.currencies()
}
\description{
Wrapper for Python function economy.currencies from OpenBB Terminal SDK
}
