% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.events}
\alias{economy.events}
\title{Get economic calendar for countries between specified dates}
\usage{
economy.events(countries, start_date, end_date)
}
\description{
Wrapper for Python function economy.events from OpenBB Terminal SDK
}
