% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.fred}
\alias{economy.fred}
\title{Get Series data. \link{Source: FRED}}
\usage{
economy.fred(series_ids, start_date, end_date)
}
\description{
Wrapper for Python function economy.fred from OpenBB Terminal SDK
}
