% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.fred_chart}
\alias{economy.fred_chart}
\title{Display (multiple) series from https://fred.stlouisfed.org. \link{Source: FRED}}
\usage{
economy.fred_chart(
  series_ids,
  start_date,
  end_date,
  limit = 10,
  get_data = FALSE,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{limit}{(integer length 1) Number of data points to display.}

\item{raw}{(logical length 1) Output only raw data}

\item{export}{(character length 1) Export data to csv,json,xlsx or png,jpg,pdf,svg file}
}
\description{
Wrapper for Python function economy.fred_chart from OpenBB Terminal SDK
}
\examples{
economy.fred_chart(limit=10, raw=FALSE)
}
