% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.fred_ids}
\alias{economy.fred_ids}
\title{Get Series IDs. \link{Source: FRED}}
\usage{
economy.fred_ids(search_query, limit = -1)
}
\arguments{
\item{search_query}{(character length 1) Text query to search on fred series notes database}

\item{limit}{(integer length 1) Maximum number of series IDs to output}
}
\description{
Wrapper for Python function economy.fred_ids from OpenBB Terminal SDK
}
\examples{
economy.fred_ids(limit=-1)
}
