% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.fred_notes}
\alias{economy.fred_notes}
\title{Get series notes. \link{Source: FRED}}
\usage{
economy.fred_notes(search_query, limit = -1)
}
\arguments{
\item{search_query}{(character length 1) Text query to search on fred series notes database}

\item{limit}{(integer length 1) Maximum number of series notes to display}
}
\description{
Wrapper for Python function economy.fred_notes from OpenBB Terminal SDK
}
\examples{
economy.fred_notes(limit=-1)
}
