% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.future}
\alias{economy.future}
\title{Get futures data. \link{Source: Finviz}}
\usage{
economy.future(future_type = "Indices", sortby = "ticker", ascend = FALSE)
}
\arguments{
\item{future_type}{(character length 1) From the following: Indices, Energy, Metals, Meats, Grains, Softs, Bonds, Currencies}

\item{sortby}{(character length 1) Column to sort by}

\item{ascend}{(logical length 1) Flag to sort in ascending order}
}
\description{
Wrapper for Python function economy.future from OpenBB Terminal SDK
}
\examples{
economy.future(future_type='Indices', sortby='ticker', ascend=FALSE)
}
