% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.futures}
\alias{economy.futures}
\title{Get futures data.}
\usage{
economy.futures(source, future_type = "Indices")
}
\arguments{
\item{source}{(Python type: inspect._empty) Data source for futures data.  From the following: WSJ, Finviz}

\item{future_type}{(character length 1) (Finviz only) Future type to get.  Can be: Indices, Energy, Metals, Meats, Grains, Softs, Bonds, Currencies.}
}
\description{
Wrapper for Python function economy.futures from OpenBB Terminal SDK
}
\examples{
economy.futures(source='WSJ', future_type='Indices')
}
