% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.index_chart}
\alias{economy.index_chart}
\title{Load (and show) the selected indices over time \link{Source: Yahoo Finance}}
\usage{
economy.index_chart(
  indices,
  interval = "1d",
  start_date,
  end_date,
  column = "Adj Close",
  returns = FALSE,
  raw = FALSE,
  external_axes,
  export = "",
  sheet_name
)
}
\arguments{
\item{indices}{(Python type: list) A list of indices you wish to load (and plot).
Available indices can be accessed through economy.available_indices().}

\item{interval}{(character length 1) Valid intervals: 1m,2m,5m,15m,30m,60m,90m,1h,1d,5d,1wk,1mo,3mo
Intraday data cannot extend last 60 days}

\item{column}{(character length 1) Which column to load in, by default this is the Adjusted Close.}

\item{returns}{(logical length 1) Flag to show cumulative returns on index}

\item{raw}{(logical length 1) Whether to display the raw output.}

\item{export}{(character length 1) Export data to csv,json,xlsx or png,jpg,pdf,svg file}
}
\description{
Wrapper for Python function economy.index_chart from OpenBB Terminal SDK
}
\examples{
economy.index_chart(interval='1d', column='Adj Close', returns=FALSE, raw=FALSE)
}
