% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.macro}
\alias{economy.macro}
\title{This functions groups the data queried from the EconDB database \link{Source: EconDB}}
\usage{
economy.macro(
  parameters,
  countries,
  transform = "",
  start_date = "1900-01-01",
  end_date,
  symbol = ""
)
}
\arguments{
\item{transform}{(character length 1) The selected transform. Available transforms can be accessed through get_macro_transform().}

\item{start_date}{(character length 1) The starting date, format "YEAR-MONTH-DAY", i.e. 2010-12-31.}

\item{symbol}{(character length 1) In what currency you wish to convert all values.}
}
\description{
Wrapper for Python function economy.macro from OpenBB Terminal SDK
}
\examples{
economy.macro(start_date='1900-01-01')
}
