% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.macro_countries}
\alias{economy.macro_countries}
\title{This function returns the available countries and respective currencies.}
\usage{
economy.macro_countries()
}
\description{
Wrapper for Python function economy.macro_countries from OpenBB Terminal SDK
}
