% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.macro_parameters}
\alias{economy.macro_parameters}
\title{This function returns the available macro parameters with detail.}
\usage{
economy.macro_parameters()
}
\description{
Wrapper for Python function economy.macro_parameters from OpenBB Terminal SDK
}
