% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.performance}
\alias{economy.performance}
\title{Get group (sectors, industry or country) performance data. \link{Source: Finviz}}
\usage{
economy.performance(group = "sector", sortby = "Name", ascend = TRUE)
}
\arguments{
\item{group}{(character length 1) Group by category. Available groups can be accessed through get_groups().}

\item{sortby}{(character length 1) Column to sort by}

\item{ascend}{(logical length 1) Flag to sort in ascending order}
}
\description{
Wrapper for Python function economy.performance from OpenBB Terminal SDK
}
\examples{
economy.performance(group='sector', sortby='Name', ascend=TRUE)
}
