% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.rtps_chart}
\alias{economy.rtps_chart}
\title{Display Real-Time Performance sector. \link{Source: AlphaVantage}}
\usage{
economy.rtps_chart(raw = FALSE, export = "", sheet_name, external_axes)
}
\arguments{
\item{raw}{(logical length 1) Output only raw data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function economy.rtps_chart from OpenBB Terminal SDK
}
\examples{
economy.rtps_chart(raw=FALSE)
}
