% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.search_index}
\alias{economy.search_index}
\title{Search indices by keyword. \link{Source: FinanceDatabase}}
\usage{
economy.search_index(keyword, limit = 10)
}
\arguments{
\item{keyword}{(Python type: list) The keyword you wish to search for. This can include spaces.}

\item{limit}{(integer length 1) The amount of views you want to show, by default this is set to 10.}
}
\description{
Wrapper for Python function economy.search_index from OpenBB Terminal SDK
}
