% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.spectrum}
\alias{economy.spectrum}
\title{Display finviz spectrum in system viewer \link{Source: Finviz}}
\usage{
economy.spectrum(group = "sector", export = "", sheet_name)
}
\arguments{
\item{group}{(character length 1) Group by category. Available groups can be accessed through get_groups().}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function economy.spectrum from OpenBB Terminal SDK
}
\examples{
economy.spectrum(group='sector')
}
