% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{economy.treasury_maturities}
\alias{economy.treasury_maturities}
\title{Get treasury maturity options \link{Source: EconDB}}
\usage{
economy.treasury_maturities()
}
\description{
Wrapper for Python function economy.treasury_maturities from OpenBB Terminal SDK
}
