% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.disc.mover}
\alias{etf.disc.mover}
\title{Scrape data for top etf movers.}
\usage{
etf.disc.mover(sort_type = "gainers", export = FALSE)
}
\arguments{
\item{sort_type}{(character length 1) Data to get. Can be "gainers", "decliners" or "active"}
}
\description{
Wrapper for Python function etf.disc.mover from OpenBB Terminal SDK
}
\examples{
etf.disc.mover(sort_type='gainers')
}
