% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.etf_by_category}
\alias{etf.etf_by_category}
\title{Return a selection of ETFs based on category filtered by total assets.}
\usage{
etf.etf_by_category(category)
}
\arguments{
\item{category}{(character length 1) Search by category to find ETFs matching the criteria.}
}
\description{
Wrapper for Python function etf.etf_by_category from OpenBB Terminal SDK
}
