% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.etf_by_name}
\alias{etf.etf_by_name}
\title{Get an ETF symbol and name based on ETF string to search. \link{Source: StockAnalysis}}
\usage{
etf.etf_by_name(name_to_search)
}
\arguments{
\item{name_to_search}{(character length 1) ETF name to match}
}
\description{
Wrapper for Python function etf.etf_by_name from OpenBB Terminal SDK
}
