% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.holdings}
\alias{etf.holdings}
\title{Get ETF holdings}
\usage{
etf.holdings(symbol)
}
\arguments{
\item{symbol}{(character length 1) Symbol to get holdings for}
}
\description{
Wrapper for Python function etf.holdings from OpenBB Terminal SDK
}
\examples{
etf.holdings(symbol='AMZN')
}
