% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.ld}
\alias{etf.ld}
\title{Return a selection of ETFs based on description filtered by total assets.}
\usage{
etf.ld(description)
}
\arguments{
\item{description}{(character length 1) Search by description to find ETFs matching the criteria.}
}
\description{
Wrapper for Python function etf.ld from OpenBB Terminal SDK
}
