% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.ln}
\alias{etf.ln}
\title{Return a selection of ETFs based on name filtered by total assets. \link{Source: Finance Database}}
\usage{
etf.ln(name)
}
\arguments{
\item{name}{(character length 1) Search by name to find ETFs matching the criteria.}
}
\description{
Wrapper for Python function etf.ln from OpenBB Terminal SDK
}
