% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.news}
\alias{etf.news}
\title{Get news for a given term. \link{Source: NewsAPI}}
\usage{
etf.news(query, limit = 10, start_date, show_newest = TRUE, sources = "")
}
\arguments{
\item{query}{(character length 1) term to search on the news articles}

\item{show_newest}{(logical length 1) flag to show newest articles first}

\item{sources}{(character length 1) sources to exclusively show news from (comma separated)}
}
\description{
Wrapper for Python function etf.news from OpenBB Terminal SDK
}
\examples{
etf.news(show_newest=TRUE)
}
