% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.news_chart}
\alias{etf.news_chart}
\title{Prints table showing news for a given term. \link{Source: NewsAPI}}
\usage{
etf.news_chart(
  query,
  limit = 3,
  start_date,
  show_newest = TRUE,
  sources = "",
  export = "",
  sheet_name
)
}
\arguments{
\item{query}{(character length 1) term to search on the news articles}

\item{limit}{(integer length 1) number of articles to display}

\item{show_newest}{(logical length 1) flag to show newest articles first}

\item{sources}{(character length 1) sources to exclusively show news from}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function etf.news_chart from OpenBB Terminal SDK
}
\examples{
etf.news_chart(limit=3, show_newest=TRUE)
}
