% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{etf.weights}
\alias{etf.weights}
\title{Return sector weightings allocation of ETF. \link{Source: FinancialModelingPrep}}
\usage{
etf.weights(name)
}
\arguments{
\item{name}{(character length 1) ETF name}
}
\description{
Wrapper for Python function etf.weights from OpenBB Terminal SDK
}
