% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.ameribor}
\alias{fixedincome.ameribor}
\title{Obtain data for American Interbank Offered Rate (AMERIBOR)}
\usage{
fixedincome.ameribor(parameter = "overnight", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) The parameter to get data for. Choose from:
"overnight": "AMERIBOR",
"term_30": "AMBOR30T",
"term_90": "AMBOR90T",
"1_week_term_structure": "AMBOR1W",
"1_month_term_structure": "AMBOR1M",
"3_month_term_structure": "AMBOR3M",
"6_month_term_structure": "AMBOR6M",
"1_year_term_structure": "AMBOR1Y",
"2_year_term_structure": "AMBOR2Y",
"30_day_ma": "AMBOR30",
"90_day_ma": "AMBOR90",}
}
\description{
Wrapper for Python function fixedincome.ameribor from OpenBB Terminal SDK
}
