% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.cp}
\alias{fixedincome.cp}
\title{Obtain Commercial Paper data}
\usage{
fixedincome.cp(
  maturity = "30d",
  category = "financial",
  grade = "aa",
  start_date,
  end_date
)
}
\arguments{
\item{maturity}{(character length 1) The maturity you want to see, either "overnight", "7d", "15d", "30d", "60d" or "90d"}

\item{category}{(character length 1) The category you want to see, either "asset_backed", "financial" or "non_financial"}

\item{grade}{(character length 1) The type of grade you want to see, either "a2_p2" or "aa"}
}
\description{
Wrapper for Python function fixedincome.cp from OpenBB Terminal SDK
}
