% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.dwpcr}
\alias{fixedincome.dwpcr}
\title{Obtain data for the Discount Window Primary Credit Rate.}
\usage{
fixedincome.dwpcr(parameter = "daily_excl_weekend", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) The parameter to get data for. Choose from:
"daily_excl_weekend"
"monthly"
"weekly"
"daily"
"annual"}
}
\description{
Wrapper for Python function fixedincome.dwpcr from OpenBB Terminal SDK
}
