% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.ecb}
\alias{fixedincome.ecb}
\title{Obtain data for ECB interest rates.}
\usage{
fixedincome.ecb(interest_type, start_date, end_date)
}
\description{
Wrapper for Python function fixedincome.ecb from OpenBB Terminal SDK
}
