% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.ecbycrv}
\alias{fixedincome.ecbycrv}
\title{Gets euro area yield curve data from ECB.}
\usage{
fixedincome.ecbycrv(
  date = "",
  yield_type = "spot_rate",
  return_date = FALSE,
  detailed = FALSE,
  any_rating = TRUE
)
}
\arguments{
\item{date}{(character length 1) Date to get curve for. If empty, gets most recent date (format yyyy-mm-dd)}

\item{yield_type}{(character length 1) What type of yield curve to get, options: \link{'spot_rate', 'instantaneous_forward', 'par_yield'}}

\item{return_date}{(logical length 1) If True, returns date of yield curve}

\item{detailed}{(logical length 1) If True, returns detailed data. Note that this is very slow.}
}
\description{
Wrapper for Python function fixedincome.ecbycrv from OpenBB Terminal SDK
}
