% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.fed}
\alias{fixedincome.fed}
\title{Obtain data for Effective Federal Funds Rate.}
\usage{
fixedincome.fed(
  parameter = "monthly",
  start_date,
  end_date,
  overnight = FALSE,
  quantiles = FALSE,
  target = FALSE
)
}
\arguments{
\item{parameter}{(character length 1) The parameter to get data for. Choose from:
"monthly"
"daily"
"weekly"
"daily_excl_weekend"
"annual"
"biweekly"
"volume"}

\item{overnight}{(logical length 1) Whether you want to plot the Overnight Banking Federal Rate}

\item{quantiles}{(logical length 1) Whether you want to see the 1, 25, 75 and 99 percentiles}

\item{target}{(logical length 1) Whether you want to see the high and low target range}
}
\description{
Wrapper for Python function fixedincome.fed from OpenBB Terminal SDK
}
