% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.ffrmc}
\alias{fixedincome.ffrmc}
\title{Get data for Selected Treasury Constant Maturity Minus Federal Funds Rate}
\usage{
fixedincome.ffrmc(parameter = "10_year", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) FRED ID of FFRMC data to plot}
}
\description{
Wrapper for Python function fixedincome.ffrmc from OpenBB Terminal SDK
}
