% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.hqm}
\alias{fixedincome.hqm}
\title{The HQM yield curve represents the high quality corporate bond market, i.e.,}
\usage{
fixedincome.hqm(date, par = FALSE)
}
\arguments{
\item{par}{(logical length 1) Whether you wish to plot the par yield curve as well}
}
\description{
Wrapper for Python function fixedincome.hqm from OpenBB Terminal SDK
}
