% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.icebofa}
\alias{fixedincome.icebofa}
\title{Get data for ICE BofA US Corporate Bond Indices.}
\usage{
fixedincome.icebofa(
  data_type = "yield",
  category = "all",
  area = "us",
  grade = "non_sovereign",
  options = FALSE,
  start_date,
  end_date
)
}
\arguments{
\item{data_type}{(character length 1) The type of data you want to see, either "yield", "yield_to_worst", "total_return", or "spread"}

\item{category}{(character length 1) The type of category you want to see, either "all", "duration", "eur" or "usd".}

\item{area}{(character length 1) The type of area you want to see, either "asia", "emea", "eu", "ex_g10", "latin_america" or "us"}

\item{grade}{(character length 1) The type of grade you want to see, either "a", "aa", "aaa", "b", "bb", "bbb", "ccc", "crossover",
"high_grade", "high_yield", "non_financial", "non_sovereign", "private_sector", "public_sector"}

\item{options}{(logical length 1) Set to True to obtain the available options.}
}
\description{
Wrapper for Python function fixedincome.icebofa from OpenBB Terminal SDK
}
