% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.iorb}
\alias{fixedincome.iorb}
\title{Obtain data for Interest Rate on Reserve Balances.}
\usage{
fixedincome.iorb(start_date, end_date)
}
\description{
Wrapper for Python function fixedincome.iorb from OpenBB Terminal SDK
}
