% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.moody}
\alias{fixedincome.moody}
\title{Get data for Moody Corporate Bond Index}
\usage{
fixedincome.moody(data_type = "aaa", spread, start_date, end_date)
}
\arguments{
\item{data_type}{(character length 1) The type of data you want to see, either "aaa" or "baa"}
}
\description{
Wrapper for Python function fixedincome.moody from OpenBB Terminal SDK
}
