% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.projection}
\alias{fixedincome.projection}
\title{Obtain data for the Federal Reserve's projection of the federal funds rate.}
\usage{
fixedincome.projection(long_run = FALSE)
}
\arguments{
\item{long_run}{(logical length 1) Whether to obtain data for the long run projection.}
}
\description{
Wrapper for Python function fixedincome.projection from OpenBB Terminal SDK
}
