% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.sofr}
\alias{fixedincome.sofr}
\title{Obtain data for Secured Overnight Financing Rate (SOFR)}
\usage{
fixedincome.sofr(parameter = "overnight", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) The parameter to get data for. Choose from:
"overnight"
"30_day_average"
"90_day_average"
"180_day_average"
"index"}
}
\description{
Wrapper for Python function fixedincome.sofr from OpenBB Terminal SDK
}
