% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.sonia}
\alias{fixedincome.sonia}
\title{Obtain data for Sterling Overnight Index Average (SONIA)}
\usage{
fixedincome.sonia(parameter = "rate", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) The parameter to get data for.}
}
\description{
Wrapper for Python function fixedincome.sonia from OpenBB Terminal SDK
}
