% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.spot}
\alias{fixedincome.spot}
\title{The spot rate for any maturity is the yield on a bond that provides}
\usage{
fixedincome.spot(maturity, category, start_date, end_date)
}
\description{
Wrapper for Python function fixedincome.spot from OpenBB Terminal SDK
}
