% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.tbffr}
\alias{fixedincome.tbffr}
\title{Get data for Selected Treasury Bill Minus Federal Funds Rate.}
\usage{
fixedincome.tbffr(parameter = "3_month", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) FRED ID of TBFFR data to plot, options: \link{'3_month', '6_month'}}
}
\description{
Wrapper for Python function fixedincome.tbffr from OpenBB Terminal SDK
}
