% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.tmc}
\alias{fixedincome.tmc}
\title{Get data for 10-Year Treasury Constant Maturity Minus Selected Treasury Constant Maturity.}
\usage{
fixedincome.tmc(parameter = "3_month", start_date, end_date)
}
\arguments{
\item{parameter}{(character length 1) FRED ID of TMC data to plot, options: \link{'T10Y3M', 'T10Y3M'}}
}
\description{
Wrapper for Python function fixedincome.tmc from OpenBB Terminal SDK
}
