% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.treasury}
\alias{fixedincome.treasury}
\title{Gets interest rates data from selected countries (3 month and 10 year)}
\usage{
fixedincome.treasury(
  short_term,
  long_term,
  forecast = FALSE,
  start_date,
  end_date
)
}
\arguments{
\item{forecast}{(logical length 1) If True, plot forecasts for short term interest rates}
}
\description{
Wrapper for Python function fixedincome.treasury from OpenBB Terminal SDK
}
