% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixedincome.usrates}
\alias{fixedincome.usrates}
\title{Plot various treasury rates from the United States}
\usage{
fixedincome.usrates(
  parameter = "tbills",
  maturity = "3_months",
  start_date,
  end_date
)
}
\arguments{
\item{parameter}{(character length 1) Either "tbills", "cmn", or "tips".}

\item{maturity}{(character length 1) Depending on the chosen parameter, a set of maturities is available.
"4_week": {"tbill": "DTB4WK"},
"1_month": {"cmn": "DGS1MO"},
"3_month": {"tbill": "TB3MS", "cmn": "DGS3MO"},
"6_month": {"tbill": "DTB6", "cmn": "DGS6MO"},
"1_year": {"tbill": "DTB1YR", "cmn": "DGS1"},
"2_year": {"cmn": "DGS2"},
"3_year": {"cmn": "DGS3"},
"5_year": {"tips": "DFII5", "cmn": "DGS5"},
"7_year": {"tips": "DFII7", "cmn": "DGS7"},
"10_year": {"tips": "DFII10", "cmn": "DGS10"},
"20_year": {"tips": "DFII20", "cmn": "DGS20"},
"30_year": {"tips": "DFII30", "cmn": "DGS30"},}
}
\description{
Wrapper for Python function fixedincome.usrates from OpenBB Terminal SDK
}
