% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.anom}
\alias{forecast.anom}
\title{Get Quantile Anomaly Detection Data}
\usage{
forecast.anom(data, target_column = "close", train_split = 0.6)
}
\description{
Wrapper for Python function forecast.anom from OpenBB Terminal SDK
}
