% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.anom_chart}
\alias{forecast.anom_chart}
\title{Display Quantile Anomaly Detection}
\usage{
forecast.anom_chart(
  data,
  dataset_name,
  target_column = "close",
  train_split = 0.6,
  start_date,
  end_date,
  external_axes
)
}
\description{
Wrapper for Python function forecast.anom_chart from OpenBB Terminal SDK
}
