% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.atr}
\alias{forecast.atr}
\title{Calculate the Average True Range of a variable based on a a specific stock ticker.}
\usage{
forecast.atr(
  dataset,
  close_column = "close",
  high_column = "high",
  low_column = "low"
)
}
\description{
Wrapper for Python function forecast.atr from OpenBB Terminal SDK
}
