% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.autoselect}
\alias{forecast.autoselect}
\title{Performs Automatic Statistical forecasting}
\usage{
forecast.autoselect(
  data,
  target_column = "close",
  seasonal_periods = 7,
  n_predict = 5,
  start_window = 0.85,
  forecast_horizon = 5
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{seasonal_periods}{(integer length 1) Number of seasonal periods in a year (7 for daily data)
If not set, inferred from frequency of the series.}

\item{n_predict}{(integer length 1) Number of days to forecast}

\item{start_window}{(numeric length 1) Size of sliding window from start of timeseries and onwards}

\item{forecast_horizon}{(integer length 1) Number of days to forecast when backtesting and retraining historical}
}
\description{
Wrapper for Python function forecast.autoselect from OpenBB Terminal SDK
}
\examples{
forecast.autoselect(target_column='close', seasonal_periods=7, n_predict=5, start_window=0.85, forecast_horizon=5)
}
