% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.clean}
\alias{forecast.clean}
\title{Clean up NaNs from the dataset}
\usage{
forecast.clean(dataset, fill, drop, limit)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to clean}
}
\description{
Wrapper for Python function forecast.clean from OpenBB Terminal SDK
}
