% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.combine}
\alias{forecast.combine}
\title{Adds the given column of df2 to df1}
\usage{
forecast.combine(df1, df2, column, dataset = "")
}
\arguments{
\item{df1}{(data.frame) The dataframe to add a column to}

\item{df2}{(data.frame) The dataframe to lose a column}

\item{column}{(character length 1) The column to transfer}

\item{dataset}{(character length 1) A name for df2 (shows in name of new column)}
}
\description{
Wrapper for Python function forecast.combine from OpenBB Terminal SDK
}
