% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.corr}
\alias{forecast.corr}
\title{Returns correlation for a given df}
\usage{
forecast.corr(data)
}
\arguments{
\item{data}{(data.frame) The df to produce statistics for}
}
\description{
Wrapper for Python function forecast.corr from OpenBB Terminal SDK
}
