% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.corr_chart}
\alias{forecast.corr_chart}
\title{Plot correlation coefficients for dataset features}
\usage{
forecast.corr_chart(dataset, export = "", sheet_name, external_axes)
}
\arguments{
\item{dataset}{(data.frame) The dataset fore calculating correlation coefficients}

\item{export}{(character length 1) Format to export image}
}
\description{
Wrapper for Python function forecast.corr_chart from OpenBB Terminal SDK
}
