% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.delete}
\alias{forecast.delete}
\title{Delete a column from a dataframe}
\usage{
forecast.delete(data, column)
}
\arguments{
\item{data}{(data.frame) The dataframe to delete a column from}

\item{column}{(character length 1) The column to delete}
}
\description{
Wrapper for Python function forecast.delete from OpenBB Terminal SDK
}
