% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.delta}
\alias{forecast.delta}
\title{Calculate the \%change of a variable based on a specific column}
\usage{
forecast.delta(dataset, target_column = "close")
}
\description{
Wrapper for Python function forecast.delta from OpenBB Terminal SDK
}
