% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.desc}
\alias{forecast.desc}
\title{Returns statistics for a given df}
\usage{
forecast.desc(data)
}
\arguments{
\item{data}{(data.frame) The df to produce statistics for}
}
\description{
Wrapper for Python function forecast.desc from OpenBB Terminal SDK
}
