% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.desc_chart}
\alias{forecast.desc_chart}
\title{Show descriptive statistics for a dataframe}
\usage{
forecast.desc_chart(data, name = "", export = "", sheet_name)
}
\arguments{
\item{data}{(data.frame) The dataframe to show}

\item{name}{(character length 1) The name of the dataframe}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function forecast.desc_chart from OpenBB Terminal SDK
}
