% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.ema}
\alias{forecast.ema}
\title{A moving average provides an indication of the trend of the price movement}
\usage{
forecast.ema(dataset, target_column = "close", period = 10)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to clean}

\item{target_column}{(character length 1) The column you wish to add the EMA to}

\item{period}{(integer length 1) Time Span}
}
\description{
Wrapper for Python function forecast.ema from OpenBB Terminal SDK
}
\examples{
forecast.ema(target_column='close', period=10)
}
