% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.load}
\alias{forecast.load}
\title{Load custom file into dataframe.}
\usage{
forecast.load(file, data_files, data_examples)
}
\arguments{
\item{file}{(character length 1) Path to file}
}
\description{
Wrapper for Python function forecast.load from OpenBB Terminal SDK
}
