% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.mom}
\alias{forecast.mom}
\title{A momentum oscillator, which measures the percentage change between the current}
\usage{
forecast.mom(dataset, target_column = "close", period = 10)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to calculate with}

\item{target_column}{(character length 1) The column you wish to add the MOM to}

\item{period}{(integer length 1) Time Span}
}
\description{
Wrapper for Python function forecast.mom from OpenBB Terminal SDK
}
\examples{
forecast.mom(target_column='close', period=10)
}
