% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.plot}
\alias{forecast.plot}
\title{Plot data from a dataset}
\usage{
forecast.plot(data, columns, export = "", sheet_name, external_axes)
}
\arguments{
\item{data}{(data.frame) The dataframe to plot}

\item{export}{(character length 1) Format to export image}
}
\description{
Wrapper for Python function forecast.plot from OpenBB Terminal SDK
}
