% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.regr}
\alias{forecast.regr}
\title{Perform Regression Forecasting}
\usage{
forecast.regr(
  data,
  target_column = "close",
  n_predict = 5,
  past_covariates,
  train_split = 0.85,
  forecast_horizon = 5,
  output_chunk_length = 5,
  lags,
  metric = "mape"
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{n_predict}{(integer length 1) Days to predict. Defaults to 5.}

\item{train_split}{(numeric length 1) Train/val split. Defaults to 0.85.}

\item{forecast_horizon}{(integer length 1) Forecast horizon when performing historical forecasting. Defaults to 5.}

\item{output_chunk_length}{(integer length 1) The length of the forecast of the model. Defaults to 1.}

\item{metric}{(character length 1) Metric to use for evaluation. Defaults to "mape".}
}
\description{
Wrapper for Python function forecast.regr from OpenBB Terminal SDK
}
\examples{
forecast.regr(n_predict=5, target_column='close', train_split=0.85, forecast_horizon=5, output_chunk_length=1)
}
