% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.rename}
\alias{forecast.rename}
\title{Rename a column in a dataframe}
\usage{
forecast.rename(data, old_column, new_column)
}
\arguments{
\item{data}{(data.frame) The dataframe to have a column renamed}

\item{old_column}{(character length 1) The column that will have its name changed}

\item{new_column}{(character length 1) The name to update to}
}
\description{
Wrapper for Python function forecast.rename from OpenBB Terminal SDK
}
