% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.rnn_chart}
\alias{forecast.rnn_chart}
\title{Display RNN forecast}
\usage{
forecast.rnn_chart(
  data,
  target_column = "close",
  dataset_name = "",
  n_predict = 5,
  train_split = 0.85,
  forecast_horizon = 5,
  model_type = "LSTM",
  hidden_dim = 20,
  dropout = 0,
  batch_size = 32,
  n_epochs = 100,
  learning_rate = 0.001,
  model_save_name = "rnn_model",
  training_length = 20,
  input_chunk_size = 14,
  force_reset = TRUE,
  save_checkpoints = TRUE,
  export = "",
  sheet_name,
  residuals = FALSE,
  forecast_only = FALSE,
  start_date,
  end_date,
  naive = FALSE,
  export_pred_raw = FALSE,
  metric = "mape",
  external_axes
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{dataset_name}{(character length 1) The name of the ticker to be predicted}

\item{n_predict}{(integer length 1) Days to predict. Defaults to 5.}

\item{train_split}{(numeric length 1) Train/val split. Defaults to 0.85.}

\item{forecast_horizon}{(integer length 1) Forecast horizon when performing historical forecasting. Defaults to 5.}

\item{model_type}{(character length 1) Either a string specifying the RNN module type ("RNN", "LSTM" or "GRU"). Defaults to "LSTM".}

\item{hidden_dim}{(integer length 1) Size for feature maps for each hidden RNN layer.. Defaults to 20.}

\item{dropout}{(numeric length 1) Fraction of neurons affected by Dropout. Defaults to 0.0.}

\item{batch_size}{(integer length 1) Number of time series (input and output sequences) used in each training pass. Defaults to 32.}

\item{n_epochs}{(integer length 1) Number of epochs over which to train the model. Defaults to 100.}

\item{learning_rate}{(numeric length 1) Defaults to 1e-3.}

\item{model_save_name}{(character length 1) Name for model. Defaults to "brnn_model".}

\item{force_reset}{(logical length 1) If set to True, any previously-existing model with the same name will be reset
(all checkpoints will be discarded). Defaults to True.}

\item{save_checkpoints}{(logical length 1) Whether or not to automatically save the untrained model and checkpoints from training. Defaults to True.}

\item{export}{(character length 1) Format to export data}

\item{residuals}{(logical length 1) Whether to show residuals for the model. Defaults to False.}

\item{forecast_only}{(logical length 1) Whether to only show dates in the forecasting range. Defaults to False.}

\item{naive}{(logical length 1) Whether to show the naive baseline. This just assumes the closing price will be the same
as the previous day's closing price. Defaults to False.}

\item{export_pred_raw}{(logical length 1) Whether to export the raw predictions. Defaults to False.}

\item{metric}{(character length 1) The metric to use for the forecast. Defaults to "mape".}
}
\description{
Wrapper for Python function forecast.rnn_chart from OpenBB Terminal SDK
}
\examples{
forecast.rnn_chart(n_predict=5, target_column='close', train_split=0.85, forecast_horizon=5, model_type='LSTM', hidden_dim=20, dropout=0, batch_size=16, n_epochs=100, learning_rate=0.001, model_save_name='rnn_model', force_reset=TRUE, save_checkpoints=TRUE, forecast_only=FALSE, residuals=FALSE, naive=FALSE)
}
