% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.roc}
\alias{forecast.roc}
\title{A momentum oscillator, which measures the percentage change between the current}
\usage{
forecast.roc(dataset, target_column = "close", period = 10)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to calculate with}

\item{target_column}{(character length 1) The column you wish to add the ROC to}

\item{period}{(integer length 1) Time Span}
}
\description{
Wrapper for Python function forecast.roc from OpenBB Terminal SDK
}
\examples{
forecast.roc(target_column='close', period=10)
}
