% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.rsi}
\alias{forecast.rsi}
\title{A momentum indicator that measures the magnitude of recent price changes to evaluate}
\usage{
forecast.rsi(dataset, target_column = "close", period = 10)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to calculate for}

\item{target_column}{(character length 1) The column you wish to add the RSI to}

\item{period}{(integer length 1) Time Span}
}
\description{
Wrapper for Python function forecast.rsi from OpenBB Terminal SDK
}
\examples{
forecast.rsi(target_column='close', period=10)
}
