% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.rwd}
\alias{forecast.rwd}
\title{Performs Random Walk with Drift forecasting}
\usage{
forecast.rwd(
  data,
  target_column = "close",
  n_predict = 5,
  start_window = 0.85,
  forecast_horizon = 5
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{n_predict}{(integer length 1) Number of days to forecast}

\item{start_window}{(numeric length 1) Size of sliding window from start of timeseries and onwards}

\item{forecast_horizon}{(integer length 1) Number of days to forecast when backtesting and retraining historical}
}
\description{
Wrapper for Python function forecast.rwd from OpenBB Terminal SDK
}
\examples{
forecast.rwd(target_column='close', n_predict=5, start_window=0.85, forecast_horizon=5)
}
