% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.rwd_chart}
\alias{forecast.rwd_chart}
\title{Display Random Walk with Drift Model}
\usage{
forecast.rwd_chart(
  data,
  target_column = "close",
  dataset_name = "",
  n_predict = 5,
  start_window = 0.85,
  forecast_horizon = 5,
  export = "",
  sheet_name,
  residuals = FALSE,
  forecast_only = FALSE,
  start_date,
  end_date,
  naive = FALSE,
  export_pred_raw = FALSE,
  external_axes
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{dataset_name}{(character length 1) The name of the ticker to be predicted}

\item{n_predict}{(integer length 1) Number of days to forecast}

\item{start_window}{(numeric length 1) Size of sliding window from start of timeseries and onwards}

\item{forecast_horizon}{(integer length 1) Number of days to forecast when backtesting and retraining historical}

\item{export}{(character length 1) Format to export data}

\item{residuals}{(logical length 1) Whether to show residuals for the model. Defaults to False.}

\item{forecast_only}{(logical length 1) Whether to only show dates in the forecasting range. Defaults to False.}

\item{naive}{(logical length 1) Whether to show the naive baseline. This just assumes the closing price will be the same
as the previous day's closing price. Defaults to False.}
}
\description{
Wrapper for Python function forecast.rwd_chart from OpenBB Terminal SDK
}
\examples{
forecast.rwd_chart(target_column='close', n_predict=5, start_window=0.85, forecast_horizon=5, residuals=FALSE, forecast_only=FALSE, naive=FALSE)
}
