% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.season_chart}
\alias{forecast.season_chart}
\title{Plot seasonality from a dataset}
\usage{
forecast.season_chart(
  data,
  column = "close",
  export = "",
  sheet_name,
  m,
  max_lag = 24,
  alpha = 0.05,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) The dataframe to plot}

\item{column}{(character length 1) The column of the dataframe to analyze}

\item{export}{(character length 1) Format to export image}

\item{max_lag}{(integer length 1) The maximal lag order to consider. Default is 24.}

\item{alpha}{(numeric length 1) The confidence interval to display. Default is 0.05.}
}
\description{
Wrapper for Python function forecast.season_chart from OpenBB Terminal SDK
}
\examples{
forecast.season_chart(column='close', max_lag=24, alpha=0.05)
}
