% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.show}
\alias{forecast.show}
\title{Show a dataframe in a table}
\usage{
forecast.show(
  data,
  limit = 15,
  limit_col = 10,
  name = "",
  export = "",
  sheet_name
)
}
\arguments{
\item{data}{(data.frame) The dataframe to show}

\item{limit}{(integer length 1) The number of rows to show}

\item{limit_col}{(integer length 1) The number of columns to show}

\item{name}{(character length 1) The name of the dataframe}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function forecast.show from OpenBB Terminal SDK
}
