% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.signal}
\alias{forecast.signal}
\title{A price signal based on short/long term price.}
\usage{
forecast.signal(dataset, target_column = "close")
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to calculate with}
}
\description{
Wrapper for Python function forecast.signal from OpenBB Terminal SDK
}
