% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.sto}
\alias{forecast.sto}
\title{Stochastic Oscillator \%K and \%D : A stochastic oscillator is a momentum indicator comparing a particular closing}
\usage{
forecast.sto(
  dataset,
  close_column = "close",
  high_column = "high",
  low_column = "low",
  period = 10
)
}
\arguments{
\item{dataset}{(data.frame) The dataset you wish to calculate for}

\item{close_column}{(character length 1) The column name for the close price}

\item{high_column}{(character length 1) The column name for the high price}

\item{low_column}{(character length 1) The column name for the low price}

\item{period}{(integer length 1) Span of time to calculate over}
}
\description{
Wrapper for Python function forecast.sto from OpenBB Terminal SDK
}
\examples{
forecast.sto(period=10)
}
