% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.tft}
\alias{forecast.tft}
\title{Performs Temporal Fusion Transformer forecasting}
\usage{
forecast.tft(
  data,
  target_column = "close",
  n_predict = 5,
  past_covariates,
  train_split = 0.85,
  forecast_horizon = 5,
  input_chunk_length = 14,
  output_chunk_length = 5,
  hidden_size = 16,
  lstm_layers = 1,
  num_attention_heads = 4,
  full_attention = FALSE,
  dropout = 0.1,
  hidden_continuous_size = 8,
  n_epochs = 200,
  batch_size = 32,
  model_save_name = "tft_model",
  force_reset = TRUE,
  save_checkpoints = TRUE,
  metric = "mape"
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{n_predict}{(integer length 1) Days to predict. Defaults to 5.}

\item{train_split}{(numeric length 1) Train/val split. Defaults to 0.85.}

\item{forecast_horizon}{(integer length 1) Forecast horizon when performing historical forecasting. Defaults to 5.}

\item{input_chunk_length}{(integer length 1) Number of past time steps that are fed to the forecasting module at prediction time.
Defaults to 14.}

\item{output_chunk_length}{(integer length 1) The length of the forecast of the model. Defaults to 5.}

\item{hidden_size}{(integer length 1) Hidden state size of the TFT. Defaults to 16.}

\item{lstm_layers}{(integer length 1) Number of layers for the Long Short Term Memory Encoder and Decoder. Defaults to 16.}

\item{full_attention}{(logical length 1) Whether to apply a multi-head attention query. Defaults to False>}

\item{dropout}{(numeric length 1) Fraction of neurons affected by dropout. Defaults to 0.1.}

\item{hidden_continuous_size}{(integer length 1) Default hidden size for processing continuous variables. Defaults to 8.}

\item{n_epochs}{(integer length 1) Number of epochs to run during training. Defaults to 200.}

\item{batch_size}{(integer length 1) Number of samples to pass through network during a single epoch. Defaults to 32.}

\item{model_save_name}{(character length 1) The name for the model. Defaults to tft_model}

\item{force_reset}{(logical length 1) If set to True, any previously-existing model with the same name will be reset
(all checkpoints will be discarded). Defaults to True.}

\item{save_checkpoints}{(logical length 1) Whether or not to automatically save the untrained model and checkpoints from training.
Defaults to True.}

\item{metric}{(character length 1) Metric to use for model selection. Defaults to "mape".}
}
\description{
Wrapper for Python function forecast.tft from OpenBB Terminal SDK
}
\examples{
forecast.tft(target_column='close')
}
