% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast.theta_chart}
\alias{forecast.theta_chart}
\title{Display Theta forecast}
\usage{
forecast.theta_chart(
  data,
  target_column = "close",
  dataset_name = "",
  seasonal = "M",
  seasonal_periods = 7,
  n_predict = 5,
  start_window = 0.85,
  forecast_horizon = 5,
  export = "",
  sheet_name,
  residuals = FALSE,
  forecast_only = FALSE,
  start_date,
  end_date,
  naive = FALSE,
  export_pred_raw = FALSE,
  metric = "mape",
  external_axes
)
}
\arguments{
\item{target_column}{(character length 1) Target column to forecast. Defaults to "close".}

\item{dataset_name}{(character length 1) The name of the ticker to be predicted}

\item{seasonal}{(character length 1) Seasonal component.  One of \link{N, A, M}
Defaults to MULTIPLICATIVE.}

\item{seasonal_periods}{(integer length 1) Number of seasonal periods in a year
If not set, inferred from frequency of the series.}

\item{n_predict}{(integer length 1) Number of days to forecast}

\item{start_window}{(numeric length 1) Size of sliding window from start of timeseries and onwards}

\item{forecast_horizon}{(integer length 1) Number of days to forecast when backtesting and retraining historical}

\item{export}{(character length 1) Format to export data}

\item{residuals}{(logical length 1) Whether to show residuals for the model. Defaults to False.}

\item{forecast_only}{(logical length 1) Whether to only show dates in the forecasting range. Defaults to False.}

\item{naive}{(logical length 1) Whether to show the naive baseline. This just assumes the closing price will be the same
as the previous day's closing price. Defaults to False.}

\item{export_pred_raw}{(logical length 1) Whether to export the raw predicted values. Defaults to False.}

\item{metric}{(character length 1) The metric to use for backtesting. Defaults to "mape".}
}
\description{
Wrapper for Python function forecast.theta_chart from OpenBB Terminal SDK
}
\examples{
forecast.theta_chart(target_column='close', seasonal='M', seasonal_periods=7, n_predict=5, start_window=0.85, forecast_horizon=5, residuals=FALSE, forecast_only=FALSE, naive=FALSE)
}
