% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.candle}
\alias{forex.candle}
\title{Show candle plot for fx data.}
\usage{
forex.candle(
  data,
  to_symbol = "",
  from_symbol = "",
  ma,
  external_axes,
  use_matplotlib = TRUE,
  add_trend = FALSE,
  yscale = "linear"
)
}
\arguments{
\item{data}{(data.frame) Loaded fx historical data}

\item{to_symbol}{(character length 1) To forex symbol}

\item{from_symbol}{(character length 1) From forex symbol}
}
\description{
Wrapper for Python function forex.candle from OpenBB Terminal SDK
}
