% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.fwd}
\alias{forex.fwd}
\title{Gets forward rates from fxempire}
\usage{
forex.fwd(to_symbol = "USD", from_symbol = "EUR")
}
\arguments{
\item{to_symbol}{(character length 1) To currency}

\item{from_symbol}{(character length 1) From currency}
}
\description{
Wrapper for Python function forex.fwd from OpenBB Terminal SDK
}
\examples{
forex.fwd(to_symbol='USD', from_symbol='EUR')
}
