% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.get_currency_list}
\alias{forex.get_currency_list}
\title{Load AV currency codes from a local file.}
\usage{
forex.get_currency_list()
}
\description{
Wrapper for Python function forex.get_currency_list from OpenBB Terminal SDK
}
