% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.load}
\alias{forex.load}
\title{Load forex for two given symbols.}
\usage{
forex.load(
  to_symbol,
  from_symbol,
  resolution = "d",
  interval = "1day",
  start_date,
  end_date,
  source = "YahooFinance",
  verbose = FALSE
)
}
\arguments{
\item{to_symbol}{(character length 1) The from currency symbol. Ex: USD, EUR, GBP, YEN}

\item{from_symbol}{(character length 1) The from currency symbol. Ex: USD, EUR, GBP, YEN}

\item{resolution}{(character length 1) The resolution for the data, by default "d"}

\item{interval}{(character length 1) What interval to get data for, by default "1day"}

\item{source}{(character length 1) Where to get data from, by default "YahooFinance"}

\item{verbose}{(logical length 1) Display verbose information on what was the pair that was loaded, by default True}
}
\description{
Wrapper for Python function forex.load from OpenBB Terminal SDK
}
\examples{
forex.load(resolution='d', interval='1day', source='YahooFinance', verbose=FALSE)
}
