% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.calendar}
\alias{forex.oanda.calendar}
\title{Request data of significant events calendar.}
\usage{
forex.oanda.calendar(days = 14, instrument)
}
\arguments{
\item{days}{(integer length 1) Number of days in advance}
}
\description{
Wrapper for Python function forex.oanda.calendar from OpenBB Terminal SDK
}
\examples{
forex.oanda.calendar(days=14)
}
