% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.calendar_chart}
\alias{forex.oanda.calendar_chart}
\title{View calendar of significant events.}
\usage{
forex.oanda.calendar_chart(instrument, days = 7)
}
\arguments{
\item{instrument}{(character length 1) The loaded currency pair}

\item{days}{(integer length 1) Number of days in advance}
}
\description{
Wrapper for Python function forex.oanda.calendar_chart from OpenBB Terminal SDK
}
\examples{
forex.oanda.calendar_chart(days=7)
}
