% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.cancel}
\alias{forex.oanda.cancel}
\title{Request cancellation of a pending order.}
\usage{
forex.oanda.cancel(orderID, accountID = "REPLACE_ME")
}
\arguments{
\item{orderID}{(character length 1) The pending order ID to cancel.}

\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.cancel from OpenBB Terminal SDK
}
\examples{
forex.oanda.cancel(accountID='REPLACE_ME')
}
