% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.cancel_chart}
\alias{forex.oanda.cancel_chart}
\title{Cancel a Pending Order.}
\usage{
forex.oanda.cancel_chart(accountID, orderID = "")
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}

\item{orderID}{(character length 1) The pending order ID to cancel.}
}
\description{
Wrapper for Python function forex.oanda.cancel_chart from OpenBB Terminal SDK
}
