% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.candles}
\alias{forex.oanda.candles}
\title{Request data for candle chart.}
\usage{
forex.oanda.candles(instrument, granularity = "D", candlecount = 180)
}
\arguments{
\item{granularity}{(character length 1) Data granularity, by default "D"}

\item{candlecount}{(integer length 1) Limit for the number of data points, by default 180}
}
\description{
Wrapper for Python function forex.oanda.candles from OpenBB Terminal SDK
}
\examples{
forex.oanda.candles(granularity='D', candlecount=180)
}
