% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.candles_chart}
\alias{forex.oanda.candles_chart}
\title{Show candle chart.}
\usage{
forex.oanda.candles_chart(
  instrument = "",
  granularity = "D",
  candlecount = 180,
  additional_charts,
  external_axes
)
}
\arguments{
\item{instrument}{(character length 1) The loaded currency pair}

\item{granularity}{(character length 1) The timeframe to get for the candle chart. Seconds: S5, S10, S15, S30
Minutes: M1, M2, M4, M5, M10, M15, M30 Hours: H1, H2, H3, H4, H6, H8, H12
Day (default): D, Week: W Month: M,}

\item{candlecount}{(integer length 1) Limit for the number of data points}
}
\description{
Wrapper for Python function forex.oanda.candles_chart from OpenBB Terminal SDK
}
\examples{
forex.oanda.candles_chart(granularity='D', candlecount=180)
}
