% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.close}
\alias{forex.oanda.close}
\title{Close a trade.}
\usage{
forex.oanda.close(orderID, units, accountID = "REPLACE_ME")
}
\arguments{
\item{orderID}{(character length 1) ID of the order to close}

\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.close from OpenBB Terminal SDK
}
\examples{
forex.oanda.close(units=0, accountID='REPLACE_ME')
}
