% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.close_chart}
\alias{forex.oanda.close_chart}
\title{Close a trade.}
\usage{
forex.oanda.close_chart(accountID, orderID = "", units)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}

\item{orderID}{(character length 1) ID of the order to close}
}
\description{
Wrapper for Python function forex.oanda.close_chart from OpenBB Terminal SDK
}
