% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.listorders}
\alias{forex.oanda.listorders}
\title{Request the orders list from Oanda.}
\usage{
forex.oanda.listorders(
  order_state = "PENDING",
  order_count = 0,
  accountID = "REPLACE_ME"
)
}
\arguments{
\item{order_state}{(character length 1) Filter orders by a specific state ("PENDING", "CANCELLED", etc.)}

\item{order_count}{(integer length 1) Limit the number of orders to retrieve}

\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.listorders from OpenBB Terminal SDK
}
\examples{
forex.oanda.listorders(order_state='PENDING', order_count=0, accountID='REPLACE_ME')
}
