% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.listorders_chart}
\alias{forex.oanda.listorders_chart}
\title{List order history.}
\usage{
forex.oanda.listorders_chart(
  accountID,
  order_state = "PENDING",
  order_count = 0
)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}

\item{order_state}{(character length 1) Filter orders by a specific state ("PENDING", "CANCELLED", etc.)}

\item{order_count}{(integer length 1) Limit the number of orders to retrieve}
}
\description{
Wrapper for Python function forex.oanda.listorders_chart from OpenBB Terminal SDK
}
\examples{
forex.oanda.listorders_chart(order_state='PENDING', order_count=0)
}
