% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.openpositions}
\alias{forex.oanda.openpositions}
\title{Request information on open positions.}
\usage{
forex.oanda.openpositions(accountID = "REPLACE_ME")
}
\arguments{
\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.openpositions from OpenBB Terminal SDK
}
\examples{
forex.oanda.openpositions(accountID='REPLACE_ME')
}
