% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.order}
\alias{forex.oanda.order}
\title{Request creation of buy/sell trade order.}
\usage{
forex.oanda.order(price = 0, units = 0, instrument, accountID = "REPLACE_ME")
}
\arguments{
\item{price}{(integer length 1) The price to set for the limit order.}

\item{units}{(integer length 1) The number of units to place in the order request.}

\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.order from OpenBB Terminal SDK
}
\examples{
forex.oanda.order(price=0, units=0, accountID='REPLACE_ME')
}
