% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.order_chart}
\alias{forex.oanda.order_chart}
\title{Create a buy/sell order.}
\usage{
forex.oanda.order_chart(accountID, instrument = "", price = 0, units = 0)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}

\item{instrument}{(character length 1) The loaded currency pair}

\item{price}{(integer length 1) The price to set for the limit order.}

\item{units}{(integer length 1) The number of units to place in the order request.}
}
\description{
Wrapper for Python function forex.oanda.order_chart from OpenBB Terminal SDK
}
\examples{
forex.oanda.order_chart(price=0, units=0)
}
