% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.orderbook_chart}
\alias{forex.oanda.orderbook_chart}
\title{Plot the orderbook for the instrument if Oanda provides one.}
\usage{
forex.oanda.orderbook_chart(accountID, instrument = "", external_axes)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}

\item{instrument}{(character length 1) The loaded currency pair}
}
\description{
Wrapper for Python function forex.oanda.orderbook_chart from OpenBB Terminal SDK
}
