% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.pending}
\alias{forex.oanda.pending}
\title{Request information on pending orders.}
\usage{
forex.oanda.pending(accountID = "REPLACE_ME")
}
\arguments{
\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.pending from OpenBB Terminal SDK
}
\examples{
forex.oanda.pending(accountID='REPLACE_ME')
}
