% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.pending_chart}
\alias{forex.oanda.pending_chart}
\title{Get information about pending orders.}
\usage{
forex.oanda.pending_chart(accountID)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}
}
\description{
Wrapper for Python function forex.oanda.pending_chart from OpenBB Terminal SDK
}
