% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.positionbook}
\alias{forex.oanda.positionbook}
\title{Request position book data for plotting.}
\usage{
forex.oanda.positionbook(instrument, accountID = "REPLACE_ME")
}
\arguments{
\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.positionbook from OpenBB Terminal SDK
}
\examples{
forex.oanda.positionbook(accountID='REPLACE_ME')
}
