% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.price_chart}
\alias{forex.oanda.price_chart}
\title{View price for loaded currency pair.}
\usage{
forex.oanda.price_chart(account, instrument)
}
\description{
Wrapper for Python function forex.oanda.price_chart from OpenBB Terminal SDK
}
