% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.summary}
\alias{forex.oanda.summary}
\title{Request Oanda account summary.}
\usage{
forex.oanda.summary(accountID = "REPLACE_ME")
}
\arguments{
\item{accountID}{(character length 1) Oanda account ID, by default cfg.OANDA_ACCOUNT}
}
\description{
Wrapper for Python function forex.oanda.summary from OpenBB Terminal SDK
}
\examples{
forex.oanda.summary(accountID='REPLACE_ME')
}
