% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.oanda.summary_chart}
\alias{forex.oanda.summary_chart}
\title{Print Oanda account summary.}
\usage{
forex.oanda.summary_chart(accountID)
}
\arguments{
\item{accountID}{(character length 1) Oanda user account ID}
}
\description{
Wrapper for Python function forex.oanda.summary_chart from OpenBB Terminal SDK
}
