% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forex.quote}
\alias{forex.quote}
\title{Get forex quote.}
\usage{
forex.quote(symbol, source = "YahooFinance")
}
\arguments{
\item{symbol}{(character length 1) Forex symbol to get quote for.}

\item{source}{(character length 1) Source to get quote from, by default "YahooFinance"}
}
\description{
Wrapper for Python function forex.quote from OpenBB Terminal SDK
}
\examples{
forex.quote(symbol='EURUSD', source='YahooFinance')
}
