% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{futures.curve}
\alias{futures.curve}
\title{Get curve futures \link{Source: Yahoo Finance}}
\usage{
futures.curve(symbol = "")
}
\arguments{
\item{symbol}{(character length 1) symbol to get forward curve}
}
\description{
Wrapper for Python function futures.curve from OpenBB Terminal SDK
}
