% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{futures.curve_chart}
\alias{futures.curve_chart}
\title{Display curve futures \link{Source: Yahoo Finance}}
\usage{
futures.curve_chart(
  symbol,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Curve future symbol to display}

\item{raw}{(logical length 1) Display futures timeseries in raw format}

\item{export}{(character length 1) Type of format to export data}
}
\description{
Wrapper for Python function futures.curve_chart from OpenBB Terminal SDK
}
\examples{
futures.curve_chart(symbol='AMZN', raw=FALSE)
}
