% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{futures.historical}
\alias{futures.historical}
\title{Get historical futures data}
\usage{
futures.historical(symbols, start_date, end_date, source, expiry)
}
\description{
Wrapper for Python function futures.historical from OpenBB Terminal SDK
}
