% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{futures.historical_chart}
\alias{futures.historical_chart}
\title{Display historical futures \link{Source: Yahoo Finance}}
\usage{
futures.historical_chart(
  symbols,
  expiry = "",
  start_date,
  end_date,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{expiry}{(character length 1) Future expiry date with format YYYY-MM}

\item{raw}{(logical length 1) Display futures timeseries in raw format}

\item{export}{(character length 1) Type of format to export data}
}
\description{
Wrapper for Python function futures.historical_chart from OpenBB Terminal SDK
}
\examples{
futures.historical_chart(raw=FALSE)
}
