% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{futures.search}
\alias{futures.search}
\title{Get search futures \link{Source: Yahoo Finance}}
\usage{
futures.search(category = "", exchange = "", description = "")
}
\arguments{
\item{category}{(character length 1) Select the category where the future exists}

\item{exchange}{(character length 1) Select the exchange where the future exists}

\item{description}{(character length 1) Select the description where the future exists}
}
\description{
Wrapper for Python function futures.search from OpenBB Terminal SDK
}
