% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the OpenBB environment}
\usage{
init()
}
\value{
A logical value indicating whether the Python environment is available.
}
\description{
This function sets up the Python virtual environment, imports necessary Python
libraries, and initializes the OpenBB environment.
}
