% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.binance}
\alias{keys.binance}
\title{Set Binance key}
\usage{
keys.binance(key, secret, persist = FALSE, show_output = FALSE)
}
\arguments{
\item{key}{(character length 1) API key}

\item{secret}{(character length 1) API secret}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.binance from OpenBB Terminal SDK
}
\examples{
keys.binance(persist=FALSE, show_output=FALSE)
}
