% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.degiro}
\alias{keys.degiro}
\title{Set Degiro key}
\usage{
keys.degiro(
  username,
  password,
  secret = "",
  persist = FALSE,
  show_output = FALSE
)
}
\arguments{
\item{username}{(character length 1) User username}

\item{password}{(character length 1) User password}

\item{secret}{(character length 1) User secret}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.degiro from OpenBB Terminal SDK
}
\examples{
keys.degiro(persist=FALSE, show_output=FALSE)
}
