% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.messari}
\alias{keys.messari}
\title{Set Messari key.}
\usage{
keys.messari(key, persist = FALSE, show_output = FALSE)
}
\arguments{
\item{key}{(character length 1) API key}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.messari from OpenBB Terminal SDK
}
\examples{
keys.messari(persist=FALSE, show_output=FALSE)
}
