% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.mykeys}
\alias{keys.mykeys}
\title{Get currently set API keys.}
\usage{
keys.mykeys(show = FALSE)
}
\arguments{
\item{show}{(logical length 1) Flag to choose whether to show actual keys or not.
By default, False.}
}
\description{
Wrapper for Python function keys.mykeys from OpenBB Terminal SDK
}
\examples{
keys.mykeys(show=FALSE)
}
