% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.oanda}
\alias{keys.oanda}
\title{Set Oanda key}
\usage{
keys.oanda(
  account,
  access_token,
  account_type = "",
  persist = FALSE,
  show_output = FALSE
)
}
\arguments{
\item{account}{(character length 1) User account}

\item{access_token}{(character length 1) User token}

\item{account_type}{(character length 1) User account type}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.oanda from OpenBB Terminal SDK
}
\examples{
keys.oanda(persist=FALSE, show_output=FALSE)
}
