% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.openbb}
\alias{keys.openbb}
\title{Set OpenBB Personal Access Token.}
\usage{
keys.openbb(key, persist = FALSE, show_output = FALSE)
}
\arguments{
\item{key}{(character length 1) Personal Access Token}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.openbb from OpenBB Terminal SDK
}
