% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.reddit}
\alias{keys.reddit}
\title{Set Reddit key}
\usage{
keys.reddit(
  client_id,
  client_secret,
  password,
  username,
  useragent,
  persist = FALSE,
  show_output = FALSE
)
}
\arguments{
\item{client_id}{(character length 1) Client ID}

\item{client_secret}{(character length 1) Client secret}

\item{password}{(character length 1) User password}

\item{username}{(character length 1) User username}

\item{useragent}{(character length 1) User useragent}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.reddit from OpenBB Terminal SDK
}
\examples{
keys.reddit(persist=FALSE, show_output=FALSE)
}
