% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.set_keys}
\alias{keys.set_keys}
\title{Set API keys in bundle.}
\usage{
keys.set_keys(keys_dict, persist = FALSE, show_output = FALSE)
}
\arguments{
\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.set_keys from OpenBB Terminal SDK
}
\examples{
keys.set_keys(persist=FALSE, show_output=FALSE)
}
