% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keys.twitter}
\alias{keys.twitter}
\title{Set Twitter key}
\usage{
keys.twitter(access_token, persist = FALSE, show_output = FALSE)
}
\arguments{
\item{access_token}{(character length 1) API token}

\item{persist}{(logical length 1) If False, api key change will be contained to where it was changed. For example, a Jupyter notebook session.
If True, api key change will be global, i.e. it will affect terminal environment variables.
By default, False.}

\item{show_output}{(logical length 1) Display status string or not. By default, False.}
}
\description{
Wrapper for Python function keys.twitter from OpenBB Terminal SDK
}
\examples{
keys.twitter(persist=FALSE, show_output=FALSE)
}
