% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{login}
\alias{login}
\title{Login and load user info.}
\usage{
login(email = "", password = "", token = "", keep_session = FALSE)
}
\arguments{
\item{email}{(character length 1) The email.}

\item{password}{(character length 1) The password.}

\item{token}{(character length 1) The OpenBB Personal Access Token.}

\item{keep_session}{(logical length 1) Keep the session, i.e., next time the user logs in,
there is no need to enter the email and password or the token.}
}
\description{
Wrapper for Python function login from OpenBB Terminal SDK
}
