% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{news}
\alias{news}
\title{Get news for a given term and source. \link{Source: Feedparser}}
\usage{
news(term = "", sources = "", sort = "published")
}
\arguments{
\item{term}{(character length 1) term to search on the news articles}

\item{sources}{(character length 1) sources to exclusively show news from (separated by commas)}

\item{sort}{(character length 1) the column to sort by}
}
\description{
Wrapper for Python function news from OpenBB Terminal SDK
}
\examples{
news(sort='published')
}
