% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptos.R
\name{openbb_crypto_load}
\alias{openbb_crypto_load}
\title{openbb_crypto_load}
\usage{
openbb_crypto_load(
  symbol,
  start_date = "1950-01-01",
  end_date = NULL,
  interval = 1440,
  exchange = "binance",
  to_symbol = "usd",
  source = "CoinGecko"
)
}
\arguments{
\item{symbol}{string. Coin to get data}

\item{start_date}{Date or NULL. Start date to get data from with. - datetime or string format (YYYY-MM-DD). Default: NULL}

\item{end_date}{Date or NULL. End date to get data from with. - datetime or string format (YYYY-MM-DD). Default: NULL}

\item{interval}{numeric or string. The interval between data points in minutes. Choose from: 1, 15, 30, 60, 240, 1440, 10080, 43200. Default: 1440}

\item{exchange}{string. The exchange to get data from. Default: binance}

\item{to_symbol}{string. Quote Currency (Defaults to usdt). Default: usd}

\item{source}{string. The source of the data. Choose from: CCXT, CoinGecko, YahooFinance. Default: CoinGecko}
}
\description{
A wrapper for the openbb.crypto.load() function in Python
}
