% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{openbb_eod_option_chain}
\alias{openbb_eod_option_chain}
\title{openbb_eod_option_chain =====================================================================================}
\usage{
openbb_eod_option_chain(symbol, date, quiet = FALSE)
}
\arguments{
\item{symbol}{string. Symbol to get option chain for}

\item{date}{string. Date to get EOD chain for (YYYY-MM-DD)}

\item{quiet}{boolean. Default: FALSE - Flag to suppress progress bar}
}
\description{
A wrapper for the openbb.stocks.options.eodchain() function in Python
}
