% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etfs.R
\name{openbb_etf_load}
\alias{openbb_etf_load}
\title{openbb_etf_load}
\usage{
openbb_etf_load(
  symbol,
  start_date = "1950-01-01",
  end_date = NULL,
  interval = 1440,
  prepost = F,
  source = "YahooFinance",
  weekly = F,
  monthly = F,
  verbose = F
)
}
\arguments{
\item{symbol}{string. Ticker to get data}

\item{start_date}{Date or NULL. Start date to get data from with. - datetime or string format (YYYY-MM-DD). Default: NULL}

\item{end_date}{Date or NULL. End date to get data from with. - datetime or string format (YYYY-MM-DD). Default: NULL}

\item{interval}{numeric. Default: 1440 - Interval (in minutes) to get data 1, 5, 15, 30, 60, or 1440 (24hrs)}

\item{prepost}{boolean. Default: F - Append Pre/Post hours data}

\item{source}{string. Default: YahooFinance -  Other sources:}

\item{weekly}{boolean. Default: F}

\item{monthly}{boolean. Default: F}

\item{verbose}{boolean. Default: F}
}
\description{
A wrapper for the openbb.etf.load() function in Python
}
