% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{openbb_iv_surface}
\alias{openbb_iv_surface}
\title{openbb_iv_surface ===========================================================================================}
\usage{
openbb_iv_surface(symbol)
}
\arguments{
\item{symbol}{string. Stock ticker symbol to get}
}
\description{
A wrapper for the openbb.stocks.options.vsurf() function in Python
}
