% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{openbb_option_chains}
\alias{openbb_option_chains}
\title{openbb_option_chains ========================================================================================}
\usage{
openbb_option_chains(symbol, source = "Nasdaq", expiration = NULL)
}
\arguments{
\item{symbol}{string. Symbol to get chain for}

\item{source}{string. Default: "Nasdaq" - Source to get data from. Can be "YahooFinance", "Tradier", "Nasdaq", or "Intrinio"}

\item{expiration}{Date or NULL. Date to get chain for. By default returns all dates. Default: NULL}
}
\description{
A wrapper for the openbb.stocks.options.chains() function in Python
}
