% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api1.R
\name{openbb_prices}
\alias{openbb_prices}
\title{openbb_prices}
\usage{
openbb_prices(
  symbol,
  start_date,
  end_date,
  interval,
  prepost,
  source,
  weekly,
  monthly,
  verbose
)
}
\arguments{
\item{interval}{numeric. Default: 1440 - Interval (in minutes) to get data 1, 5, 15, 30, 60 or 1440 (24hrs)}

\item{prepost}{boolean. Default: F - Append Pre/Post hours data}

\item{source}{string. Default: yahooFinance -  Other sources:
AlphaVantage (https://www.alphavantage.co/documentation/)
Eod Historical Data (https://eodhistoricaldata.com/financial-apis/)
@param weekly boolean. Default: F
@param monthly boolean. Default: F
@param verbose boolean. Default: F}
}
\description{
a wrapper for openbb.stocks.load() function in python
}
