% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{openbb_stocks_load}
\alias{openbb_stocks_load}
\title{openbb_stocks_load}
\usage{
openbb_stocks_load(
  symbol = NULL,
  start_date = "1050-01-01",
  end_date = Sys.Date(),
  interval = 1440,
  prepost = F,
  source = "YahooFinance",
  weekly = F,
  monthly = F,
  verbose = F
)
}
\arguments{
\item{interval}{numeric. Default: 1440 - Interval (in minutes) to get data 1, 5, 15, 30, 60 or 1440 (24hrs)}

\item{prepost}{boolean. Default: F - Append Pre/Post hours data}

\item{source}{string. Default: yahooFinance -  Other sources:
AlphaVantage (https://www.alphavantage.co/documentation/)
Eod Historical Data (https://eodhistoricaldata.com/financial-apis/)
@param weekly boolean. Default: F
@param monthly boolean. Default: F
@param verbose boolean. Default: F}
}
\description{
a wrapper for openbb.stocks.load() function in python
}
