% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.alloc.assets}
\alias{portfolio.alloc.assets}
\title{Display portfolio asset allocation compared to the benchmark}
\usage{
portfolio.alloc.assets(
  portfolio_engine,
  tables = FALSE,
  limit = 10,
  recalculate = FALSE
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{tables}{(logical length 1) Whether to include separate allocation tables}

\item{limit}{(integer length 1) The amount of assets you wish to show, by default this is set to 10}

\item{recalculate}{(logical length 1) Flag to force recalculate allocation if already exists}
}
\description{
Wrapper for Python function portfolio.alloc.assets from OpenBB Terminal SDK
}
\examples{
portfolio.alloc.assets(tables=FALSE, limit=10, recalculate=FALSE)
}
