% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.bench}
\alias{portfolio.bench}
\title{Load benchmark into portfolio}
\usage{
portfolio.bench(portfolio_engine, symbol, full_shares = FALSE)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{symbol}{(character length 1) Benchmark symbol to download data}

\item{full_shares}{(logical length 1) Whether to mimic the portfolio trades exactly (partial shares) or round down the
quantity to the nearest number}
}
\description{
Wrapper for Python function portfolio.bench from OpenBB Terminal SDK
}
\examples{
portfolio.bench(symbol='AMZN', full_shares=FALSE)
}
