% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.distr_chart}
\alias{portfolio.distr_chart}
\title{Display daily returns}
\usage{
portfolio.distr_chart(
  portfolio_engine,
  window = "all",
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{raw}{(logical length 1) Display raw data from cumulative return}

\item{export}{(character length 1) Export certain type of data}
}
\description{
Wrapper for Python function portfolio.distr_chart from OpenBB Terminal SDK
}
