% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.dret_chart}
\alias{portfolio.dret_chart}
\title{Display daily returns}
\usage{
portfolio.dret_chart(
  portfolio_engine,
  window = "all",
  raw = FALSE,
  limit = 10,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) interval to compare cumulative returns and benchmark}

\item{raw}{(logical length 1) Display raw data from cumulative return}

\item{limit}{(integer length 1) Last daily returns to display}

\item{export}{(character length 1) Export certain type of data}
}
\description{
Wrapper for Python function portfolio.dret_chart from OpenBB Terminal SDK
}
\examples{
portfolio.dret_chart(window='all', limit=10)
}
