% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.es}
\alias{portfolio.es}
\title{Get portfolio expected shortfall}
\usage{
portfolio.es(
  portfolio_engine,
  use_mean = FALSE,
  distribution = "normal",
  percentile = 99.9
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{use_mean}{(logical length 1) if one should use the data mean return}

\item{distribution}{(character length 1) choose distribution to use: logistic, laplace, normal}

\item{percentile}{(numeric length 1) es percentile (\%)}
}
\description{
Wrapper for Python function portfolio.es from OpenBB Terminal SDK
}
\examples{
portfolio.es(distribution='normal', percentile=99.9)
}
