% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.holdp_chart}
\alias{portfolio.holdp_chart}
\title{Display holdings of assets (in percentage)}
\usage{
portfolio.holdp_chart(
  portfolio_engine,
  unstack = FALSE,
  raw = FALSE,
  limit = 10,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{unstack}{(logical length 1) Individual assets over time}

\item{raw}{(logical length 1) To display raw data}

\item{limit}{(integer length 1) Number of past market days to display holdings}

\item{export}{(character length 1) Format to export plot}
}
\description{
Wrapper for Python function portfolio.holdp_chart from OpenBB Terminal SDK
}
\examples{
portfolio.holdp_chart(unstack=FALSE, raw=FALSE, limit=10)
}
