% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.holdv}
\alias{portfolio.holdv}
\title{Get holdings of assets (absolute value)}
\usage{
portfolio.holdv(portfolio_engine)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}
}
\description{
Wrapper for Python function portfolio.holdv from OpenBB Terminal SDK
}
