% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.load}
\alias{portfolio.load}
\title{Get PortfolioEngine object}
\usage{
portfolio.load(
  transactions_file_path,
  benchmark_symbol = "SPY",
  full_shares = FALSE,
  risk_free_rate = 0
)
}
\arguments{
\item{transactions_file_path}{(character length 1) Path to transactions file}

\item{benchmark_symbol}{(character length 1) Benchmark ticker to download data}

\item{full_shares}{(logical length 1) Whether to mimic the portfolio trades exactly (partial shares) or round down the
quantity to the nearest number}

\item{risk_free_rate}{(numeric length 1) Risk free rate in float format}
}
\description{
Wrapper for Python function portfolio.load from OpenBB Terminal SDK
}
\examples{
portfolio.load(benchmark_symbol='SPY', full_shares=FALSE, risk_free_rate=0)
}
