% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.maxdd}
\alias{portfolio.maxdd}
\title{Calculate the drawdown (MDD) of historical series.  Note that the calculation is done}
\usage{
portfolio.maxdd(portfolio_engine, is_returns = FALSE)
}
\arguments{
\item{is_returns}{(logical length 1) Flag to indicate inputs are returns}
}
\description{
Wrapper for Python function portfolio.maxdd from OpenBB Terminal SDK
}
\examples{
portfolio.maxdd(is_returns=FALSE)
}
