% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.maxdd_chart}
\alias{portfolio.maxdd_chart}
\title{Display maximum drawdown curve}
\usage{
portfolio.maxdd_chart(portfolio_engine, export = "", sheet_name, external_axes)
}
\arguments{
\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function portfolio.maxdd_chart from OpenBB Terminal SDK
}
