% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.gaintopain}
\alias{portfolio.metric.gaintopain}
\title{Get Pain-to-Gain ratio based on historical data}
\usage{
portfolio.metric.gaintopain(portfolio_engine)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}
}
\description{
Wrapper for Python function portfolio.metric.gaintopain from OpenBB Terminal SDK
}
