% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.jensens}
\alias{portfolio.metric.jensens}
\title{Get jensen's alpha}
\usage{
portfolio.metric.jensens(portfolio_engine, risk_free_rate = 0, window = "1y")
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{risk_free_rate}{(numeric length 1) Risk free rate}

\item{window}{(character length 1) Interval used for rolling values}
}
\description{
Wrapper for Python function portfolio.metric.jensens from OpenBB Terminal SDK
}
\examples{
portfolio.metric.jensens(window='1y', risk_free_rate=0)
}
