% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.payoff}
\alias{portfolio.metric.payoff}
\title{Get payoff ratio}
\usage{
portfolio.metric.payoff(portfolio_engine)
}
\description{
Wrapper for Python function portfolio.metric.payoff from OpenBB Terminal SDK
}
