% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.skew}
\alias{portfolio.metric.skew}
\title{Get skewness for portfolio and benchmark selected}
\usage{
portfolio.metric.skew(portfolio_engine)
}
\description{
Wrapper for Python function portfolio.metric.skew from OpenBB Terminal SDK
}
