% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.sortino}
\alias{portfolio.metric.sortino}
\title{Get sortino ratio for portfolio and benchmark selected}
\usage{
portfolio.metric.sortino(portfolio_engine, risk_free_rate = 0)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{risk_free_rate}{(numeric length 1) Risk free rate value}
}
\description{
Wrapper for Python function portfolio.metric.sortino from OpenBB Terminal SDK
}
\examples{
portfolio.metric.sortino(risk_free_rate=0)
}
