% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.metric.trackerr}
\alias{portfolio.metric.trackerr}
\title{Get tracking error}
\usage{
portfolio.metric.trackerr(portfolio_engine, window = 252)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(integer length 1) Interval used for rolling values}
}
\description{
Wrapper for Python function portfolio.metric.trackerr from OpenBB Terminal SDK
}
\examples{
portfolio.metric.trackerr(window=252)
}
