% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.mret_chart}
\alias{portfolio.mret_chart}
\title{Display monthly returns}
\usage{
portfolio.mret_chart(
  portfolio_engine,
  window = "all",
  instrument = "both",
  graph = FALSE,
  show_vals = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) interval to compare cumulative returns and benchmark}

\item{instrument}{(character length 1) Display raw data from cumulative return, default is showing both the portfolio and benchmark in one table}

\item{show_vals}{(logical length 1) Show values on heatmap}

\item{export}{(character length 1) Export certain type of data}
}
\description{
Wrapper for Python function portfolio.mret_chart from OpenBB Terminal SDK
}
\examples{
portfolio.mret_chart(window='all')
}
