% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.om}
\alias{portfolio.om}
\title{Get omega ratio}
\usage{
portfolio.om(portfolio_engine, threshold_start = 0, threshold_end = 1.5)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{threshold_start}{(numeric length 1) annualized target return threshold start of plotted threshold range}

\item{threshold_end}{(numeric length 1) annualized target return threshold end of plotted threshold range}
}
\description{
Wrapper for Python function portfolio.om from OpenBB Terminal SDK
}
\examples{
portfolio.om(threshold_start=0, threshold_end=1.5)
}
