% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.perf}
\alias{portfolio.perf}
\title{Get portfolio performance vs the benchmark}
\usage{
portfolio.perf(portfolio_engine, show_all_trades = FALSE)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{show_all_trades}{(logical length 1) Whether to also show all trades made and their performance (default is False)}
}
\description{
Wrapper for Python function portfolio.perf from OpenBB Terminal SDK
}
\examples{
portfolio.perf(show_all_trades=FALSE)
}
