% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.blacklitterman}
\alias{portfolio.po.blacklitterman}
\title{Optimize decorrelation weights}
\usage{
portfolio.po.blacklitterman(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.blacklitterman from OpenBB Terminal SDK
}
