% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.ef}
\alias{portfolio.po.ef}
\title{Get Efficient Frontier}
\usage{
portfolio.po.ef(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.ef from OpenBB Terminal SDK
}
