% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.ef_chart}
\alias{portfolio.po.ef_chart}
\title{Display efficient frontier}
\usage{
portfolio.po.ef_chart(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.ef_chart from OpenBB Terminal SDK
}
