% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.file}
\alias{portfolio.po.file}
\title{Load portfolio optimization engine from file}
\usage{
portfolio.po.file(portfolio_engine, parameters_file_path)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_optimization.po_engine.PoEngine) Portfolio optimization engine, by default None
Use \code{portfolio.po.load} to load a portfolio engine}

\item{parameters_file_path}{(character length 1) Parameters file full path, by default None}
}
\description{
Wrapper for Python function portfolio.po.file from OpenBB Terminal SDK
}
