% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.herc}
\alias{portfolio.po.herc}
\title{Optimize with Hierarchical Equal Risk Contribution (HERC) method.}
\usage{
portfolio.po.herc(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.herc from OpenBB Terminal SDK
}
