% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.hrp}
\alias{portfolio.po.hrp}
\title{Optimize with Hierarchical Risk Parity}
\usage{
portfolio.po.hrp(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.hrp from OpenBB Terminal SDK
}
