% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.load}
\alias{portfolio.po.load}
\title{Load portfolio optimization engine}
\usage{
portfolio.po.load(symbols_categories, symbols_file_path, parameters_file_path)
}
\description{
Wrapper for Python function portfolio.po.load from OpenBB Terminal SDK
}
