% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.load_bl_views}
\alias{portfolio.po.load_bl_views}
\title{Load a Excel file with views for Black Litterman model.}
\usage{
portfolio.po.load_bl_views(excel_file = "")
}
\arguments{
\item{excel_file}{(character length 1) The location of the Excel file that needs to be loaded.}
}
\description{
Wrapper for Python function portfolio.po.load_bl_views from OpenBB Terminal SDK
}
