% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.maxdecorr}
\alias{portfolio.po.maxdecorr}
\title{Optimize decorrelation weights}
\usage{
portfolio.po.maxdecorr(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.maxdecorr from OpenBB Terminal SDK
}
