% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.maxdiv}
\alias{portfolio.po.maxdiv}
\title{Optimize diversification weights}
\usage{
portfolio.po.maxdiv(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.maxdiv from OpenBB Terminal SDK
}
