% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.maxret}
\alias{portfolio.po.maxret}
\title{Optimize maximum return weights}
\usage{
portfolio.po.maxret(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.maxret from OpenBB Terminal SDK
}
