% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.maxsharpe}
\alias{portfolio.po.maxsharpe}
\title{Optimize Sharpe ratio weights}
\usage{
portfolio.po.maxsharpe(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.maxsharpe from OpenBB Terminal SDK
}
