% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.maxutil}
\alias{portfolio.po.maxutil}
\title{Optimize maximum utility weights}
\usage{
portfolio.po.maxutil(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.maxutil from OpenBB Terminal SDK
}
