% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.minrisk}
\alias{portfolio.po.minrisk}
\title{Optimize minimum risk weights}
\usage{
portfolio.po.minrisk(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.minrisk from OpenBB Terminal SDK
}
