% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.nco}
\alias{portfolio.po.nco}
\title{Optimize with Non-Convex Optimization (NCO) model.}
\usage{
portfolio.po.nco(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.nco from OpenBB Terminal SDK
}
