% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.plot}
\alias{portfolio.po.plot}
\title{Display efficient frontier}
\usage{
portfolio.po.plot(portfolio_engine, chart_type = "pie", kwargs)
}
\arguments{
\item{chart_type}{(character length 1) Chart type, by default "pie"
Options are "pie", "hist", "dd" or "rc"}
}
\description{
Wrapper for Python function portfolio.po.plot from OpenBB Terminal SDK
}
\examples{
portfolio.po.plot(chart_type='pie')
}
