% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.relriskparity}
\alias{portfolio.po.relriskparity}
\title{Optimize with Relaxed Risk Parity using the least squares approach}
\usage{
portfolio.po.relriskparity(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.relriskparity from OpenBB Terminal SDK
}
