% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.riskparity}
\alias{portfolio.po.riskparity}
\title{Optimize with Risk Parity using the risk budgeting approach}
\usage{
portfolio.po.riskparity(portfolio_engine, kwargs)
}
\description{
Wrapper for Python function portfolio.po.riskparity from OpenBB Terminal SDK
}
