% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.po.show}
\alias{portfolio.po.show}
\title{Show portfolio optimization results}
\usage{
portfolio.po.show(portfolio_engine, category)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_optimization.po_engine.PoEngine) Portfolio optimization engine
Use \code{portfolio.po.load} to load a portfolio engine}
}
\description{
Wrapper for Python function portfolio.po.show from OpenBB Terminal SDK
}
