% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rbeta}
\alias{portfolio.rbeta}
\title{Get rolling beta using portfolio and benchmark returns}
\usage{
portfolio.rbeta(portfolio_engine, window = "1y")
}
\arguments{
\item{window}{(character length 1) Interval used for rolling values.
Possible options: mtd, qtd, ytd, 1d, 5d, 10d, 1m, 3m, 6m, 1y, 3y, 5y, 10y.}
}
\description{
Wrapper for Python function portfolio.rbeta from OpenBB Terminal SDK
}
\examples{
portfolio.rbeta(window='1y')
}
