% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rbeta_chart}
\alias{portfolio.rbeta_chart}
\title{Display rolling beta}
\usage{
portfolio.rbeta_chart(
  portfolio_engine,
  window = "1y",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{window}{(character length 1) interval for window to consider
Possible options: mtd, qtd, ytd, 1d, 5d, 10d, 1m, 3m, 6m, 1y, 3y, 5y, 10y.}

\item{export}{(character length 1) Export to file}
}
\description{
Wrapper for Python function portfolio.rbeta_chart from OpenBB Terminal SDK
}
\examples{
portfolio.rbeta_chart(window='1y')
}
