% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rsharpe_chart}
\alias{portfolio.rsharpe_chart}
\title{Display rolling sharpe}
\usage{
portfolio.rsharpe_chart(
  portfolio_engine,
  risk_free_rate = 0,
  window = "1y",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{risk_free_rate}{(numeric length 1) Value to use for risk free rate in sharpe/other calculations}

\item{window}{(character length 1) interval for window to consider}

\item{export}{(character length 1) Export to file}
}
\description{
Wrapper for Python function portfolio.rsharpe_chart from OpenBB Terminal SDK
}
\examples{
portfolio.rsharpe_chart(risk_free_rate=0, window='1y')
}
