% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rsort}
\alias{portfolio.rsort}
\title{Get rolling sortino}
\usage{
portfolio.rsort(portfolio_engine, risk_free_rate = 0, window = "1y")
}
\arguments{
\item{risk_free_rate}{(numeric length 1) Value to use for risk free rate in sharpe/other calculations}

\item{window}{(character length 1) interval for window to consider
Possible options: mtd, qtd, ytd, 1d, 5d, 10d, 1m, 3m, 6m, 1y, 3y, 5y, 10y}
}
\description{
Wrapper for Python function portfolio.rsort from OpenBB Terminal SDK
}
\examples{
portfolio.rsort(window='1y', risk_free_rate=0)
}
