% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rvol}
\alias{portfolio.rvol}
\title{Get rolling volatility}
\usage{
portfolio.rvol(portfolio_engine, window = "1y")
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) Rolling window size to use
Possible options: mtd, qtd, ytd, 1d, 5d, 10d, 1m, 3m, 6m, 1y, 3y, 5y, 10y}
}
\description{
Wrapper for Python function portfolio.rvol from OpenBB Terminal SDK
}
\examples{
portfolio.rvol(window='1y')
}
