% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.rvol_chart}
\alias{portfolio.rvol_chart}
\title{Display rolling volatility}
\usage{
portfolio.rvol_chart(
  portfolio_engine,
  window = "1y",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{export}{(character length 1) Export to file}
}
\description{
Wrapper for Python function portfolio.rvol_chart from OpenBB Terminal SDK
}
