% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.summary}
\alias{portfolio.summary}
\title{Get portfolio and benchmark returns summary}
\usage{
portfolio.summary(portfolio_engine, window = "all", risk_free_rate = 0)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) interval to compare cumulative returns and benchmark}

\item{risk_free_rate}{(numeric length 1) Risk free rate for calculations}
}
\description{
Wrapper for Python function portfolio.summary from OpenBB Terminal SDK
}
\examples{
portfolio.summary(window='all', risk_free_rate=0)
}
