% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.var}
\alias{portfolio.var}
\title{Get portfolio VaR}
\usage{
portfolio.var(
  portfolio_engine,
  use_mean = FALSE,
  adjusted_var = FALSE,
  student_t = FALSE,
  percentile = 99.9
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{use_mean}{(logical length 1) if one should use the data mean return}

\item{adjusted_var}{(logical length 1) if one should have VaR adjusted for skew and kurtosis (Cornish-Fisher-Expansion)}

\item{student_t}{(logical length 1) If one should use the student-t distribution}

\item{percentile}{(numeric length 1) var percentile (\%)}
}
\description{
Wrapper for Python function portfolio.var from OpenBB Terminal SDK
}
\examples{
portfolio.var(use_mean=FALSE, adjusted_var=FALSE, student_t=FALSE, percentile=99.9)
}
