% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.yret}
\alias{portfolio.yret}
\title{Get yearly returns}
\usage{
portfolio.yret(portfolio_engine, window = "all")
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) interval to compare cumulative returns and benchmark}
}
\description{
Wrapper for Python function portfolio.yret from OpenBB Terminal SDK
}
\examples{
portfolio.yret(window='all')
}
