% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio.yret_chart}
\alias{portfolio.yret_chart}
\title{Display yearly returns}
\usage{
portfolio.yret_chart(
  portfolio_engine,
  window = "all",
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{portfolio_engine}{(Python type: openbb_terminal.portfolio.portfolio_engine.PortfolioEngine) PortfolioEngine class instance, this will hold transactions and perform calculations.
Use \code{portfolio.load} to create a PortfolioEngine.}

\item{window}{(character length 1) interval to compare cumulative returns and benchmark}

\item{raw}{(logical length 1) Display raw data from cumulative return}

\item{export}{(character length 1) Export certain type of data}
}
\description{
Wrapper for Python function portfolio.yret_chart from OpenBB Terminal SDK
}
\examples{
portfolio.yret_chart(window='all')
}
