% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.acf}
\alias{qa.acf}
\title{Plots Auto and Partial Auto Correlation of returns and change in returns}
\usage{
qa.acf(data, target, symbol = "", lags = 15, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe to look at}

\item{target}{(character length 1) Data column to look at}

\item{symbol}{(character length 1) Name of dataset}

\item{lags}{(integer length 1) Max number of lags to look at}
}
\description{
Wrapper for Python function qa.acf from OpenBB Terminal SDK
}
\examples{
qa.acf(lags=15)
}
