% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.bw}
\alias{qa.bw}
\title{Plots box and whisker plots}
\usage{
qa.bw(data, target, symbol = "", yearly = TRUE, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe to look at}

\item{target}{(character length 1) Data column to look at}

\item{symbol}{(character length 1) Name of dataset}

\item{yearly}{(logical length 1) Flag to indicate yearly accumulation}
}
\description{
Wrapper for Python function qa.bw from OpenBB Terminal SDK
}
\examples{
qa.bw(yearly=TRUE)
}
