% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.calculate_adjusted_var}
\alias{qa.calculate_adjusted_var}
\title{Calculates VaR, which is adjusted for skew and kurtosis (Cornish-Fischer-Expansion)}
\usage{
qa.calculate_adjusted_var(kurtosis, skew, ndp, std, mean)
}
\arguments{
\item{kurtosis}{(numeric length 1) kurtosis of data}

\item{skew}{(numeric length 1) skew of data}

\item{ndp}{(numeric length 1) normal distribution percentage number (99\% -> -2.326)}

\item{std}{(numeric length 1) standard deviation of data}

\item{mean}{(numeric length 1) mean of data}
}
\description{
Wrapper for Python function qa.calculate_adjusted_var from OpenBB Terminal SDK
}
