% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.cdf}
\alias{qa.cdf}
\title{Plots Cumulative Distribution Function}
\usage{
qa.cdf(data, target, symbol = "", export = "", sheet_name, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe to look at}

\item{target}{(character length 1) Data column}

\item{symbol}{(character length 1) Name of dataset}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.cdf from OpenBB Terminal SDK
}
