% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.cusum}
\alias{qa.cusum}
\title{Plots Cumulative sum algorithm (CUSUM) to detect abrupt changes in data}
\usage{
qa.cusum(data, target, threshold = 5, drift = 2.1, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe}

\item{target}{(character length 1) Column of data to look at}

\item{threshold}{(numeric length 1) Threshold value}

\item{drift}{(numeric length 1) Drift parameter}
}
\description{
Wrapper for Python function qa.cusum from OpenBB Terminal SDK
}
\examples{
qa.cusum(threshold=5, drift=2.1)
}
