% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.decompose}
\alias{qa.decompose}
\title{Perform seasonal decomposition}
\usage{
qa.decompose(data, multiplicative = FALSE)
}
\arguments{
\item{data}{(data.frame) Dataframe of targeted data}

\item{multiplicative}{(logical length 1) Boolean to indicate multiplication instead of addition}
}
\description{
Wrapper for Python function qa.decompose from OpenBB Terminal SDK
}
\examples{
qa.decompose(multiplicative=FALSE)
}
