% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.es}
\alias{qa.es}
\title{Gets Expected Shortfall for specified stock dataframe.}
\usage{
qa.es(
  data,
  use_mean = FALSE,
  distribution = "normal",
  percentile,
  portfolio = FALSE
)
}
\arguments{
\item{data}{(data.frame) Data dataframe}

\item{use_mean}{(logical length 1) If one should use the data mean for calculation}

\item{distribution}{(character length 1) Type of distribution, options: laplace, student_t, normal}

\item{portfolio}{(logical length 1) If the data is a portfolio}
}
\description{
Wrapper for Python function qa.es from OpenBB Terminal SDK
}
\examples{
qa.es(use_mean=FALSE, distribution='normal', percentile=99.9, portfolio=FALSE)
}
