% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.es_chart}
\alias{qa.es_chart}
\title{Prints table showing expected shortfall.}
\usage{
qa.es_chart(
  data,
  symbol = "",
  use_mean = FALSE,
  distribution = "normal",
  percentile = 99.9,
  portfolio = FALSE
)
}
\arguments{
\item{data}{(data.frame) Data dataframe}

\item{symbol}{(character length 1) name of the data}

\item{use_mean}{(logical length 1) if one should use the data mean return}

\item{distribution}{(character length 1) choose distribution to use: logistic, laplace, normal}

\item{percentile}{(numeric length 1) es percentile}

\item{portfolio}{(logical length 1) If the data is a portfolio}
}
\description{
Wrapper for Python function qa.es_chart from OpenBB Terminal SDK
}
\examples{
qa.es_chart(distribution='normal', percentile=99, portfolio=FALSE)
}
