% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.hist}
\alias{qa.hist}
\title{Plots histogram of data}
\usage{
qa.hist(data, target, symbol = "", bins = 15, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe to look at}

\item{target}{(character length 1) Data column to get histogram of the dataframe}

\item{symbol}{(character length 1) Name of dataset}

\item{bins}{(integer length 1) Number of bins in histogram}
}
\description{
Wrapper for Python function qa.hist from OpenBB Terminal SDK
}
\examples{
qa.hist(bins=15)
}
