% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.kurtosis}
\alias{qa.kurtosis}
\title{Kurtosis Indicator}
\usage{
qa.kurtosis(data, window = 14)
}
\arguments{
\item{data}{(data.frame) Dataframe of targeted data}

\item{window}{(integer length 1) Length of window}
}
\description{
Wrapper for Python function qa.kurtosis from OpenBB Terminal SDK
}
\examples{
qa.kurtosis(window=14)
}
