% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.kurtosis_chart}
\alias{qa.kurtosis_chart}
\title{Plots rolling kurtosis}
\usage{
qa.kurtosis_chart(
  symbol,
  data,
  target,
  window = 14,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker}

\item{data}{(data.frame) Dataframe of stock prices}

\item{target}{(character length 1) Column in data to look at}

\item{window}{(integer length 1) Length of window}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.kurtosis_chart from OpenBB Terminal SDK
}
\examples{
qa.kurtosis_chart(symbol='AMZN', window=14)
}
