% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.line}
\alias{qa.line}
\title{Display line plot of data}
\usage{
qa.line(
  data,
  title = "",
  log_y = TRUE,
  markers_lines,
  markers_scatter,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Data to plot}

\item{title}{(character length 1) Title for plot}

\item{log_y}{(logical length 1) Flag for showing y on log scale}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.line from OpenBB Terminal SDK
}
\examples{
qa.line(log_y=TRUE)
}
