% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.normality}
\alias{qa.normality}
\title{Look at the distribution of returns and generate statistics on the relation to the normal curve.}
\usage{
qa.normality(data)
}
\arguments{
\item{data}{(data.frame) Dataframe of targeted data}
}
\description{
Wrapper for Python function qa.normality from OpenBB Terminal SDK
}
