% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.normality_chart}
\alias{qa.normality_chart}
\title{Prints table showing normality statistics}
\usage{
qa.normality_chart(data, target, export = "", sheet_name)
}
\arguments{
\item{data}{(data.frame) DataFrame}

\item{target}{(character length 1) Column in data to look at}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.normality_chart from OpenBB Terminal SDK
}
