% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.omega_chart}
\alias{qa.omega_chart}
\title{Plots the omega ratio}
\usage{
qa.omega_chart(data, threshold_start = 0, threshold_end = 1.5)
}
\arguments{
\item{data}{(data.frame) stock dataframe}

\item{threshold_start}{(numeric length 1) annualized target return threshold start of plotted threshold range}

\item{threshold_end}{(numeric length 1) annualized target return threshold end of plotted threshold range}
}
\description{
Wrapper for Python function qa.omega_chart from OpenBB Terminal SDK
}
\examples{
qa.omega_chart(threshold_start=0, threshold_end=1.5)
}
