% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.qqplot}
\alias{qa.qqplot}
\title{Plots QQ plot for data against normal quantiles}
\usage{
qa.qqplot(data, target, symbol = "", external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe}

\item{target}{(character length 1) Column in data to look at}

\item{symbol}{(character length 1) Stock ticker}
}
\description{
Wrapper for Python function qa.qqplot from OpenBB Terminal SDK
}
