% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.quantile}
\alias{qa.quantile}
\title{Overlay Median & Quantile}
\usage{
qa.quantile(data, window = 14, quantile_pct = 0.5)
}
\arguments{
\item{data}{(data.frame) Dataframe of targeted data}

\item{window}{(integer length 1) Length of window}

\item{quantile_pct}{(numeric length 1) Quantile to display}
}
\description{
Wrapper for Python function qa.quantile from OpenBB Terminal SDK
}
\examples{
qa.quantile(window=14, quantile_pct=0.5)
}
