% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.quantile_chart}
\alias{qa.quantile_chart}
\title{Plots rolling quantile}
\usage{
qa.quantile_chart(
  data,
  target,
  symbol = "",
  window = 14,
  quantile = 0.5,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe}

\item{target}{(character length 1) Column in data to look at}

\item{symbol}{(character length 1) Stock ticker}

\item{window}{(integer length 1) Length of window}

\item{quantile}{(numeric length 1) Quantile to get}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.quantile_chart from OpenBB Terminal SDK
}
\examples{
qa.quantile_chart(window=14, quantile=0.5)
}
