% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.rolling}
\alias{qa.rolling}
\title{Return rolling mean and standard deviation}
\usage{
qa.rolling(data, window = 14)
}
\arguments{
\item{data}{(data.frame) Dataframe of target data}

\item{window}{(integer length 1) Length of rolling window}
}
\description{
Wrapper for Python function qa.rolling from OpenBB Terminal SDK
}
\examples{
qa.rolling(window=14)
}
