% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.sharpe}
\alias{qa.sharpe}
\title{Calculates the sharpe ratio}
\usage{
qa.sharpe(data, rfr = 0, window = 252)
}
\arguments{
\item{data}{(data.frame) selected dataframe column}

\item{rfr}{(numeric length 1) risk free rate}

\item{window}{(numeric length 1) length of the rolling window}
}
\description{
Wrapper for Python function qa.sharpe from OpenBB Terminal SDK
}
\examples{
qa.sharpe(rfr=0, window=252)
}
