% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.skew}
\alias{qa.skew}
\title{Skewness Indicator}
\usage{
qa.skew(data, window = 14)
}
\arguments{
\item{data}{(data.frame) Dataframe of targeted data}

\item{window}{(integer length 1) Length of window}
}
\description{
Wrapper for Python function qa.skew from OpenBB Terminal SDK
}
\examples{
qa.skew(window=14)
}
