% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.sortino}
\alias{qa.sortino}
\title{Calculates the sortino ratio}
\usage{
qa.sortino(data, target_return = 0, window = 252, adjusted = FALSE)
}
\arguments{
\item{data}{(data.frame) selected dataframe}

\item{target_return}{(numeric length 1) target return of the asset}

\item{window}{(numeric length 1) length of the rolling window}

\item{adjusted}{(logical length 1) adjust the sortino ratio}
}
\description{
Wrapper for Python function qa.sortino from OpenBB Terminal SDK
}
\examples{
qa.sortino(target_return=0, window=252, adjusted=FALSE)
}
