% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.sortino_chart}
\alias{qa.sortino_chart}
\title{Plots the sortino ratio}
\usage{
qa.sortino_chart(data, target_return, window, adjusted = TRUE)
}
\arguments{
\item{data}{(data.frame) selected dataframe}

\item{target_return}{(numeric length 1) target return of the asset}

\item{window}{(numeric length 1) length of the rolling window}

\item{adjusted}{(logical length 1) adjust the sortino ratio}
}
\description{
Wrapper for Python function qa.sortino_chart from OpenBB Terminal SDK
}
