% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.spread}
\alias{qa.spread}
\title{Standard Deviation and Variance}
\usage{
qa.spread(data, window = 14)
}
\arguments{
\item{data}{(data.frame) DataFrame of targeted data}

\item{window}{(integer length 1) Length of window}
}
\description{
Wrapper for Python function qa.spread from OpenBB Terminal SDK
}
\examples{
qa.spread(window=14)
}
