% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.spread_chart}
\alias{qa.spread_chart}
\title{Plots rolling spread}
\usage{
qa.spread_chart(
  data,
  target,
  symbol = "",
  window = 14,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe}

\item{target}{(character length 1) Column in data to look at}

\item{symbol}{(character length 1) Stock ticker}

\item{window}{(integer length 1) Length of window}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.spread_chart from OpenBB Terminal SDK
}
\examples{
qa.spread_chart(window=14)
}
