% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.summary}
\alias{qa.summary}
\title{Print summary statistics}
\usage{
qa.summary(data)
}
\arguments{
\item{data}{(data.frame) Dataframe to get summary statistics for}
}
\description{
Wrapper for Python function qa.summary from OpenBB Terminal SDK
}
