% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.summary_chart}
\alias{qa.summary_chart}
\title{Prints table showing summary statistics}
\usage{
qa.summary_chart(data, export = "", sheet_name)
}
\arguments{
\item{data}{(data.frame) DataFrame to get statistics of}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function qa.summary_chart from OpenBB Terminal SDK
}
