% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.unitroot_chart}
\alias{qa.unitroot_chart}
\title{Prints table showing unit root test calculations}
\usage{
qa.unitroot_chart(
  data,
  target,
  fuller_reg = "c",
  kpss_reg = "c",
  export = "",
  sheet_name
)
}
\arguments{
\item{data}{(data.frame) DataFrame}

\item{target}{(character length 1) Column of data to look at}

\item{fuller_reg}{(character length 1) Type of regression of ADF test. Can be ‘c’,’ct’,’ctt’,’nc’ 'c' - Constant and t - trend order}

\item{kpss_reg}{(character length 1) Type of regression for KPSS test. Can be ‘c’,’ct'}

\item{export}{(character length 1) Format for exporting data}
}
\description{
Wrapper for Python function qa.unitroot_chart from OpenBB Terminal SDK
}
\examples{
qa.unitroot_chart(fuller_reg='c', kpss_reg='c')
}
