% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.var}
\alias{qa.var}
\title{Gets value at risk for specified stock dataframe.}
\usage{
qa.var(
  data,
  use_mean = FALSE,
  adjusted_var = FALSE,
  student_t = FALSE,
  percentile,
  portfolio = FALSE
)
}
\arguments{
\item{data}{(data.frame) Data dataframe}

\item{use_mean}{(logical length 1) If one should use the data mean for calculation}

\item{adjusted_var}{(logical length 1) If one should return VaR adjusted for skew and kurtosis}

\item{student_t}{(logical length 1) If one should use the student-t distribution}

\item{portfolio}{(logical length 1) If the data is a portfolio}
}
\description{
Wrapper for Python function qa.var from OpenBB Terminal SDK
}
\examples{
qa.var(use_mean=FALSE, adjusted_var=FALSE, student_t=FALSE, percentile=99.9, portfolio=FALSE)
}
