% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qa.var_chart}
\alias{qa.var_chart}
\title{Prints table showing VaR of dataframe.}
\usage{
qa.var_chart(
  data,
  symbol = "",
  use_mean = FALSE,
  adjusted_var = FALSE,
  student_t = FALSE,
  percentile = 99.9,
  data_range = 0,
  portfolio = FALSE
)
}
\arguments{
\item{data}{(data.frame) Data dataframe}

\item{symbol}{(character length 1) name of the data}

\item{use_mean}{(logical length 1) if one should use the data mean return}

\item{adjusted_var}{(logical length 1) if one should have VaR adjusted for skew and kurtosis (Cornish-Fisher-Expansion)}

\item{student_t}{(logical length 1) If one should use the student-t distribution}

\item{percentile}{(numeric length 1) var percentile}

\item{data_range}{(integer length 1) Number of rows you want to use VaR over}

\item{portfolio}{(logical length 1) If the data is a portfolio}
}
\description{
Wrapper for Python function qa.var_chart from OpenBB Terminal SDK
}
\examples{
qa.var_chart(use_mean=FALSE, adjusted_var=FALSE, student_t=FALSE, percentile=99, data_range=0, portfolio=FALSE)
}
