% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptos.R
\name{run_cryptos}
\alias{run_cryptos}
\title{Run crptos API}
\usage{
run_cryptos(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run crptos API
}
\examples{
\dontrun{
run_cryptos()
run_scryptosswagger = TRUE, port = 8000)
}
}
