% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etfs.R
\name{run_etfs}
\alias{run_etfs}
\title{Run etfs API}
\usage{
run_etfs(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run etfs API
}
\examples{
\dontrun{
run_etfs()
run_etfs(swagger = TRUE, port = 8000)
}
}
