% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{run_options}
\alias{run_options}
\title{Run options API =============================================================================================}
\usage{
run_options(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run options API =============================================================================================
}
\examples{
\dontrun{
run_options()
run_options(swagger = TRUE, port = 8000)
}
}
