% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{run_stocks}
\alias{run_stocks}
\title{Run stocks API}
\usage{
run_stocks(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run stocks API
}
\examples{
\dontrun{
run_stocks()
run_stocks(swagger = TRUE, port = 8000)
}
}
