% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.cnews}
\alias{stocks.ba.cnews}
\title{Get news from a company. \link{Source: Finnhub}}
\usage{
stocks.ba.cnews(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{(character length 1) company ticker to look for news articles}
}
\description{
Wrapper for Python function stocks.ba.cnews from OpenBB Terminal SDK
}
\examples{
stocks.ba.cnews(symbol='AMZN')
}
