% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.getdd}
\alias{stocks.ba.getdd}
\title{Gets due diligence posts from list of subreddits \link{Source: reddit}.}
\usage{
stocks.ba.getdd(symbol, limit = 5, n_days = 3, show_all_flairs = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{limit}{(integer length 1) Number of posts to get}

\item{n_days}{(integer length 1) Number of days back to get posts}

\item{show_all_flairs}{(logical length 1) Search through all flairs (apart from Yolo and Meme)}
}
\description{
Wrapper for Python function stocks.ba.getdd from OpenBB Terminal SDK
}
\examples{
stocks.ba.getdd(symbol='AMZN', limit=5, n_days=3, show_all_flairs=FALSE)
}
