% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.headlines}
\alias{stocks.ba.headlines}
\title{Gets Sentiment analysis provided by FinBrain's API \link{Source: finbrain}.}
\usage{
stocks.ba.headlines(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the sentiment analysis from}
}
\description{
Wrapper for Python function stocks.ba.headlines from OpenBB Terminal SDK
}
\examples{
stocks.ba.headlines(symbol='AMZN')
}
