% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.headlines_chart}
\alias{stocks.ba.headlines_chart}
\title{Plots Sentiment analysis from FinBrain. Prints table if raw is True. \link{Source: FinBrain}}
\usage{
stocks.ba.headlines_chart(
  symbol,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the sentiment analysis from}

\item{raw}{(logical length 1) Display raw table data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.ba.headlines_chart from OpenBB Terminal SDK
}
\examples{
stocks.ba.headlines_chart(symbol='AMZN')
}
