% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.infer_chart}
\alias{stocks.ba.infer_chart}
\title{Prints Inference sentiment from past n tweets.}
\usage{
stocks.ba.infer_chart(symbol, limit = 100, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number of tweets to analyze}

\item{export}{(character length 1) Format to export tweet dataframe}
}
\description{
Wrapper for Python function stocks.ba.infer_chart from OpenBB Terminal SDK
}
\examples{
stocks.ba.infer_chart(symbol='AMZN', limit=100)
}
