% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.mentions}
\alias{stocks.ba.mentions}
\title{Get interest over time from google api \link{Source: google}.}
\usage{
stocks.ba.mentions(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.ba.mentions from OpenBB Terminal SDK
}
\examples{
stocks.ba.mentions(symbol='AMZN')
}
