% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.messages}
\alias{stocks.ba.messages}
\title{Get last messages for a given ticker \link{Source: stocktwits}.}
\usage{
stocks.ba.messages(symbol, limit = 30)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number of messages to get}
}
\description{
Wrapper for Python function stocks.ba.messages from OpenBB Terminal SDK
}
\examples{
stocks.ba.messages(symbol='AMZN', limit=30)
}
