% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.popular}
\alias{stocks.ba.popular}
\title{Get popular tickers from list of subreddits \link{Source: reddit}.}
\usage{
stocks.ba.popular(limit = 10, post_limit = 50, subreddits = "")
}
\arguments{
\item{limit}{(integer length 1) Number of top tickers to get}

\item{post_limit}{(integer length 1) How many posts to analyze in each subreddit}

\item{subreddits}{(character length 1) String of comma separated subreddits.}
}
\description{
Wrapper for Python function stocks.ba.popular from OpenBB Terminal SDK
}
\examples{
stocks.ba.popular(limit=10, post_limit=50)
}
