% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.queries}
\alias{stocks.ba.queries}
\title{Get related queries from google api \link{Source: google}.}
\usage{
stocks.ba.queries(symbol, limit = 10)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol to compare}

\item{limit}{(integer length 1) Number of queries to show}
}
\description{
Wrapper for Python function stocks.ba.queries from OpenBB Terminal SDK
}
\examples{
stocks.ba.queries(symbol='AMZN', limit=10)
}
