% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.redditsent}
\alias{stocks.ba.redditsent}
\title{Finds posts related to a specific search term in Reddit.}
\usage{
stocks.ba.redditsent(
  symbol,
  limit = 100,
  sortby = "relevance",
  time_frame = "week",
  full_search = TRUE,
  subreddits = "all"
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to search for}

\item{limit}{(integer length 1) Number of posts to get per subreddit}

\item{sortby}{(character length 1) Search type (Possibilities: "relevance", "hot", "top", "new", or "comments")}

\item{time_frame}{(character length 1) Relative time of post (Possibilities: "hour", "day", "week", "month", "year", "all")}

\item{full_search}{(logical length 1) Enable comprehensive search for ticker}

\item{subreddits}{(character length 1) Comma-separated list of subreddits}
}
\description{
Wrapper for Python function stocks.ba.redditsent from OpenBB Terminal SDK
}
\examples{
stocks.ba.redditsent(symbol='AMZN', limit=100, sortby='relevance', time_frame='week', full_search=TRUE, subreddits='all')
}
