% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.regions}
\alias{stocks.ba.regions}
\title{Get interest by region from google api \link{Source: google}.}
\usage{
stocks.ba.regions(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to look at}
}
\description{
Wrapper for Python function stocks.ba.regions from OpenBB Terminal SDK
}
\examples{
stocks.ba.regions(symbol='AMZN')
}
