% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.regions_chart}
\alias{stocks.ba.regions_chart}
\title{Plots bars of regions based on stock's interest. \link{Source: Google}.}
\usage{
stocks.ba.regions_chart(
  symbol,
  limit = 5,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol}

\item{limit}{(integer length 1) Number of regions to show}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.ba.regions_chart from OpenBB Terminal SDK
}
\examples{
stocks.ba.regions_chart(symbol='AMZN', limit=5)
}
