% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.rise}
\alias{stocks.ba.rise}
\title{Get top rising related queries with this stock's query \link{Source: google}.}
\usage{
stocks.ba.rise(symbol, limit = 10)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number of queries to show}
}
\description{
Wrapper for Python function stocks.ba.rise from OpenBB Terminal SDK
}
\examples{
stocks.ba.rise(symbol='AMZN', limit=10)
}
