% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.sentiment}
\alias{stocks.ba.sentiment}
\title{Get sentiments from symbol.}
\usage{
stocks.ba.sentiment(symbol, n_tweets = 15, n_days_past = 2)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol to get sentiment for}

\item{n_tweets}{(integer length 1) Number of tweets to get per hour}

\item{n_days_past}{(integer length 1) Number of days to extract tweets for}
}
\description{
Wrapper for Python function stocks.ba.sentiment from OpenBB Terminal SDK
}
\examples{
stocks.ba.sentiment(symbol='AMZN', n_tweets=15, n_days_past=2)
}
