% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.sentiment_chart}
\alias{stocks.ba.sentiment_chart}
\title{Plots sentiments from symbol}
\usage{
stocks.ba.sentiment_chart(
  symbol,
  n_tweets = 15,
  n_days_past = 2,
  compare = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol to get sentiment for}

\item{n_tweets}{(integer length 1) Number of tweets to get per hour}

\item{n_days_past}{(integer length 1) Number of days to extract tweets for}

\item{compare}{(logical length 1) Show corresponding change in stock price}

\item{export}{(character length 1) Format to export tweet dataframe}
}
\description{
Wrapper for Python function stocks.ba.sentiment_chart from OpenBB Terminal SDK
}
\examples{
stocks.ba.sentiment_chart(symbol='AMZN', n_tweets=15, n_days_past=2, compare=FALSE)
}
