% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.snews}
\alias{stocks.ba.snews}
\title{Get headlines sentiment using VADER model over time. \link{Source: Finnhub}}
\usage{
stocks.ba.snews(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker of company}
}
\description{
Wrapper for Python function stocks.ba.snews from OpenBB Terminal SDK
}
\examples{
stocks.ba.snews(symbol='AMZN')
}
