% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.snews_chart}
\alias{stocks.ba.snews_chart}
\title{Display stock price and headlines sentiment using VADER model over time. \link{Source: Finnhub}}
\usage{
stocks.ba.snews_chart(symbol, export = "", sheet_name, external_axes)
}
\arguments{
\item{symbol}{(character length 1) Ticker of company}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.ba.snews_chart from OpenBB Terminal SDK
}
\examples{
stocks.ba.snews_chart(symbol='AMZN')
}
