% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.spac}
\alias{stocks.ba.spac}
\title{Get posts containing SPAC from top subreddits \link{Source: reddit}.}
\usage{
stocks.ba.spac(limit = 5)
}
\arguments{
\item{limit}{(integer length 1) Number of posts to get for each subreddit, by default 5}
}
\description{
Wrapper for Python function stocks.ba.spac from OpenBB Terminal SDK
}
\examples{
stocks.ba.spac(limit=5)
}
