% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.spacc}
\alias{stocks.ba.spacc}
\title{Get top tickers from r/SPACs \link{Source: reddit}.}
\usage{
stocks.ba.spacc(limit = 10, popular = FALSE)
}
\arguments{
\item{limit}{(integer length 1) Number of posts to look at}

\item{popular}{(logical length 1) Search by hot instead of new}
}
\description{
Wrapper for Python function stocks.ba.spacc from OpenBB Terminal SDK
}
\examples{
stocks.ba.spacc(limit=10, popular=FALSE)
}
