% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.stalker}
\alias{stocks.ba.stalker}
\title{Gets messages from given user \link{Source: stocktwits}.}
\usage{
stocks.ba.stalker(user, limit = 30)
}
\arguments{
\item{user}{(character length 1) User to get posts for}

\item{limit}{(integer length 1) Number of posts to get, by default 30}
}
\description{
Wrapper for Python function stocks.ba.stalker from OpenBB Terminal SDK
}
\examples{
stocks.ba.stalker(limit=30)
}
