% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.text_sent}
\alias{stocks.ba.text_sent}
\title{Find the sentiment of a post and related comments.}
\usage{
stocks.ba.text_sent(post_data)
}
\description{
Wrapper for Python function stocks.ba.text_sent from OpenBB Terminal SDK
}
