% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.trending}
\alias{stocks.ba.trending}
\title{Get trending tickers from stocktwits \link{Source: stocktwits}.}
\usage{
stocks.ba.trending()
}
\description{
Wrapper for Python function stocks.ba.trending from OpenBB Terminal SDK
}
