% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.watchlist}
\alias{stocks.ba.watchlist}
\title{Get reddit users watchlists \link{Source: reddit}.}
\usage{
stocks.ba.watchlist(limit = 5)
}
\arguments{
\item{limit}{(integer length 1) Number of posts to look through}
}
\description{
Wrapper for Python function stocks.ba.watchlist from OpenBB Terminal SDK
}
\examples{
stocks.ba.watchlist(limit=5)
}
