% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ba.wsb}
\alias{stocks.ba.wsb}
\title{Get wsb posts \link{Source: reddit}.}
\usage{
stocks.ba.wsb(limit = 10, new = FALSE)
}
\arguments{
\item{limit}{(integer length 1) Number of posts to get, by default 10}

\item{new}{(logical length 1) Flag to sort by new instead of hot, by default False}
}
\description{
Wrapper for Python function stocks.ba.wsb from OpenBB Terminal SDK
}
\examples{
stocks.ba.wsb(limit=10, new=FALSE)
}
