% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.cashflow}
\alias{stocks.ca.cashflow}
\title{Get cashflow data. \link{Source: Marketwatch}}
\usage{
stocks.ca.cashflow(similar, timeframe = "2022", quarter = FALSE)
}
\arguments{
\item{timeframe}{(character length 1) Column header to compare}

\item{quarter}{(logical length 1) Whether to use quarterly statements, by default False}
}
\description{
Wrapper for Python function stocks.ca.cashflow from OpenBB Terminal SDK
}
\examples{
stocks.ca.cashflow(timeframe='2021', quarter=FALSE)
}
