% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.hcorr}
\alias{stocks.ca.hcorr}
\title{Get historical price correlation. \link{Source: Yahoo Finance}}
\usage{
stocks.ca.hcorr(similar, start_date, end_date, candle_type = "a")
}
\arguments{
\item{candle_type}{(character length 1) OHLCA column to use for candles or R for returns, by default "a" for Adjusted Close}
}
\description{
Wrapper for Python function stocks.ca.hcorr from OpenBB Terminal SDK
}
\examples{
stocks.ca.hcorr(candle_type='a')
}
