% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.hcorr_chart}
\alias{stocks.ca.hcorr_chart}
\title{Correlation heatmap based on historical price comparison}
\usage{
stocks.ca.hcorr_chart(
  similar,
  start_date,
  end_date,
  candle_type = "a",
  display_full_matrix = FALSE,
  raw = FALSE,
  external_axes,
  export = "",
  sheet_name
)
}
\arguments{
\item{candle_type}{(character length 1) OHLCA column to use for candles or R for returns, by default "a" for Adjusted Close}

\item{display_full_matrix}{(logical length 1) Optionally display all values in the matrix, rather than masking off half, by default False}

\item{raw}{(logical length 1) Whether to display raw data}

\item{export}{(character length 1) Format to export correlation prices, by default ""}
}
\description{
Wrapper for Python function stocks.ca.hcorr_chart from OpenBB Terminal SDK
}
\examples{
stocks.ca.hcorr_chart(candle_type='a', display_full_matrix=FALSE, raw=FALSE)
}
