% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.hist}
\alias{stocks.ca.hist}
\title{Get historical prices for all comparison stocks}
\usage{
stocks.ca.hist(similar, start_date, end_date, candle_type = "a")
}
\arguments{
\item{candle_type}{(character length 1) Candle variable to compare, by default "a" for Adjusted Close. Possible values are: o, h, l, c, a, v, r}
}
\description{
Wrapper for Python function stocks.ca.hist from OpenBB Terminal SDK
}
\examples{
stocks.ca.hist(candle_type='a')
}
