% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.hist_chart}
\alias{stocks.ca.hist_chart}
\title{Display historical stock prices. \link{Source: Yahoo Finance}}
\usage{
stocks.ca.hist_chart(
  similar,
  start_date,
  end_date,
  candle_type = "a",
  normalize = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{candle_type}{(character length 1) OHLCA column to use or R to use daily returns calculated from Adjusted Close, by default "a" for Adjusted Close}

\item{normalize}{(logical length 1) Boolean to normalize all stock prices using MinMax defaults True}

\item{export}{(character length 1) Format to export historical prices, by default ""}
}
\description{
Wrapper for Python function stocks.ca.hist_chart from OpenBB Terminal SDK
}
\examples{
stocks.ca.hist_chart(candle_type='a', normalize=TRUE)
}
