% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.income_chart}
\alias{stocks.ca.income_chart}
\title{Display income data. \link{Source: Marketwatch}.}
\usage{
stocks.ca.income_chart(
  symbols,
  timeframe = "2022",
  quarter = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{timeframe}{(character length 1) What year to look at}

\item{quarter}{(logical length 1) Whether to use quarterly statements, by default False}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.ca.income_chart from OpenBB Terminal SDK
}
\examples{
stocks.ca.income_chart(timeframe='2021', quarter=FALSE)
}
