% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.scorr}
\alias{stocks.ca.scorr}
\title{Get correlation sentiments across similar companies. \link{Source: FinBrain}.}
\usage{
stocks.ca.scorr(similar)
}
\description{
Wrapper for Python function stocks.ca.scorr from OpenBB Terminal SDK
}
