% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.screener}
\alias{stocks.ca.screener}
\title{Screener Overview.}
\usage{
stocks.ca.screener(similar, data_type = "overview")
}
\arguments{
\item{data_type}{(character length 1) Data type between: overview, valuation, financial, ownership, performance, technical}
}
\description{
Wrapper for Python function stocks.ca.screener from OpenBB Terminal SDK
}
\examples{
stocks.ca.screener(data_type='overview')
}
