% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.sentiment}
\alias{stocks.ca.sentiment}
\title{Gets Sentiment analysis from several symbols provided by FinBrain's API.}
\usage{
stocks.ca.sentiment(symbols)
}
\description{
Wrapper for Python function stocks.ca.sentiment from OpenBB Terminal SDK
}
