% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.sentiment_chart}
\alias{stocks.ca.sentiment_chart}
\title{Display sentiment for all ticker. \link{Source: FinBrain}.}
\usage{
stocks.ca.sentiment_chart(
  similar,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{raw}{(logical length 1) Output raw values, by default False}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.ca.sentiment_chart from OpenBB Terminal SDK
}
\examples{
stocks.ca.sentiment_chart(raw=FALSE)
}
