% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.similar}
\alias{stocks.ca.similar}
\title{Find similar tickers to a given symbol.}
\usage{
stocks.ca.similar(symbol, source)
}
\arguments{
\item{symbol}{(character length 1) Symbol to find similar tickers to.}

\item{source}{(Python type: inspect._empty) Source to get data for, by default "Finviz".  Can be "Finviz", "Polygon", "Finnhub", or "TSNE"}
}
\description{
Wrapper for Python function stocks.ca.similar from OpenBB Terminal SDK
}
\examples{
stocks.ca.similar(symbol='AMZN', source='Finviz')
}
