% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.volume}
\alias{stocks.ca.volume}
\title{Get stock volume. \link{Source: Yahoo Finance}}
\usage{
stocks.ca.volume(similar, start_date, end_date)
}
\description{
Wrapper for Python function stocks.ca.volume from OpenBB Terminal SDK
}
