% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ca.volume_chart}
\alias{stocks.ca.volume_chart}
\title{Display stock volume. \link{Source: Yahoo Finance}}
\usage{
stocks.ca.volume_chart(
  similar,
  start_date,
  end_date,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{export}{(character length 1) Format to export historical prices, by default ""}
}
\description{
Wrapper for Python function stocks.ca.volume_chart from OpenBB Terminal SDK
}
