% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.candle}
\alias{stocks.candle}
\title{Show candle plot of loaded ticker.}
\usage{
stocks.candle(
  symbol,
  data,
  use_matplotlib = TRUE,
  intraday = FALSE,
  add_trend = FALSE,
  ma,
  asset_type = "",
  start_date,
  interval = 1440,
  end_date,
  prepost = FALSE,
  source = "YahooFinance",
  weekly = FALSE,
  monthly = FALSE,
  external_axes,
  raw = FALSE,
  yscale = "linear"
)
}
\arguments{
\item{symbol}{(character length 1) Ticker name}

\item{use_matplotlib}{(logical length 1) Flag to use matplotlib instead of interactive plotly chart}

\item{intraday}{(logical length 1) Flag for intraday data for plotly range breaks}

\item{add_trend}{(logical length 1) Flag to add high and low trends to chart}

\item{interval}{(integer length 1) Interval (in minutes) to get data 1, 5, 15, 30, 60 or 1440}

\item{prepost}{(logical length 1) Pre and After hours data}

\item{source}{(character length 1) Source of data extracted}

\item{weekly}{(logical length 1) Flag to get weekly data}

\item{monthly}{(logical length 1) Flag to get monthly data}

\item{raw}{(logical length 1) Flag to display raw data, by default False}

\item{yscale}{(character length 1) Linear or log for yscale}
}
\description{
Wrapper for Python function stocks.candle from OpenBB Terminal SDK
}
\examples{
stocks.candle(symbol='AMZN', use_matplotlib=TRUE, intraday=FALSE, add_trend=FALSE, interval=1440, prepost=FALSE, source='YahooFinance', weekly=FALSE, monthly=FALSE, raw=FALSE, yscale='linear')
}
