% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.active}
\alias{stocks.disc.active}
\title{Get stocks ordered in descending order by intraday trade volume. \link{Source: Yahoo Finance}}
\usage{
stocks.disc.active()
}
\description{
Wrapper for Python function stocks.disc.active from OpenBB Terminal SDK
}
