% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.arkord}
\alias{stocks.disc.arkord}
\title{Returns ARK orders in a Dataframe}
\usage{
stocks.disc.arkord(buys_only = FALSE, sells_only = FALSE, fund = "")
}
\arguments{
\item{buys_only}{(logical length 1) Flag to filter on buys only}

\item{sells_only}{(logical length 1) Flag to sort on sells only}

\item{fund}{(character length 1) Optional filter by fund}
}
\description{
Wrapper for Python function stocks.disc.arkord from OpenBB Terminal SDK
}
\examples{
stocks.disc.arkord(buys_only=FALSE, sells_only=FALSE)
}
