% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.asc}
\alias{stocks.disc.asc}
\title{Get Yahoo Finance small cap stocks with earnings growth rates better than 25\%.}
\usage{
stocks.disc.asc()
}
\description{
Wrapper for Python function stocks.disc.asc from OpenBB Terminal SDK
}
