% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.dividends}
\alias{stocks.disc.dividends}
\title{Gets dividend calendar for given date.  Date represents Ex-Dividend Date}
\usage{
stocks.disc.dividends(date)
}
\description{
Wrapper for Python function stocks.disc.dividends from OpenBB Terminal SDK
}
