% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.filings}
\alias{stocks.disc.filings}
\title{Get SEC Filings RSS feed, disseminated by FMP}
\usage{
stocks.disc.filings(pages = 1)
}
\arguments{
\item{pages}{(integer length 1) The range of most-rececnt pages to get entries from (1000 per page; maximum of 30 pages)}
}
\description{
Wrapper for Python function stocks.disc.filings from OpenBB Terminal SDK
}
