% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.filings_chart}
\alias{stocks.disc.filings_chart}
\title{Display recent forms submitted to the SEC}
\usage{
stocks.disc.filings_chart(
  pages = 1,
  limit = 5,
  today = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{pages}{(integer length 1) The range of most-rececnt pages to get entries from (1000 per page, max 30 pages)}

\item{limit}{(integer length 1) Limit the number of entries to display (default: 5)}

\item{today}{(logical length 1) Show all from today}

\item{export}{(character length 1) Export data as csv, json, or xlsx}
}
\description{
Wrapper for Python function stocks.disc.filings_chart from OpenBB Terminal SDK
}
