% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.gainers}
\alias{stocks.disc.gainers}
\title{Get top gainers. \link{Source: Yahoo Finance}}
\usage{
stocks.disc.gainers()
}
\description{
Wrapper for Python function stocks.disc.gainers from OpenBB Terminal SDK
}
