% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.gtech}
\alias{stocks.disc.gtech}
\title{Get technology stocks with revenue and earnings growth in excess of 25\%. \link{Source: Yahoo Finance}}
\usage{
stocks.disc.gtech()
}
\description{
Wrapper for Python function stocks.disc.gtech from OpenBB Terminal SDK
}
