% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.hotpenny}
\alias{stocks.disc.hotpenny}
\title{Returns today hot penny stocks}
\usage{
stocks.disc.hotpenny()
}
\description{
Wrapper for Python function stocks.disc.hotpenny from OpenBB Terminal SDK
}
