% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.losers}
\alias{stocks.disc.losers}
\title{Get top losers. \link{Source: Yahoo Finance}}
\usage{
stocks.disc.losers()
}
\description{
Wrapper for Python function stocks.disc.losers from OpenBB Terminal SDK
}
