% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.lowfloat}
\alias{stocks.disc.lowfloat}
\title{Returns low float DataFrame}
\usage{
stocks.disc.lowfloat()
}
\description{
Wrapper for Python function stocks.disc.lowfloat from OpenBB Terminal SDK
}
