% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.news}
\alias{stocks.disc.news}
\title{Gets news. \link{Source: SeekingAlpha}}
\usage{
stocks.disc.news(news_type = "Top-News", limit = 5)
}
\arguments{
\item{news_type}{(character length 1) From: Top-News, On-The-Move, Market-Pulse, Notable-Calls, Buybacks, Commodities, Crypto, Issuance, Global,
Guidance, IPOs, SPACs, Politics, M-A, Consumer, Energy, Financials, Healthcare, MLPs, REITs, Technology}

\item{limit}{(integer length 1) Number of news to display}
}
\description{
Wrapper for Python function stocks.disc.news from OpenBB Terminal SDK
}
\examples{
stocks.disc.news(news_type='Top-News', limit=5)
}
