% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.pipo}
\alias{stocks.disc.pipo}
\title{Past IPOs dates. \link{Source: Finnhub}}
\usage{
stocks.disc.pipo(num_days_behind = 5, start_date)
}
\arguments{
\item{num_days_behind}{(integer length 1) Number of days to look behind for IPOs dates}
}
\description{
Wrapper for Python function stocks.disc.pipo from OpenBB Terminal SDK
}
\examples{
stocks.disc.pipo(num_days_behind=5)
}
