% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.rtat}
\alias{stocks.disc.rtat}
\title{Gets the top 10 retail stocks per day}
\usage{
stocks.disc.rtat()
}
\description{
Wrapper for Python function stocks.disc.rtat from OpenBB Terminal SDK
}
