% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.trending}
\alias{stocks.disc.trending}
\title{Returns a list of trending articles}
\usage{
stocks.disc.trending(limit = 5)
}
\arguments{
\item{limit}{(integer length 1) Number of articles}
}
\description{
Wrapper for Python function stocks.disc.trending from OpenBB Terminal SDK
}
\examples{
stocks.disc.trending(limit=5)
}
