% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.ugs}
\alias{stocks.disc.ugs}
\title{Get stocks with earnings growth rates better than 25\% and relatively low PE and PEG ratios.}
\usage{
stocks.disc.ugs()
}
\description{
Wrapper for Python function stocks.disc.ugs from OpenBB Terminal SDK
}
