% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.ulc}
\alias{stocks.disc.ulc}
\title{Get Yahoo Finance potentially undervalued large cap stocks.}
\usage{
stocks.disc.ulc()
}
\description{
Wrapper for Python function stocks.disc.ulc from OpenBB Terminal SDK
}
