% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.disc.upcoming}
\alias{stocks.disc.upcoming}
\title{Returns a DataFrame with upcoming earnings}
\usage{
stocks.disc.upcoming(limit = 10)
}
\arguments{
\item{limit}{(integer length 1) Number of pages}
}
\description{
Wrapper for Python function stocks.disc.upcoming from OpenBB Terminal SDK
}
\examples{
stocks.disc.upcoming(limit=10)
}
