% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.ctb}
\alias{stocks.dps.ctb}
\title{Get stocks with highest cost to borrow \link{Source: Interactive Broker}}
\usage{
stocks.dps.ctb()
}
\description{
Wrapper for Python function stocks.dps.ctb from OpenBB Terminal SDK
}
