% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.dpotc}
\alias{stocks.dps.dpotc}
\title{Get all FINRA data associated with a ticker}
\usage{
stocks.dps.dpotc(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker to get data from}
}
\description{
Wrapper for Python function stocks.dps.dpotc from OpenBB Terminal SDK
}
\examples{
stocks.dps.dpotc(symbol='AMZN')
}
