% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.dpotc_chart}
\alias{stocks.dps.dpotc_chart}
\title{Display barchart of dark pool (ATS) and OTC (Non ATS) data. \link{Source: FINRA}}
\usage{
stocks.dps.dpotc_chart(symbol, export = "", sheet_name, external_axes)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.dps.dpotc_chart from OpenBB Terminal SDK
}
\examples{
stocks.dps.dpotc_chart(symbol='AMZN')
}
