% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.ftd}
\alias{stocks.dps.ftd}
\title{Display fails-to-deliver data for a given ticker. \link{Source: SEC}}
\usage{
stocks.dps.ftd(symbol, start_date, end_date, limit = 0)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{limit}{(integer length 1) Number of latest fails-to-deliver being printed}
}
\description{
Wrapper for Python function stocks.dps.ftd from OpenBB Terminal SDK
}
\examples{
stocks.dps.ftd(symbol='AMZN', limit=0)
}
