% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.ftd_chart}
\alias{stocks.dps.ftd_chart}
\title{Display fails-to-deliver data for a given ticker. \link{Source: SEC}}
\usage{
stocks.dps.ftd_chart(
  symbol,
  data,
  start_date,
  end_date,
  limit = 0,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{limit}{(integer length 1) Number of latest fails-to-deliver being printed}

\item{raw}{(logical length 1) Print raw data}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.dps.ftd_chart from OpenBB Terminal SDK
}
\examples{
stocks.dps.ftd_chart(symbol='AMZN', limit=0, raw=FALSE)
}
