% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.pos}
\alias{stocks.dps.pos}
\title{Get dark pool short positions. \link{Source: Stockgrid}}
\usage{
stocks.dps.pos(sortby = "dpp_dollar", ascend = FALSE)
}
\arguments{
\item{sortby}{(character length 1) Field for which to sort by, where 'sv': Short Vol. \link{1M},
'sv_pct': Short Vol. \%\%, 'nsv': Net Short Vol. \link{1M},
'nsv_dollar': Net Short Vol. ($100M), 'dpp': DP Position \link{1M},
'dpp_dollar': DP Position ($1B)}

\item{ascend}{(logical length 1) Data in ascending order}
}
\description{
Wrapper for Python function stocks.dps.pos from OpenBB Terminal SDK
}
\examples{
stocks.dps.pos(sortby='dpp_dollar', ascend=FALSE)
}
