% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.prom}
\alias{stocks.dps.prom}
\title{Get all FINRA ATS data, and parse most promising tickers based on linear regression}
\usage{
stocks.dps.prom(limit = 1000, tier_ats = "T1")
}
\arguments{
\item{limit}{(integer length 1) Number of tickers to filter from entire ATS data based on the sum of the total weekly shares quantity}

\item{tier_ats}{(character length 1) Tier to process data from: T1, T2 or OTCE}
}
\description{
Wrapper for Python function stocks.dps.prom from OpenBB Terminal SDK
}
\examples{
stocks.dps.prom(limit=1000)
}
