% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.prom_chart}
\alias{stocks.dps.prom_chart}
\title{Display dark pool (ATS) data of tickers with growing trades activity. \link{Source: FINRA}}
\usage{
stocks.dps.prom_chart(
  input_limit = 1000,
  limit = 10,
  tier = "T1",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{input_limit}{(integer length 1) Number of tickers to filter from entire ATS data based on
the sum of the total weekly shares quantity}

\item{limit}{(integer length 1) Number of tickers to display from most promising with
better linear regression slope}

\item{tier}{(character length 1) Tier to process data from: T1, T2 or OTCE}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.dps.prom_chart from OpenBB Terminal SDK
}
\examples{
stocks.dps.prom_chart(input_limit=1000, limit=10, tier='T1')
}
