% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.psi_q}
\alias{stocks.dps.psi_q}
\title{Plots the short interest of a stock. This corresponds to the}
\usage{
stocks.dps.psi_q(symbol, nyse = FALSE)
}
\arguments{
\item{symbol}{(character length 1) ticker to get short interest from}

\item{nyse}{(logical length 1) data from NYSE if true, otherwise NASDAQ}
}
\description{
Wrapper for Python function stocks.dps.psi_q from OpenBB Terminal SDK
}
\examples{
stocks.dps.psi_q(symbol='AMZN', nyse=FALSE)
}
