% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.psi_q_chart}
\alias{stocks.dps.psi_q_chart}
\title{Plot the short interest of a stock. This corresponds to the}
\usage{
stocks.dps.psi_q_chart(
  symbol,
  nyse = FALSE,
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) ticker to get short interest from}

\item{nyse}{(logical length 1) data from NYSE if true, otherwise NASDAQ}

\item{limit}{(integer length 1) Number of past days to show short interest}

\item{raw}{(logical length 1) Flag to print raw data instead}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.dps.psi_q_chart from OpenBB Terminal SDK
}
\examples{
stocks.dps.psi_q_chart(symbol='AMZN', nyse=FALSE, limit=10, raw=FALSE)
}
