% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.psi_sg}
\alias{stocks.dps.psi_sg}
\title{Get price vs short interest volume. \link{Source: Stockgrid}}
\usage{
stocks.dps.psi_sg(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock to get data from}
}
\description{
Wrapper for Python function stocks.dps.psi_sg from OpenBB Terminal SDK
}
\examples{
stocks.dps.psi_sg(symbol='AMZN')
}
