% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.shorted}
\alias{stocks.dps.shorted}
\title{Get most shorted stock screener \link{Source: Yahoo Finance}}
\usage{
stocks.dps.shorted()
}
\description{
Wrapper for Python function stocks.dps.shorted from OpenBB Terminal SDK
}
