% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.sidtc}
\alias{stocks.dps.sidtc}
\title{Get short interest and days to cover. \link{Source: Stockgrid}}
\usage{
stocks.dps.sidtc(sortby = "float")
}
\arguments{
\item{sortby}{(character length 1) Field for which to sort by, where 'float': Float Short \%\%,
'dtc': Days to Cover, 'si': Short Interest}
}
\description{
Wrapper for Python function stocks.dps.sidtc from OpenBB Terminal SDK
}
\examples{
stocks.dps.sidtc(sortby='float')
}
