% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.dps.spos_chart}
\alias{stocks.dps.spos_chart}
\title{Plot net short position. \link{Source: Stockgrid}}
\usage{
stocks.dps.spos_chart(
  symbol,
  limit = 84,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock to plot for}

\item{limit}{(integer length 1) Number of last open market days to show}

\item{raw}{(logical length 1) Flag to print raw data instead}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.dps.spos_chart from OpenBB Terminal SDK
}
\examples{
stocks.dps.spos_chart(symbol='AMZN', limit=84, raw=FALSE)
}
