% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.analysis}
\alias{stocks.fa.analysis}
\title{Save time reading SEC filings with the help of machine learning. \link{Source: https://eclect.us}}
\usage{
stocks.fa.analysis(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to see analysis of filings}
}
\description{
Wrapper for Python function stocks.fa.analysis from OpenBB Terminal SDK
}
\examples{
stocks.fa.analysis(symbol='AMZN')
}
