% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.balance}
\alias{stocks.fa.balance}
\title{Get balance sheet.}
\usage{
stocks.fa.balance(
  symbol,
  quarterly = FALSE,
  ratios = FALSE,
  source = "YahooFinance",
  limit = 10
)
}
\arguments{
\item{symbol}{(character length 1) Symbol to get balance sheet for}

\item{quarterly}{(logical length 1) Flag to get quarterly data}

\item{ratios}{(logical length 1) Flag to return data as a percent change.}

\item{source}{(character length 1) Data source for balance sheet, by default "YahooFinance"}

\item{limit}{(integer length 1) Number of statements to return (free tiers may be limited to 5 years)}
}
\description{
Wrapper for Python function stocks.fa.balance from OpenBB Terminal SDK
}
\examples{
stocks.fa.balance(symbol='AMZN', source='YahooFinance', quarterly=FALSE, ratios=FALSE, limit=10)
}
