% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.cal}
\alias{stocks.fa.cal}
\title{Get calendar earnings for ticker symbol}
\usage{
stocks.fa.cal(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.cal from OpenBB Terminal SDK
}
\examples{
stocks.fa.cal(symbol='AMZN')
}
