% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.customer}
\alias{stocks.fa.customer}
\title{Print customers from ticker provided}
\usage{
stocks.fa.customer(symbol, limit = 50)
}
\arguments{
\item{symbol}{(character length 1) Ticker to select customers from}

\item{limit}{(integer length 1) The maximum number of rows to show}
}
\description{
Wrapper for Python function stocks.fa.customer from OpenBB Terminal SDK
}
