% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.data}
\alias{stocks.fa.data}
\title{Get fundamental data from finviz}
\usage{
stocks.fa.data(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.data from OpenBB Terminal SDK
}
\examples{
stocks.fa.data(symbol='AMZN')
}
