% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.dcfc}
\alias{stocks.fa.dcfc}
\title{Get stocks dcf from FMP}
\usage{
stocks.fa.dcfc(symbol, limit = 5, quarterly = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number to get}

\item{quarterly}{(logical length 1) Flag to get quarterly data, by default False}
}
\description{
Wrapper for Python function stocks.fa.dcfc from OpenBB Terminal SDK
}
