% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.divs_chart}
\alias{stocks.fa.divs_chart}
\title{Display historical dividends}
\usage{
stocks.fa.divs_chart(
  symbol,
  limit = 12,
  plot = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number to show}

\item{plot}{(logical length 1) Plots historical data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.fa.divs_chart from OpenBB Terminal SDK
}
\examples{
stocks.fa.divs_chart(symbol='AMZN', limit=12, plot=TRUE)
}
