% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.dupont}
\alias{stocks.fa.dupont}
\title{Get dupont ratios}
\usage{
stocks.fa.dupont(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.dupont from OpenBB Terminal SDK
}
\examples{
stocks.fa.dupont(symbol='AMZN')
}
