% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.earnings}
\alias{stocks.fa.earnings}
\title{Get earnings data.}
\usage{
stocks.fa.earnings(symbol, source = "YahooFinance", quarterly = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{source}{(character length 1) Source to use, by default "AlphaVantage"}

\item{quarterly}{(logical length 1) Flag to get quarterly data (AlphaVantage only), by default False.}
}
\description{
Wrapper for Python function stocks.fa.earnings from OpenBB Terminal SDK
}
\examples{
stocks.fa.earnings(symbol='AMZN', source='YahooFinance', quarterly=FALSE)
}
