% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.enterprise}
\alias{stocks.fa.enterprise}
\title{Financial Modeling Prep ticker enterprise}
\usage{
stocks.fa.enterprise(symbol, limit = 5, quarterly = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Fundamental analysis ticker symbol}

\item{limit}{(integer length 1) Number to get}

\item{quarterly}{(logical length 1) Flag to get quarterly data}
}
\description{
Wrapper for Python function stocks.fa.enterprise from OpenBB Terminal SDK
}
\examples{
stocks.fa.enterprise(symbol='AMZN', limit=5, quarterly=FALSE)
}
