% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.epsfc}
\alias{stocks.fa.epsfc}
\title{Takes the ticker, asks for seekingalphaID and gets eps estimates}
\usage{
stocks.fa.epsfc(ticker)
}
\arguments{
\item{ticker}{(character length 1) ticker of company}
}
\description{
Wrapper for Python function stocks.fa.epsfc from OpenBB Terminal SDK
}
