% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.est}
\alias{stocks.fa.est}
\title{Get analysts' estimates for a given ticker. \link{Source: Business Insider}}
\usage{
stocks.fa.est(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get analysts' estimates}
}
\description{
Wrapper for Python function stocks.fa.est from OpenBB Terminal SDK
}
