% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.fama_coe}
\alias{stocks.fa.fama_coe}
\title{Use Fama and French to get the cost of equity for a company}
\usage{
stocks.fa.fama_coe(symbol)
}
\arguments{
\item{symbol}{(character length 1) The ticker symbol to be analyzed}
}
\description{
Wrapper for Python function stocks.fa.fama_coe from OpenBB Terminal SDK
}
\examples{
stocks.fa.fama_coe(symbol='AMZN')
}
