% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.fraud}
\alias{stocks.fa.fraud}
\title{Get fraud ratios based on fundamentals}
\usage{
stocks.fa.fraud(symbol, detail = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{detail}{(logical length 1) Whether to provide extra m-score details}
}
\description{
Wrapper for Python function stocks.fa.fraud from OpenBB Terminal SDK
}
\examples{
stocks.fa.fraud(symbol='AMZN', detail=FALSE)
}
