% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.growth}
\alias{stocks.fa.growth}
\title{Get financial statement growth}
\usage{
stocks.fa.growth(symbol, limit = 5, quarterly = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number to get}

\item{quarterly}{(logical length 1) Flag to get quarterly data, by default False}
}
\description{
Wrapper for Python function stocks.fa.growth from OpenBB Terminal SDK
}
\examples{
stocks.fa.growth(symbol='AMZN', limit=5, quarterly=FALSE)
}
