% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.historical_5}
\alias{stocks.fa.historical_5}
\title{Get 5 year monthly historical performance for a ticker with dividends filtered}
\usage{
stocks.fa.historical_5(symbol)
}
\arguments{
\item{symbol}{(character length 1) The ticker symbol to be analyzed}
}
\description{
Wrapper for Python function stocks.fa.historical_5 from OpenBB Terminal SDK
}
\examples{
stocks.fa.historical_5(symbol='AMZN')
}
