% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.info}
\alias{stocks.fa.info}
\title{Gets ticker symbol info}
\usage{
stocks.fa.info(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.info from OpenBB Terminal SDK
}
\examples{
stocks.fa.info(symbol='AMZN')
}
