% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.key}
\alias{stocks.fa.key}
\title{Get key metrics from overview}
\usage{
stocks.fa.key(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.key from OpenBB Terminal SDK
}
\examples{
stocks.fa.key(symbol='AMZN')
}
