% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.mgmt}
\alias{stocks.fa.mgmt}
\title{Get company managers from Business Insider}
\usage{
stocks.fa.mgmt(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.mgmt from OpenBB Terminal SDK
}
\examples{
stocks.fa.mgmt(symbol='AMZN')
}
