% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.mktcap}
\alias{stocks.fa.mktcap}
\title{Get market cap over time for ticker. \link{Source: Yahoo Finance}}
\usage{
stocks.fa.mktcap(symbol, start_date)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get market cap over time}
}
\description{
Wrapper for Python function stocks.fa.mktcap from OpenBB Terminal SDK
}
\examples{
stocks.fa.mktcap(symbol='AMZN')
}
