% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.mktcap_chart}
\alias{stocks.fa.mktcap_chart}
\title{Display market cap over time. \link{Source: Yahoo Finance}}
\usage{
stocks.fa.mktcap_chart(
  symbol,
  start_date,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.fa.mktcap_chart from OpenBB Terminal SDK
}
\examples{
stocks.fa.mktcap_chart(symbol='AMZN')
}
