% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.news}
\alias{stocks.fa.news}
\title{Get news from Finviz}
\usage{
stocks.fa.news(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.news from OpenBB Terminal SDK
}
