% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.profile}
\alias{stocks.fa.profile}
\title{Get ticker profile from FMP}
\usage{
stocks.fa.profile(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.profile from OpenBB Terminal SDK
}
\examples{
stocks.fa.profile(symbol='AMZN')
}
