% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.pt}
\alias{stocks.fa.pt}
\title{Get analysts' price targets for a given stock. \link{Source: Business Insider}}
\usage{
stocks.fa.pt(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.pt from OpenBB Terminal SDK
}
