% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.pt_chart}
\alias{stocks.fa.pt_chart}
\title{Display analysts' price targets for a given stock. \link{Source: Business Insider}}
\usage{
stocks.fa.pt_chart(
  symbol,
  data,
  start_date,
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Due diligence ticker symbol}

\item{limit}{(integer length 1) Number of latest price targets from analysts to print}

\item{raw}{(logical length 1) Display raw data only}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.fa.pt_chart from OpenBB Terminal SDK
}
