% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.rating}
\alias{stocks.fa.rating}
\title{Get ratings for a given ticker. \link{Source: Financial Modeling Prep}}
\usage{
stocks.fa.rating(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.rating from OpenBB Terminal SDK
}
