% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.revfc}
\alias{stocks.fa.revfc}
\title{Takes the ticker, asks for seekingalphaID and gets rev estimates}
\usage{
stocks.fa.revfc(ticker)
}
\arguments{
\item{ticker}{(character length 1) ticker of company}
}
\description{
Wrapper for Python function stocks.fa.revfc from OpenBB Terminal SDK
}
