% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.rot}
\alias{stocks.fa.rot}
\title{Get rating over time data. \link{Source: Finnhub}}
\usage{
stocks.fa.rot(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get ratings from}
}
\description{
Wrapper for Python function stocks.fa.rot from OpenBB Terminal SDK
}
