% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.rot_chart}
\alias{stocks.fa.rot_chart}
\title{Rating over time (monthly). \link{Source: Finnhub}}
\usage{
stocks.fa.rot_chart(
  symbol,
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{limit}{(integer length 1) Number of last months ratings to show}

\item{raw}{(logical length 1) Display raw data only}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.fa.rot_chart from OpenBB Terminal SDK
}
