% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.score}
\alias{stocks.fa.score}
\title{Gets value score from fmp}
\usage{
stocks.fa.score(symbol, years)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{years}{(integer length 1) The amount of years to use to calculate the score}
}
\description{
Wrapper for Python function stocks.fa.score from OpenBB Terminal SDK
}
\examples{
stocks.fa.score(symbol='AMZN')
}
