% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.sec}
\alias{stocks.fa.sec}
\title{Get SEC filings for a given stock ticker. \link{Source: Market Watch}}
\usage{
stocks.fa.sec(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.fa.sec from OpenBB Terminal SDK
}
