% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.sec_fmp_chart}
\alias{stocks.fa.sec_fmp_chart}
\title{Display recent forms submitted to the SEC}
\usage{
stocks.fa.sec_fmp_chart(
  ticker = "",
  pages = 1,
  limit = 20,
  export = "",
  sheet_name
)
}
\arguments{
\item{pages}{(integer length 1) The range of most-rececnt pages to get entries from (1000 per page, max 30 pages)}

\item{limit}{(integer length 1) Limit the number of entries to display (default: 20)}

\item{export}{(character length 1) Export data as csv, json, or xlsx}
}
\description{
Wrapper for Python function stocks.fa.sec_fmp_chart from OpenBB Terminal SDK
}
