% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.shrs}
\alias{stocks.fa.shrs}
\title{Get shareholders from yahoo}
\usage{
stocks.fa.shrs(symbol, holder = "institutional")
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{holder}{(character length 1) Which holder to get table for}
}
\description{
Wrapper for Python function stocks.fa.shrs from OpenBB Terminal SDK
}
\examples{
stocks.fa.shrs(symbol='AMZN', holder='institutional')
}
