% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.similar_dfs}
\alias{stocks.fa.similar_dfs}
\title{Get dataframes for similar companies}
\usage{
stocks.fa.similar_dfs(symbol, info, n, no_filter = FALSE)
}
\arguments{
\item{symbol}{(character length 1) The ticker symbol to create a dataframe for}

\item{n}{(integer length 1) The number of similar companies to produce}

\item{no_filter}{(logical length 1) True means that we do not filter based on market cap}
}
\description{
Wrapper for Python function stocks.fa.similar_dfs from OpenBB Terminal SDK
}
\examples{
stocks.fa.similar_dfs(symbol='AMZN', no_filter=FALSE)
}
