% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.splits}
\alias{stocks.fa.splits}
\title{Get splits and reverse splits events. \link{Source: Yahoo Finance}}
\usage{
stocks.fa.splits(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get forward and reverse splits}
}
\description{
Wrapper for Python function stocks.fa.splits from OpenBB Terminal SDK
}
\examples{
stocks.fa.splits(symbol='AMZN')
}
