% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.splits_chart}
\alias{stocks.fa.splits_chart}
\title{Display splits and reverse splits events. \link{Source: Yahoo Finance}}
\usage{
stocks.fa.splits_chart(symbol, export = "", sheet_name, external_axes)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.fa.splits_chart from OpenBB Terminal SDK
}
\examples{
stocks.fa.splits_chart(symbol='AMZN')
}
