% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.fa.supplier}
\alias{stocks.fa.supplier}
\title{Get suppliers from ticker provided. \link{Source: CSIMarket}}
\usage{
stocks.fa.supplier(symbol, limit = 50)
}
\arguments{
\item{symbol}{(character length 1) Ticker to select suppliers from}

\item{limit}{(integer length 1) The maximum number of rows to show}
}
\description{
Wrapper for Python function stocks.fa.supplier from OpenBB Terminal SDK
}
