% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.contracts}
\alias{stocks.gov.contracts}
\title{Get government contracts for ticker \link{Source: quiverquant.com}}
\usage{
stocks.gov.contracts(symbol, past_transaction_days = 10)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get congress trading data from}

\item{past_transaction_days}{(integer length 1) Number of days to get transactions for}
}
\description{
Wrapper for Python function stocks.gov.contracts from OpenBB Terminal SDK
}
\examples{
stocks.gov.contracts(symbol='AMZN', past_transaction_days=10)
}
