% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.contracts_chart}
\alias{stocks.gov.contracts_chart}
\title{Show government contracts for ticker \link{Source: quiverquant.com}}
\usage{
stocks.gov.contracts_chart(
  symbol,
  past_transaction_days = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get congress trading data from}

\item{past_transaction_days}{(integer length 1) Number of days to get transactions for}

\item{raw}{(logical length 1) Flag to display raw data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.contracts_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.contracts_chart(symbol='AMZN', past_transaction_days=10, raw=FALSE)
}
