% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.government_trading}
\alias{stocks.gov.government_trading}
\title{Returns the most recent transactions by members of government}
\usage{
stocks.gov.government_trading(gov_type = "congress", symbol = "")
}
\arguments{
\item{gov_type}{(character length 1) Type of government data between:
'congress', 'senate', 'house', 'contracts', 'quarter-contracts' and 'corporate-lobbying'}

\item{symbol}{(character length 1) Ticker symbol to get congress trading data from}
}
\description{
Wrapper for Python function stocks.gov.government_trading from OpenBB Terminal SDK
}
\examples{
stocks.gov.government_trading(gov_type='congress')
}
