% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.gtrades}
\alias{stocks.gov.gtrades}
\title{Government trading for specific ticker \link{Source: quiverquant.com}}
\usage{
stocks.gov.gtrades(symbol, gov_type = "congress", past_transactions_months = 6)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get congress trading data from}

\item{gov_type}{(character length 1) Type of government data between: congress, senate and house}

\item{past_transactions_months}{(integer length 1) Number of months to get transactions for}
}
\description{
Wrapper for Python function stocks.gov.gtrades from OpenBB Terminal SDK
}
\examples{
stocks.gov.gtrades(symbol='AMZN', gov_type='congress', past_transactions_months=6)
}
