% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.gtrades_chart}
\alias{stocks.gov.gtrades_chart}
\title{Government trading for specific ticker \link{Source: quiverquant.com}}
\usage{
stocks.gov.gtrades_chart(
  symbol,
  gov_type = "congress",
  past_transactions_months = 6,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get congress trading data from}

\item{gov_type}{(character length 1) Type of government data between: congress, senate and house}

\item{past_transactions_months}{(integer length 1) Number of months to get transactions for}

\item{raw}{(logical length 1) Show raw output of trades}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.gtrades_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.gtrades_chart(symbol='AMZN', gov_type='congress', past_transactions_months=6, raw=FALSE)
}
