% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.histcont}
\alias{stocks.gov.histcont}
\title{Get historical quarterly government contracts \link{Source: quiverquant.com}}
\usage{
stocks.gov.histcont(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get congress trading data from}
}
\description{
Wrapper for Python function stocks.gov.histcont from OpenBB Terminal SDK
}
\examples{
stocks.gov.histcont(symbol='AMZN')
}
