% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.histcont_chart}
\alias{stocks.gov.histcont_chart}
\title{Show historical quarterly government contracts \link{Source: quiverquant.com}}
\usage{
stocks.gov.histcont_chart(
  symbol,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get congress trading data from}

\item{raw}{(logical length 1) Flag to display raw data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.histcont_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.histcont_chart(symbol='AMZN', raw=FALSE)
}
