% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.lastcontracts}
\alias{stocks.gov.lastcontracts}
\title{Get last government contracts \link{Source: quiverquant.com}}
\usage{
stocks.gov.lastcontracts(past_transaction_days = 2)
}
\arguments{
\item{past_transaction_days}{(integer length 1) Number of days to look back}
}
\description{
Wrapper for Python function stocks.gov.lastcontracts from OpenBB Terminal SDK
}
\examples{
stocks.gov.lastcontracts(past_transaction_days=2)
}
