% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.lastcontracts_chart}
\alias{stocks.gov.lastcontracts_chart}
\title{Last government contracts \link{Source: quiverquant.com}}
\usage{
stocks.gov.lastcontracts_chart(
  past_transaction_days = 2,
  limit = 20,
  sum_contracts = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{past_transaction_days}{(integer length 1) Number of days to look back}

\item{limit}{(integer length 1) Number of contracts to show}

\item{sum_contracts}{(logical length 1) Flag to show total amount of contracts given out.}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.lastcontracts_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.lastcontracts_chart(past_transaction_days=2, limit=20, sum_contracts=FALSE)
}
