% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.lasttrades}
\alias{stocks.gov.lasttrades}
\title{Get last government trading \link{Source: quiverquant.com}}
\usage{
stocks.gov.lasttrades(gov_type = "congress", limit = -1, representative = "")
}
\arguments{
\item{gov_type}{(character length 1) Type of government data between: congress, senate and house}

\item{limit}{(integer length 1) Number of days to look back}

\item{representative}{(character length 1) Specific representative to look at}
}
\description{
Wrapper for Python function stocks.gov.lasttrades from OpenBB Terminal SDK
}
\examples{
stocks.gov.lasttrades(gov_type='congress', limit=-1)
}
