% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.lobbying}
\alias{stocks.gov.lobbying}
\title{Corporate lobbying details}
\usage{
stocks.gov.lobbying(symbol, limit = 10)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get corporate lobbying data from}

\item{limit}{(integer length 1) Number of events to show}
}
\description{
Wrapper for Python function stocks.gov.lobbying from OpenBB Terminal SDK
}
\examples{
stocks.gov.lobbying(symbol='AMZN', limit=10)
}
