% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.qtrcontracts}
\alias{stocks.gov.qtrcontracts}
\title{Analyzes quarterly contracts by ticker}
\usage{
stocks.gov.qtrcontracts(analysis = "total", limit = 5)
}
\arguments{
\item{analysis}{(character length 1) How to analyze.  Either gives total amount or sorts by high/low momentum.}

\item{limit}{(integer length 1) Number to return, by default 5}
}
\description{
Wrapper for Python function stocks.gov.qtrcontracts from OpenBB Terminal SDK
}
\examples{
stocks.gov.qtrcontracts(analysis='total', limit=5)
}
