% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.topbuys_chart}
\alias{stocks.gov.topbuys_chart}
\title{Top buy government trading \link{Source: quiverquant.com}}
\usage{
stocks.gov.topbuys_chart(
  gov_type = "congress",
  past_transactions_months = 6,
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{gov_type}{(character length 1) Type of government data between: congress, senate and house}

\item{past_transactions_months}{(integer length 1) Number of months to get trading for}

\item{limit}{(integer length 1) Number of tickers to show}

\item{raw}{(logical length 1) Display raw data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.topbuys_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.topbuys_chart(gov_type='congress', past_transactions_months=6, limit=10, raw=FALSE)
}
