% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.toplobbying}
\alias{stocks.gov.toplobbying}
\title{Corporate lobbying details}
\usage{
stocks.gov.toplobbying()
}
\description{
Wrapper for Python function stocks.gov.toplobbying from OpenBB Terminal SDK
}
