% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.toplobbying_chart}
\alias{stocks.gov.toplobbying_chart}
\title{Top lobbying tickers based on total spent}
\usage{
stocks.gov.toplobbying_chart(
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{limit}{(integer length 1) Number of tickers to show}

\item{raw}{(logical length 1) Show raw data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.gov.toplobbying_chart from OpenBB Terminal SDK
}
\examples{
stocks.gov.toplobbying_chart(limit=10, raw=FALSE)
}
