% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.gov.topsells}
\alias{stocks.gov.topsells}
\title{Get top sell government trading \link{Source: quiverquant.com}}
\usage{
stocks.gov.topsells(gov_type = "congress", past_transactions_months = 6)
}
\arguments{
\item{gov_type}{(character length 1) Type of government data between: congress, senate and house}

\item{past_transactions_months}{(integer length 1) Number of months to get trading for}
}
\description{
Wrapper for Python function stocks.gov.topsells from OpenBB Terminal SDK
}
\examples{
stocks.gov.topsells(gov_type='congress', past_transactions_months=6)
}
