% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.act}
\alias{stocks.ins.act}
\title{Get insider activity. \link{Source: Business Insider}}
\usage{
stocks.ins.act(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get insider activity data from}
}
\description{
Wrapper for Python function stocks.ins.act from OpenBB Terminal SDK
}
\examples{
stocks.ins.act(symbol='AMZN')
}
