% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.act_chart}
\alias{stocks.ins.act_chart}
\title{Display insider activity. \link{Source: Business Insider}}
\usage{
stocks.ins.act_chart(
  data,
  symbol,
  start_date,
  interval = "1440min",
  limit = 10,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Stock dataframe}

\item{symbol}{(character length 1) Due diligence ticker symbol}

\item{interval}{(character length 1) Stock data interval}

\item{limit}{(integer length 1) Number of latest days of inside activity}

\item{raw}{(logical length 1) Print to console}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.ins.act_chart from OpenBB Terminal SDK
}
\examples{
stocks.ins.act_chart(symbol='AMZN', interval='1440min', limit=10, raw=FALSE)
}
