% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.filter}
\alias{stocks.ins.filter}
\title{GEt insider trades based on preset filter}
\usage{
stocks.ins.filter(preset)
}
\arguments{
\item{preset}{(character length 1) Name of preset filter}
}
\description{
Wrapper for Python function stocks.ins.filter from OpenBB Terminal SDK
}
