% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.lins}
\alias{stocks.ins.lins}
\title{Get last insider activity for a given stock ticker. \link{Source: Finviz}}
\usage{
stocks.ins.lins(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.ins.lins from OpenBB Terminal SDK
}
\examples{
stocks.ins.lins(symbol='AMZN')
}
