% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.lins_chart}
\alias{stocks.ins.lins_chart}
\title{Display insider activity for a given stock ticker. \link{Source: Finviz}}
\usage{
stocks.ins.lins_chart(symbol, limit = 10, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{limit}{(integer length 1) Number of latest insider activity to display}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.ins.lins_chart from OpenBB Terminal SDK
}
\examples{
stocks.ins.lins_chart(symbol='AMZN', limit=10)
}
