% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.print_insider_data}
\alias{stocks.ins.print_insider_data}
\title{Print insider data}
\usage{
stocks.ins.print_insider_data(type_insider = "lcb")
}
\arguments{
\item{type_insider}{(character length 1) Insider type of data. Available types can be accessed through get_insider_types().}
}
\description{
Wrapper for Python function stocks.ins.print_insider_data from OpenBB Terminal SDK
}
\examples{
stocks.ins.print_insider_data(type_insider='lcb', limit=10)
}
