% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.print_insider_data_chart}
\alias{stocks.ins.print_insider_data_chart}
\title{Print insider data}
\usage{
stocks.ins.print_insider_data_chart(
  type_insider = "lcb",
  limit = 10,
  export = "",
  sheet_name
)
}
\arguments{
\item{type_insider}{(character length 1) Insider type of data. Available types can be accessed through get_insider_types().}

\item{limit}{(integer length 1) Limit of data rows to display}

\item{export}{(character length 1) Export data format}
}
\description{
Wrapper for Python function stocks.ins.print_insider_data_chart from OpenBB Terminal SDK
}
\examples{
stocks.ins.print_insider_data_chart(type_insider='lcb', limit=10)
}
