% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ins.stats}
\alias{stocks.ins.stats}
\title{Get OpenInsider stats for ticker}
\usage{
stocks.ins.stats(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get insider stats for}
}
\description{
Wrapper for Python function stocks.ins.stats from OpenBB Terminal SDK
}
