% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.chains}
\alias{stocks.options.chains}
\title{Get Option Chain For A Stock.  No greek data is returned}
\usage{
stocks.options.chains(symbol, source = "Nasdaq", expiration)
}
\arguments{
\item{symbol}{(character length 1) Symbol to get chain for}

\item{source}{(character length 1) Source to get data from, by default "Nasdaq".  Can be YahooFinance, Tradier, Nasdaq, or Intrinio}
}
\description{
Wrapper for Python function stocks.options.chains from OpenBB Terminal SDK
}
\examples{
stocks.options.chains(symbol='AMZN', source='Nasdaq')
}
