% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.dte}
\alias{stocks.options.dte}
\title{Gets days to expiration from yfinance option date}
\usage{
stocks.options.dte(date_value)
}
\description{
Wrapper for Python function stocks.options.dte from OpenBB Terminal SDK
}
