% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.eodchain}
\alias{stocks.options.eodchain}
\title{Get full EOD option date across all expirations}
\usage{
stocks.options.eodchain(symbol, date, quiet = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Symbol to get option chain for}

\item{date}{(character length 1) Date to get EOD chain for}

\item{quiet}{(logical length 1) Flag to suppress progress bar}
}
\description{
Wrapper for Python function stocks.options.eodchain from OpenBB Terminal SDK
}
