% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.expirations}
\alias{stocks.options.expirations}
\title{Get Option Chain Expirations}
\usage{
stocks.options.expirations(symbol, source = "Nasdaq")
}
\arguments{
\item{symbol}{(character length 1) Symbol to get chain for}

\item{source}{(character length 1) Source to get data from, by default "Nasdaq"}
}
\description{
Wrapper for Python function stocks.options.expirations from OpenBB Terminal SDK
}
\examples{
stocks.options.expirations(symbol='AMZN', source='Nasdaq')
}
