% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.generate_data}
\alias{stocks.options.generate_data}
\title{Gets x values, and y values before and after premiums}
\usage{
stocks.options.generate_data(current_price, options, underlying)
}
\description{
Wrapper for Python function stocks.options.generate_data from OpenBB Terminal SDK
}
