% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.greeks}
\alias{stocks.options.greeks}
\title{Gets the greeks for a given option}
\usage{
stocks.options.greeks(current_price, chain, expire, div_cont = 0, rf)
}
\arguments{
\item{current_price}{(numeric length 1) The current price of the underlying}

\item{chain}{(data.frame) The dataframe with option chains}

\item{expire}{(character length 1) The date of expiration}

\item{div_cont}{(numeric length 1) The dividend continuous rate}
}
\description{
Wrapper for Python function stocks.options.greeks from OpenBB Terminal SDK
}
