% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.grhist}
\alias{stocks.options.grhist}
\title{Get histoical option greeks}
\usage{
stocks.options.grhist(symbol, expiry, strike, chain_id = "", put = FALSE)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{expiry}{(character length 1) Option expiration date}

\item{strike}{(numeric length 1) Strike price to look for}

\item{chain_id}{(character length 1) OCC option symbol.  Overwrites other inputs}

\item{put}{(logical length 1) Is this a put option?}
}
\description{
Wrapper for Python function stocks.options.grhist from OpenBB Terminal SDK
}
\examples{
stocks.options.grhist(symbol='AMZN', put=FALSE)
}
