% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.grhist_chart}
\alias{stocks.options.grhist_chart}
\title{Plots historical greeks for a given option. \link{Source: Syncretism}}
\usage{
stocks.options.grhist_chart(
  symbol,
  expiry,
  strike,
  greek = "Delta",
  chain_id = "",
  put = FALSE,
  raw = FALSE,
  limit,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker}

\item{expiry}{(character length 1) Expiration date}

\item{greek}{(character length 1) Greek variable to plot}

\item{chain_id}{(character length 1) OCC option chain.  Overwrites other variables}

\item{put}{(logical length 1) Is this a put option?}

\item{raw}{(logical length 1) Print to console}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function stocks.options.grhist_chart from OpenBB Terminal SDK
}
\examples{
stocks.options.grhist_chart(symbol='AMZN', greek='Delta', put=FALSE, raw=FALSE, limit=20)
}
