% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.hist}
\alias{stocks.options.hist}
\title{Get historical option pricing.}
\usage{
stocks.options.hist(symbol, exp, strike, call = TRUE, source)
}
\arguments{
\item{symbol}{(character length 1) Symbol to get data for}

\item{exp}{(character length 1) Expiration date}

\item{call}{(logical length 1) Flag to indicate a call, by default True}

\item{source}{(Python type: inspect._empty) Source to get data from.  Can be ChartExchange or Tradier, by default "ChartExchange"}
}
\description{
Wrapper for Python function stocks.options.hist from OpenBB Terminal SDK
}
\examples{
stocks.options.hist(symbol='AMZN', call=TRUE, source='ChartExchange')
}
