% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.info_chart}
\alias{stocks.options.info_chart}
\title{Scrapes Barchart.com for the options information}
\usage{
stocks.options.info_chart(symbol, export = "", sheet_name)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get options info for}

\item{export}{(character length 1) Format of export file}
}
\description{
Wrapper for Python function stocks.options.info_chart from OpenBB Terminal SDK
}
\examples{
stocks.options.info_chart(symbol='AMZN')
}
