% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.last_price}
\alias{stocks.options.last_price}
\title{Makes api request for last price}
\usage{
stocks.options.last_price(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol}
}
\description{
Wrapper for Python function stocks.options.last_price from OpenBB Terminal SDK
}
\examples{
stocks.options.last_price(symbol='AMZN')
}
