% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.pcr_chart}
\alias{stocks.options.pcr_chart}
\title{Display put call ratio \link{Source: AlphaQuery.com}}
\usage{
stocks.options.pcr_chart(
  symbol,
  window = 30,
  start_date = "2022-04-03",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}

\item{window}{(integer length 1) Window length to look at, by default 30}

\item{start_date}{(character length 1) Starting date for data, by default (datetime.now() - timedelta(days=366)).strftime("\%Y-\%m-\%d")}

\item{export}{(character length 1) Format to export data, by default ""}
}
\description{
Wrapper for Python function stocks.options.pcr_chart from OpenBB Terminal SDK
}
\examples{
stocks.options.pcr_chart(symbol='AMZN', window=30, start_date='2021-11-24')
}
