% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.price}
\alias{stocks.options.price}
\title{Get Option current price for a stock.}
\usage{
stocks.options.price(symbol, source = "Nasdaq")
}
\arguments{
\item{symbol}{(character length 1) Symbol to get chain for}

\item{source}{(character length 1) Source to get data from, by default "Nasdaq"}
}
\description{
Wrapper for Python function stocks.options.price from OpenBB Terminal SDK
}
