% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.unu}
\alias{stocks.options.unu}
\title{Get unusual option activity from fdscanner.com}
\usage{
stocks.options.unu(limit = 100)
}
\arguments{
\item{limit}{(integer length 1) Number to show}
}
\description{
Wrapper for Python function stocks.options.unu from OpenBB Terminal SDK
}
\examples{
stocks.options.unu(limit=100)
}
