% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.unu_chart}
\alias{stocks.options.unu_chart}
\title{Displays the unusual options table}
\usage{
stocks.options.unu_chart(
  limit = 20,
  sortby = "Vol/OI",
  ascend = FALSE,
  calls_only = FALSE,
  puts_only = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{limit}{(integer length 1) Number of rows to show}

\item{sortby}{(character length 1) Data column to sort on}

\item{ascend}{(logical length 1) Whether to sort in ascend order}

\item{calls_only}{(logical length 1) Flag to only show calls}

\item{puts_only}{(logical length 1) Flag to show puts only}

\item{export}{(character length 1) File type to export}
}
\description{
Wrapper for Python function stocks.options.unu_chart from OpenBB Terminal SDK
}
\examples{
stocks.options.unu_chart(limit=20, sortby='Vol/OI', ascend=FALSE, calls_only=FALSE, puts_only=FALSE)
}
