% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.vol}
\alias{stocks.options.vol}
\title{Plot volume}
\usage{
stocks.options.vol(
  chain,
  current_price,
  symbol,
  expiry,
  min_sp = -1,
  max_sp = -1,
  calls_only = FALSE,
  puts_only = FALSE,
  raw = FALSE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{chain}{(data.frame) Dataframe with options chain}

\item{current_price}{(numeric length 1) Current price of selected symbol}

\item{symbol}{(character length 1) Ticker symbol}

\item{expiry}{(character length 1) expiration date for options}

\item{min_sp}{(numeric length 1) Min strike to consider}

\item{max_sp}{(numeric length 1) Max strike to consider}

\item{calls_only}{(logical length 1) Show calls only}

\item{puts_only}{(logical length 1) Show puts only}

\item{export}{(character length 1) Format to export file}
}
\description{
Wrapper for Python function stocks.options.vol from OpenBB Terminal SDK
}
