% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.options.vsurf}
\alias{stocks.options.vsurf}
\title{Gets IV surface for calls and puts for ticker}
\usage{
stocks.options.vsurf(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol to get}
}
\description{
Wrapper for Python function stocks.options.vsurf from OpenBB Terminal SDK
}
\examples{
stocks.options.vsurf(symbol='AMZN')
}
