% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.process_candle}
\alias{stocks.process_candle}
\title{Process DataFrame into candle style plot.}
\usage{
stocks.process_candle(data)
}
\arguments{
\item{data}{(data.frame) Stock dataframe.}
}
\description{
Wrapper for Python function stocks.process_candle from OpenBB Terminal SDK
}
