% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.qa.beta}
\alias{stocks.qa.beta}
\title{Calculate beta for a ticker and a reference ticker.}
\usage{
stocks.qa.beta(symbol, ref_symbol, data, ref_data, interval = 1440)
}
\arguments{
\item{symbol}{(character length 1) A ticker to calculate beta for}

\item{ref_symbol}{(character length 1) A reference ticker symbol for the beta calculation (default in terminal is SPY)}

\item{interval}{(integer length 1) The interval of the ref_data. This will ONLY be used if ref_data is None}
}
\description{
Wrapper for Python function stocks.qa.beta from OpenBB Terminal SDK
}
\examples{
stocks.qa.beta(symbol='AMZN', interval=1440)
}
