% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.qa.beta_chart}
\alias{stocks.qa.beta_chart}
\title{Display the beta scatterplot + linear regression.}
\usage{
stocks.qa.beta_chart(
  symbol,
  ref_symbol,
  data,
  ref_data,
  interval = 1440,
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) A ticker to calculate beta for}

\item{ref_symbol}{(character length 1) A reference ticker symbol for the beta calculation (default in terminal is SPY)}

\item{interval}{(integer length 1) The interval of the ref_data. This will ONLY be used if ref_data is None}
}
\description{
Wrapper for Python function stocks.qa.beta_chart from OpenBB Terminal SDK
}
\examples{
stocks.qa.beta_chart(symbol='AMZN', interval=1440)
}
