% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.qa.capm}
\alias{stocks.qa.capm}
\title{Provides information that relates to the CAPM model}
\usage{
stocks.qa.capm(symbol)
}
\arguments{
\item{symbol}{(character length 1) A ticker symbol in string form}
}
\description{
Wrapper for Python function stocks.qa.capm from OpenBB Terminal SDK
}
\examples{
stocks.qa.capm(symbol='AMZN')
}
