% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.qa.fama_raw}
\alias{stocks.qa.fama_raw}
\title{Gets base Fama French data to calculate risk}
\usage{
stocks.qa.fama_raw()
}
\description{
Wrapper for Python function stocks.qa.fama_raw from OpenBB Terminal SDK
}
