% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.qa.historical_5}
\alias{stocks.qa.historical_5}
\title{Get 5 year monthly historical performance for a ticker with dividends filtered}
\usage{
stocks.qa.historical_5(symbol)
}
\arguments{
\item{symbol}{(character length 1) A ticker symbol in string form}
}
\description{
Wrapper for Python function stocks.qa.historical_5 from OpenBB Terminal SDK
}
\examples{
stocks.qa.historical_5(symbol='AMZN')
}
