% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.quote}
\alias{stocks.quote}
\title{Gets ticker quote from FMP}
\usage{
stocks.quote(symbol)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.quote from OpenBB Terminal SDK
}
\examples{
stocks.quote(symbol='AMZN')
}
