% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.screener.arktrades}
\alias{stocks.screener.arktrades}
\title{Gets a dataframe of ARK trades for ticker}
\usage{
stocks.screener.arktrades(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker to get trades for}
}
\description{
Wrapper for Python function stocks.screener.arktrades from OpenBB Terminal SDK
}
