% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.screener.historical_chart}
\alias{stocks.screener.historical_chart}
\title{View historical price of stocks that meet preset}
\usage{
stocks.screener.historical_chart(
  preset_loaded = "top_gainers",
  limit = 10,
  start_date = "2022-10-06",
  type_candle = "a",
  normalize = TRUE,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{preset_loaded}{(character length 1) Preset loaded to filter for tickers}

\item{limit}{(integer length 1) Number of stocks to display}

\item{start_date}{(character length 1) Start date to display historical data, in YYYY-MM-DD format}

\item{type_candle}{(character length 1) Type of candle to display}

\item{normalize}{(logical length 1) Boolean to normalize all stock prices using MinMax}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.screener.historical_chart from OpenBB Terminal SDK
}
\examples{
stocks.screener.historical_chart(preset_loaded='top_gainers', limit=10, start_date='2022-05-29', type_candle='a', normalize=TRUE)
}
