% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.screener.screener_data}
\alias{stocks.screener.screener_data}
\title{Screener Overview}
\usage{
stocks.screener.screener_data(
  preset_loaded = "top_gainers",
  data_type = "overview",
  limit = 10,
  ascend = FALSE
)
}
\arguments{
\item{preset_loaded}{(character length 1) Loaded preset filter}

\item{data_type}{(character length 1) Data type between: overview, valuation, financial, ownership, performance, technical}

\item{limit}{(integer length 1) Limit of stocks filtered with presets to print}

\item{ascend}{(logical length 1) Ascended order of stocks filtered to print}
}
\description{
Wrapper for Python function stocks.screener.screener_data from OpenBB Terminal SDK
}
\examples{
stocks.screener.screener_data(preset_loaded='top_gainers', data_type='overview', limit=10, ascend=FALSE)
}
