% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.screener.screener_data_chart}
\alias{stocks.screener.screener_data_chart}
\title{Screener one of the following: overview, valuation, financial, ownership, performance, technical.}
\usage{
stocks.screener.screener_data_chart(
  loaded_preset = "top_gainers",
  data_type = "overview",
  limit = 10,
  ascend = FALSE,
  sortby = "",
  export = "",
  sheet_name
)
}
\arguments{
\item{loaded_preset}{(character length 1) Preset loaded to filter for tickers}

\item{data_type}{(character length 1) Data type string between: overview, valuation, financial, ownership, performance, technical}

\item{limit}{(integer length 1) Limit of stocks to display}

\item{ascend}{(logical length 1) Order of table to ascend or descend}

\item{sortby}{(character length 1) Column to sort table by}

\item{export}{(character length 1) Export dataframe data to csv,json,xlsx file}
}
\description{
Wrapper for Python function stocks.screener.screener_data_chart from OpenBB Terminal SDK
}
\examples{
stocks.screener.screener_data_chart(loaded_preset='top_gainers', data_type='overview', limit=10, ascend=FALSE)
}
