% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.search}
\alias{stocks.search}
\title{Search selected query for tickers.}
\usage{
stocks.search(
  query = "",
  country = "",
  sector = "",
  industry_group = "",
  industry = "",
  exchange_country = "",
  all_exchanges = FALSE,
  limit = 0
)
}
\arguments{
\item{query}{(character length 1) The search term used to find company tickers}

\item{country}{(character length 1) Search by country to find stocks matching the criteria}

\item{sector}{(character length 1) Search by sector to find stocks matching the criteria}

\item{industry}{(character length 1) Search by industry to find stocks matching the criteria}

\item{exchange_country}{(character length 1) Search by exchange country to find stock matching}

\item{all_exchanges}{(logical length 1) Whether to search all exchanges, without this option only the United States market is searched}

\item{limit}{(integer length 1) The limit of companies shown.}
}
\description{
Wrapper for Python function stocks.search from OpenBB Terminal SDK
}
\examples{
stocks.search(limit=0)
}
