% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.recom}
\alias{stocks.ta.recom}
\title{Get tradingview recommendation based on technical indicators}
\usage{
stocks.ta.recom(symbol, screener = "america", exchange = "", interval = "")
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the recommendation from tradingview based on technical indicators}

\item{screener}{(character length 1) Screener based on tradingview docs https://python-tradingview-ta.readthedocs.io/en/latest/usage.html}

\item{exchange}{(character length 1) Exchange based on tradingview docs https://python-tradingview-ta.readthedocs.io/en/latest/usage.html}

\item{interval}{(character length 1) Interval time to check technical indicators and correspondent recommendation}
}
\description{
Wrapper for Python function stocks.ta.recom from OpenBB Terminal SDK
}
\examples{
stocks.ta.recom(symbol='AMZN', screener='america')
}
