% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.recom_chart}
\alias{stocks.ta.recom_chart}
\title{Print tradingview recommendation based on technical indicators}
\usage{
stocks.ta.recom_chart(
  symbol,
  screener = "america",
  exchange = "",
  interval = "",
  export = "",
  sheet_name
)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get tradingview recommendation based on technical indicators}

\item{screener}{(character length 1) Screener based on tradingview docs https://python-tradingview-ta.readthedocs.io/en/latest/usage.html}

\item{exchange}{(character length 1) Exchange based on tradingview docs https://python-tradingview-ta.readthedocs.io/en/latest/usage.html}

\item{interval}{(character length 1) Interval time to check technical indicators and correspondent recommendation}

\item{export}{(character length 1) Format of export file}
}
\description{
Wrapper for Python function stocks.ta.recom_chart from OpenBB Terminal SDK
}
\examples{
stocks.ta.recom_chart(symbol='AMZN', screener='america')
}
