% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.rsp}
\alias{stocks.ta.rsp}
\title{Relative strength percentile \link{Source: https://github.com/skyte/relative-strength}}
\usage{
stocks.ta.rsp(s_ticker = "")
}
\arguments{
\item{s_ticker}{(character length 1) Stock Ticker}
}
\description{
Wrapper for Python function stocks.ta.rsp from OpenBB Terminal SDK
}
