% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.rsp_chart}
\alias{stocks.ta.rsp_chart}
\title{Display Relative Strength Percentile \link{Source: https://github.com/skyte/relative-strength}}
\usage{
stocks.ta.rsp_chart(
  s_ticker = "",
  export = "",
  sheet_name,
  tickers_show = FALSE
)
}
\arguments{
\item{s_ticker}{(character length 1) Stock ticker}

\item{export}{(character length 1) Format of export file}

\item{tickers_show}{(logical length 1) Boolean to check if tickers in the same industry as the stock should be shown}
}
\description{
Wrapper for Python function stocks.ta.rsp_chart from OpenBB Terminal SDK
}
\examples{
stocks.ta.rsp_chart(tickers_show=FALSE)
}
