% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.summary}
\alias{stocks.ta.summary}
\title{Get technical summary report provided by FinBrain's API}
\usage{
stocks.ta.summary(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the technical summary}
}
\description{
Wrapper for Python function stocks.ta.summary from OpenBB Terminal SDK
}
\examples{
stocks.ta.summary(symbol='AMZN')
}
