% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.summary_chart}
\alias{stocks.ta.summary_chart}
\title{Print technical summary report provided by FinBrain's API}
\usage{
stocks.ta.summary_chart(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol to get the technical summary}
}
\description{
Wrapper for Python function stocks.ta.summary_chart from OpenBB Terminal SDK
}
\examples{
stocks.ta.summary_chart(symbol='AMZN')
}
