% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.view}
\alias{stocks.ta.view}
\title{Get finviz image for given ticker}
\usage{
stocks.ta.view(symbol)
}
\arguments{
\item{symbol}{(character length 1) Ticker symbol}
}
\description{
Wrapper for Python function stocks.ta.view from OpenBB Terminal SDK
}
\examples{
stocks.ta.view(symbol='AMZN')
}
