% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.ta.view_chart}
\alias{stocks.ta.view_chart}
\title{View finviz image for ticker}
\usage{
stocks.ta.view_chart(symbol, external_axes)
}
\arguments{
\item{symbol}{(character length 1) Stock ticker symbol}
}
\description{
Wrapper for Python function stocks.ta.view_chart from OpenBB Terminal SDK
}
\examples{
stocks.ta.view_chart(symbol='AMZN')
}
