% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.th.check_if_open}
\alias{stocks.th.check_if_open}
\title{Check if market open helper function}
\usage{
stocks.th.check_if_open(bursa, exchange)
}
\description{
Wrapper for Python function stocks.th.check_if_open from OpenBB Terminal SDK
}
