% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.th.closed_chart}
\alias{stocks.th.closed_chart}
\title{Display closed exchanges.}
\usage{
stocks.th.closed_chart()
}
\description{
Wrapper for Python function stocks.th.closed_chart from OpenBB Terminal SDK
}
