% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.th.exchange}
\alias{stocks.th.exchange}
\title{Get current exchange open hours.}
\usage{
stocks.th.exchange(symbol)
}
\arguments{
\item{symbol}{(character length 1) Exchange symbol}
}
\description{
Wrapper for Python function stocks.th.exchange from OpenBB Terminal SDK
}
\examples{
stocks.th.exchange(symbol='AMZN')
}
