% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stocks.tob}
\alias{stocks.tob}
\title{Get top of book bid and ask for ticker on exchange \link{CBOE.com}}
\usage{
stocks.tob(symbol, exchange = "BZX")
}
\arguments{
\item{symbol}{(character length 1) Ticker to get}

\item{exchange}{(character length 1) Exchange to look at.  Can be \code{BZX},\code{EDGX}, \code{BYX}, \code{EDGA}}
}
\description{
Wrapper for Python function stocks.tob from OpenBB Terminal SDK
}
\examples{
stocks.tob(symbol='AMZN', exchange='BZX')
}
