% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.ad}
\alias{ta.ad}
\title{Calculate AD technical indicator}
\usage{
ta.ad(data, use_open = FALSE)
}
\arguments{
\item{data}{(data.frame) Dataframe of prices with OHLC and Volume}

\item{use_open}{(logical length 1) Whether to use open prices}
}
\description{
Wrapper for Python function ta.ad from OpenBB Terminal SDK
}
\examples{
ta.ad(use_open=FALSE)
}
