% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.ad_chart}
\alias{ta.ad_chart}
\title{Plots AD technical indicator}
\usage{
ta.ad_chart(
  data,
  use_open = FALSE,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{use_open}{(logical length 1) Whether to use open prices in calculation}

\item{symbol}{(character length 1) Ticker symbol}

\item{export}{(character length 1) Format to export data as}
}
\description{
Wrapper for Python function ta.ad_chart from OpenBB Terminal SDK
}
\examples{
ta.ad_chart(use_open=FALSE)
}
