% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.adosc}
\alias{ta.adosc}
\title{Calculate AD oscillator technical indicator}
\usage{
ta.adosc(data, use_open = FALSE, fast = 3, slow = 10)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{use_open}{(logical length 1) Whether to use open prices}

\item{fast}{(integer length 1) Fast value}

\item{slow}{(integer length 1) Slow value}
}
\description{
Wrapper for Python function ta.adosc from OpenBB Terminal SDK
}
\examples{
ta.adosc(use_open=FALSE, fast=3, slow=10)
}
