% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.adosc_chart}
\alias{ta.adosc_chart}
\title{Plots AD Osc Indicator}
\usage{
ta.adosc_chart(
  data,
  fast = 3,
  slow = 10,
  use_open = FALSE,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{fast}{(integer length 1) Length of fast window}

\item{slow}{(integer length 1) Length of slow window}

\item{use_open}{(logical length 1) Whether to use open prices in calculation}

\item{symbol}{(character length 1) Stock ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.adosc_chart from OpenBB Terminal SDK
}
\examples{
ta.adosc_chart(use_open=FALSE, fast=3, slow=10)
}
