% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.adx}
\alias{ta.adx}
\title{ADX technical indicator}
\usage{
ta.adx(data, window = 14, scalar = 100, drift = 1)
}
\arguments{
\item{data}{(data.frame) Dataframe with OHLC price data}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(integer length 1) Scalar variable}

\item{drift}{(integer length 1) Drift variable}
}
\description{
Wrapper for Python function ta.adx from OpenBB Terminal SDK
}
\examples{
ta.adx(window=14, scalar=100, drift=1)
}
