% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.adx_chart}
\alias{ta.adx_chart}
\title{Plots ADX indicator}
\usage{
ta.adx_chart(
  data,
  window = 14,
  scalar = 100,
  drift = 1,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe with OHLC price data}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(integer length 1) Scalar variable}

\item{drift}{(integer length 1) Drift variable}

\item{symbol}{(character length 1) Ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.adx_chart from OpenBB Terminal SDK
}
\examples{
ta.adx_chart(window=14, scalar=100, drift=1)
}
