% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.aroon}
\alias{ta.aroon}
\title{Aroon technical indicator}
\usage{
ta.aroon(data, window = 25, scalar = 100)
}
\arguments{
\item{data}{(data.frame) Dataframe with OHLC price data}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(integer length 1) Scalar variable}
}
\description{
Wrapper for Python function ta.aroon from OpenBB Terminal SDK
}
\examples{
ta.aroon(window=25, scalar=100)
}
