% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.aroon_chart}
\alias{ta.aroon_chart}
\title{Plots Aroon indicator}
\usage{
ta.aroon_chart(
  data,
  window = 25,
  scalar = 100,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe with OHLC price data}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(integer length 1) Scalar variable}

\item{symbol}{(character length 1) Ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.aroon_chart from OpenBB Terminal SDK
}
\examples{
ta.aroon_chart(window=25, scalar=100)
}
