% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.atr}
\alias{ta.atr}
\title{Average True Range}
\usage{
ta.atr(data, window = 14, mamode = "ema", offset = 0)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{window}{(integer length 1) Length of window}

\item{mamode}{(character length 1) Type of filter}

\item{offset}{(integer length 1) Offset value}
}
\description{
Wrapper for Python function ta.atr from OpenBB Terminal SDK
}
\examples{
ta.atr(window=14, mamode='ema', offset=0)
}
