% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.atr_chart}
\alias{ta.atr_chart}
\title{Plots ATR}
\usage{
ta.atr_chart(
  data,
  symbol = "",
  window = 14,
  mamode = "sma",
  offset = 0,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{symbol}{(character length 1) Ticker symbol}

\item{window}{(integer length 1) Length of window to calculate upper channel}

\item{export}{(character length 1) Format of export file}
}
\description{
Wrapper for Python function ta.atr_chart from OpenBB Terminal SDK
}
\examples{
ta.atr_chart(window=14)
}
