% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.bbands}
\alias{ta.bbands}
\title{Calculate Bollinger Bands}
\usage{
ta.bbands(data, window = 15, n_std = 2, mamode = "ema")
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{window}{(integer length 1) Length of window to calculate BB}

\item{n_std}{(numeric length 1) Number of standard deviations to show}

\item{mamode}{(character length 1) Method of calculating average}
}
\description{
Wrapper for Python function ta.bbands from OpenBB Terminal SDK
}
\examples{
ta.bbands(window=15, n_std=2, mamode='ema')
}
