% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.bbands_chart}
\alias{ta.bbands_chart}
\title{Plots bollinger bands}
\usage{
ta.bbands_chart(
  data,
  symbol = "",
  window = 15,
  n_std = 2,
  mamode = "sma",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{symbol}{(character length 1) Ticker symbol}

\item{window}{(integer length 1) Length of window to calculate BB}

\item{n_std}{(numeric length 1) Number of standard deviations to show}

\item{mamode}{(character length 1) Method of calculating average}

\item{export}{(character length 1) Format of export file}
}
\description{
Wrapper for Python function ta.bbands_chart from OpenBB Terminal SDK
}
\examples{
ta.bbands_chart(window=15, n_std=2, mamode='sma')
}
