% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.cci}
\alias{ta.cci}
\title{Commodity channel index}
\usage{
ta.cci(data, window = 14, scalar = 0.0015)
}
\arguments{
\item{window}{(integer length 1) Length of window}

\item{scalar}{(numeric length 1) Scalar variable}
}
\description{
Wrapper for Python function ta.cci from OpenBB Terminal SDK
}
\examples{
ta.cci(window=14, scalar=0.0015)
}
