% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.cg}
\alias{ta.cg}
\title{Center of gravity}
\usage{
ta.cg(values, window)
}
\arguments{
\item{values}{(Python type: pandas.core.series.Series) Data to use with close being titled values}

\item{window}{(integer length 1) Length for indicator window}
}
\description{
Wrapper for Python function ta.cg from OpenBB Terminal SDK
}
