% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.cg_chart}
\alias{ta.cg_chart}
\title{Plots center of gravity Indicator}
\usage{
ta.cg_chart(
  data,
  window = 14,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series of values}

\item{window}{(integer length 1) Length of window}

\item{symbol}{(character length 1) Stock ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.cg_chart from OpenBB Terminal SDK
}
\examples{
ta.cg_chart(window=14)
}
