% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.clenow}
\alias{ta.clenow}
\title{Gets the Clenow Volatility Adjusted Momentum.  this is defined as the regression coefficient on log prices}
\usage{
ta.clenow(values, window = 90)
}
\arguments{
\item{values}{(Python type: pandas.core.series.Series) Values to perform regression for}

\item{window}{(integer length 1) Length of lookback period}
}
\description{
Wrapper for Python function ta.clenow from OpenBB Terminal SDK
}
\examples{
ta.clenow(window=90)
}
