% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.clenow_chart}
\alias{ta.clenow_chart}
\title{Prints table and plots clenow momentum}
\usage{
ta.clenow_chart(
  data,
  symbol = "",
  window = 90,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series of values}

\item{symbol}{(character length 1) Symbol that the data corresponds to}

\item{window}{(integer length 1) Length of window}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.clenow_chart from OpenBB Terminal SDK
}
\examples{
ta.clenow_chart(window=90)
}
