% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.cones_chart}
\alias{ta.cones_chart}
\title{Plots the realized volatility quantiles for the loaded ticker.}
\usage{
ta.cones_chart(
  data,
  symbol = "",
  lower_q = 0.25,
  upper_q = 0.75,
  model = "STD",
  is_crypto = FALSE,
  export = "",
  sheet_name
)
}
\arguments{
\item{data}{(data.frame) DataFrame of OHLC prices.}

\item{symbol}{(character length 1) The ticker symbol.}

\item{lower_q}{(numeric length 1) The lower quantile to calculate for.}

\item{upper_q}{(numeric length 1) The upper quantile to for.}

\item{model}{(character length 1) The model to use for volatility calculation. Choices are:
\link{"STD", "Parkinson", "Garman-Klass", "Hodges-Tompkins", "Rogers-Satchell", "Yang-Zhang"}}

\item{is_crypto}{(logical length 1) If true, volatility is calculated for 365 days instead of 252.}
}
\description{
Wrapper for Python function ta.cones_chart from OpenBB Terminal SDK
}
