% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.demark}
\alias{ta.demark}
\title{Get the integer value for demark sequential indicator}
\usage{
ta.demark(values)
}
\arguments{
\item{values}{(Python type: pandas.core.series.Series) Series of close values}
}
\description{
Wrapper for Python function ta.demark from OpenBB Terminal SDK
}
