% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.donchian}
\alias{ta.donchian}
\title{Calculate Donchian Channels}
\usage{
ta.donchian(data, upper_length = 20, lower_length = 20)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{upper_length}{(integer length 1) Length of window to calculate upper channel}

\item{lower_length}{(integer length 1) Length of window to calculate lower channel}
}
\description{
Wrapper for Python function ta.donchian from OpenBB Terminal SDK
}
\examples{
ta.donchian(upper_length=20, lower_length=20)
}
