% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.donchian_chart}
\alias{ta.donchian_chart}
\title{Plots donchian channels}
\usage{
ta.donchian_chart(
  data,
  symbol = "",
  upper_length = 20,
  lower_length = 20,
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{symbol}{(character length 1) Ticker symbol}

\item{upper_length}{(integer length 1) Length of window to calculate upper channel}

\item{lower_length}{(integer length 1) Length of window to calculate lower channel}

\item{export}{(character length 1) Format of export file}
}
\description{
Wrapper for Python function ta.donchian_chart from OpenBB Terminal SDK
}
\examples{
ta.donchian_chart(upper_length=20, lower_length=20)
}
