% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.ema}
\alias{ta.ema}
\title{Gets exponential moving average (EMA) for stock}
\usage{
ta.ema(data, length = 50, offset = 0)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of dates and prices}

\item{length}{(integer length 1) Length of EMA window}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.ema from OpenBB Terminal SDK
}
\examples{
ta.ema(length=50, offset=0)
}
