% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.fib}
\alias{ta.fib}
\title{Calculate Fibonacci levels}
\usage{
ta.fib(data, limit = 120, start_date, end_date)
}
\arguments{
\item{data}{(data.frame) Dataframe of prices}

\item{limit}{(integer length 1) Days to look back for retracement}
}
\description{
Wrapper for Python function ta.fib from OpenBB Terminal SDK
}
\examples{
ta.fib(limit=120)
}
