% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.fib_chart}
\alias{ta.fib_chart}
\title{Plots Calculated fibonacci retracement levels}
\usage{
ta.fib_chart(
  data,
  limit = 120,
  start_date,
  end_date,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) OHLC data}

\item{limit}{(integer length 1) Days to lookback}

\item{symbol}{(character length 1) Ticker symbol}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.fib_chart from OpenBB Terminal SDK
}
\examples{
ta.fib_chart(limit=120)
}
