% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.fisher}
\alias{ta.fisher}
\title{Fisher Transform}
\usage{
ta.fisher(data, window = 14)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{window}{(integer length 1) Length for indicator window}
}
\description{
Wrapper for Python function ta.fisher from OpenBB Terminal SDK
}
\examples{
ta.fisher(window=14)
}
