% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.fisher_chart}
\alias{ta.fisher_chart}
\title{Plots Fisher Indicator}
\usage{
ta.fisher_chart(
  data,
  window = 14,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{window}{(integer length 1) Length of window}

\item{symbol}{(character length 1) Ticker string}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.fisher_chart from OpenBB Terminal SDK
}
\examples{
ta.fisher_chart(window=14)
}
