% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.hma}
\alias{ta.hma}
\title{Gets hull moving average (HMA) for stock}
\usage{
ta.hma(data, length = 50, offset = 0)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of dates and prices}

\item{length}{(integer length 1) Length of SMA window}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.hma from OpenBB Terminal SDK
}
\examples{
ta.hma(length=50, offset=0)
}
