% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.kc}
\alias{ta.kc}
\title{Keltner Channels}
\usage{
ta.kc(data, window = 20, scalar = 2, mamode = "ema", offset = 0)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(numeric length 1) Scalar value}

\item{mamode}{(character length 1) Type of filter}

\item{offset}{(integer length 1) Offset value}
}
\description{
Wrapper for Python function ta.kc from OpenBB Terminal SDK
}
\examples{
ta.kc(window=20, scalar=2, mamode='ema', offset=0)
}
