% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.kc_chart}
\alias{ta.kc_chart}
\title{Plots Keltner Channels Indicator}
\usage{
ta.kc_chart(
  data,
  window = 20,
  scalar = 2,
  mamode = "ema",
  offset = 0,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\description{
Wrapper for Python function ta.kc_chart from OpenBB Terminal SDK
}
\examples{
ta.kc_chart(window=20, scalar=2, mamode='ema', offset=0)
}
