% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.ma_chart}
\alias{ta.ma_chart}
\title{Plots MA technical indicator}
\usage{
ta.ma_chart(
  data,
  window,
  offset = 0,
  ma_type = "EMA",
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Series of prices}

\item{offset}{(integer length 1) Offset variable}

\item{ma_type}{(character length 1) Type of moving average.  Either "EMA" "ZLMA" or "SMA"}

\item{symbol}{(character length 1) Ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.ma_chart from OpenBB Terminal SDK
}
\examples{
ta.ma_chart(offset=0, ma_type='EMA')
}
