% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.macd}
\alias{ta.macd}
\title{Moving average convergence divergence}
\usage{
ta.macd(data, n_fast = 12, n_slow = 26, n_signal = 9)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Values for calculation}

\item{n_fast}{(integer length 1) Fast period}

\item{n_slow}{(integer length 1) Slow period}

\item{n_signal}{(integer length 1) Signal period}
}
\description{
Wrapper for Python function ta.macd from OpenBB Terminal SDK
}
\examples{
ta.macd(n_fast=12, n_slow=26, n_signal=9)
}
