% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.macd_chart}
\alias{ta.macd_chart}
\title{Plots MACD signal}
\usage{
ta.macd_chart(
  data,
  n_fast = 12,
  n_slow = 26,
  n_signal = 9,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Values to input}

\item{n_fast}{(integer length 1) Fast period}

\item{n_slow}{(integer length 1) Slow period}

\item{n_signal}{(integer length 1) Signal period}

\item{symbol}{(character length 1) Stock ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.macd_chart from OpenBB Terminal SDK
}
\examples{
ta.macd_chart(n_fast=12, n_slow=26, n_signal=9)
}
