% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.obv}
\alias{ta.obv}
\title{On Balance Volume}
\usage{
ta.obv(data)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}
}
\description{
Wrapper for Python function ta.obv from OpenBB Terminal SDK
}
