% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.obv_chart}
\alias{ta.obv_chart}
\title{Plots OBV technical indicator}
\usage{
ta.obv_chart(data, symbol = "", export = "", sheet_name, external_axes)
}
\arguments{
\item{data}{(data.frame) Dataframe of ohlc prices}

\item{symbol}{(character length 1) Ticker}

\item{export}{(character length 1) Format to export data as}
}
\description{
Wrapper for Python function ta.obv_chart from OpenBB Terminal SDK
}
