% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.rsi}
\alias{ta.rsi}
\title{Relative strength index}
\usage{
ta.rsi(data, window = 14, scalar = 100, drift = 1)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of prices}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(numeric length 1) Scalar variable}

\item{drift}{(integer length 1) Drift variable}
}
\description{
Wrapper for Python function ta.rsi from OpenBB Terminal SDK
}
\examples{
ta.rsi(window=14, scalar=100, drift=1)
}
