% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.rsi_chart}
\alias{ta.rsi_chart}
\title{Plots RSI Indicator}
\usage{
ta.rsi_chart(
  data,
  window = 14,
  scalar = 100,
  drift = 1,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Values to input}

\item{window}{(integer length 1) Length of window}

\item{scalar}{(numeric length 1) Scalar variable}

\item{drift}{(integer length 1) Drift variable}

\item{symbol}{(character length 1) Stock ticker}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.rsi_chart from OpenBB Terminal SDK
}
\examples{
ta.rsi_chart(window=14, scalar=100, drift=1)
}
