% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.rvol_std}
\alias{ta.rvol_std}
\title{Standard deviation measures how widely returns are dispersed from the average return.}
\usage{
ta.rvol_std(
  data,
  window = 30,
  trading_periods = 252,
  is_crypto = FALSE,
  clean = TRUE
)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices.}

\item{window}{(integer length 1) Length of window to calculate standard deviation.}

\item{trading_periods}{(integer length 1) Number of trading periods in a year.}

\item{is_crypto}{(logical length 1) If true, trading_periods is defined as 365.}

\item{clean}{(logical length 1) Whether to clean the data or not by dropping NaN values.}
}
\description{
Wrapper for Python function ta.rvol_std from OpenBB Terminal SDK
}
