% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.rvol_yang_zhang}
\alias{ta.rvol_yang_zhang}
\title{Yang-Zhang volatility is the combination of the overnight (close-to-open volatility).}
\usage{
ta.rvol_yang_zhang(
  data,
  window = 30,
  trading_periods = 252,
  is_crypto = TRUE,
  clean
)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices.}

\item{window}{(integer length 1) Length of window to calculate standard deviation.}

\item{trading_periods}{(integer length 1) Number of trading periods in a year.}

\item{is_crypto}{(logical length 1) If true, trading_periods is defined as 365.}

\item{clean}{(Python type: inspect._empty) Whether to clean the data or not by dropping NaN values.}
}
\description{
Wrapper for Python function ta.rvol_yang_zhang from OpenBB Terminal SDK
}
