% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.sma}
\alias{ta.sma}
\title{Gets simple moving average (SMA) for stock}
\usage{
ta.sma(data, length = 50, offset = 0)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of dates and prices}

\item{length}{(integer length 1) Length of SMA window}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.sma from OpenBB Terminal SDK
}
\examples{
ta.sma(length=50, offset=0)
}
