% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.stoch}
\alias{ta.stoch}
\title{Stochastic oscillator}
\usage{
ta.stoch(data, fastkperiod = 14, slowdperiod = 3, slowkperiod = 3)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{fastkperiod}{(integer length 1) Fast k period}

\item{slowdperiod}{(integer length 1) Slow d period}

\item{slowkperiod}{(integer length 1) Slow k period}
}
\description{
Wrapper for Python function ta.stoch from OpenBB Terminal SDK
}
\examples{
ta.stoch(fastkperiod=14, slowdperiod=3, slowkperiod=3)
}
