% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.stoch_chart}
\alias{ta.stoch_chart}
\title{Plots stochastic oscillator signal}
\usage{
ta.stoch_chart(
  data,
  fastkperiod = 14,
  slowdperiod = 3,
  slowkperiod = 3,
  symbol = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{fastkperiod}{(integer length 1) Fast k period}

\item{slowdperiod}{(integer length 1) Slow d period}

\item{slowkperiod}{(integer length 1) Slow k period}

\item{symbol}{(character length 1) Stock ticker symbol}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.stoch_chart from OpenBB Terminal SDK
}
\examples{
ta.stoch_chart(fastkperiod=14, slowdperiod=3, slowkperiod=3)
}
