% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.vwap}
\alias{ta.vwap}
\title{Gets volume weighted average price (VWAP)}
\usage{
ta.vwap(data, offset = 0)
}
\arguments{
\item{data}{(data.frame) Dataframe of dates and prices}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.vwap from OpenBB Terminal SDK
}
\examples{
ta.vwap(offset=0)
}
