% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.vwap_chart}
\alias{ta.vwap_chart}
\title{Plots VWMA technical indicator}
\usage{
ta.vwap_chart(
  data,
  symbol = "",
  start_date,
  end_date,
  offset = 0,
  interval = "",
  export = "",
  sheet_name,
  external_axes
)
}
\arguments{
\item{data}{(data.frame) Dataframe of OHLC prices}

\item{symbol}{(character length 1) Ticker}

\item{offset}{(integer length 1) Offset variable}

\item{interval}{(character length 1) Interval of data}

\item{export}{(character length 1) Format to export data}
}
\description{
Wrapper for Python function ta.vwap_chart from OpenBB Terminal SDK
}
\examples{
ta.vwap_chart(offset=0)
}
