% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.wma}
\alias{ta.wma}
\title{Gets weighted moving average (WMA) for stock}
\usage{
ta.wma(data, length = 50, offset = 0)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of dates and prices}

\item{length}{(integer length 1) Length of SMA window}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.wma from OpenBB Terminal SDK
}
\examples{
ta.wma(length=50, offset=0)
}
