% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ta.zlma}
\alias{ta.zlma}
\title{Gets zero-lagged exponential moving average (ZLEMA) for stock}
\usage{
ta.zlma(data, length = 50, offset = 0)
}
\arguments{
\item{data}{(Python type: pandas.core.series.Series) Dataframe of dates and prices}

\item{length}{(integer length 1) Length of EMA window}

\item{offset}{(integer length 1) Length of offset}
}
\description{
Wrapper for Python function ta.zlma from OpenBB Terminal SDK
}
\examples{
ta.zlma(length=50, offset=0)
}
